### 2025-05-26 #8281 @editorialbot
[PRE REVIEW]: GeneForgeLang: A Symbolic Language for Causal Reasoning in Genetic Engineering and Molecular Bioinformatics
**Submitting author:** @manuelmenendezgonzalez (<a href="http://orcid.org/0000-0002-5218-0774">Manuel Menendez-Gonzalez</a>)
**Repository:** https://github.com/Fundacion-de-Neurociencias/GeneForgeLang
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
