### 2025-05-26 #8281 @manuelmenendezgonzalez
Hi,
As reviewer perhaps Charles Tapley Hoyt](https://reviewers.joss.theoj.org/reviewers/1942) or [Yen-Chung Chen](https://reviewers.joss.theoj.org/reviewers/2012)
License is MIT: https://github.com/Fundacion-de-Neurociencias/GeneForge/blob/main/LICENSE
Best,
Manuel
