### 2025-05-27 e263f68
feat: extended grammar, parser + probabilistic rule engine
