### 2025-05-30 a210f6f
feat: add residue(ncAA) construct to grammar & parser
