+++ gfl/grammar.md
@@ -1,0 +1,44 @@
+﻿### 🔄 2025-05  New constructs
+* **`vector()`** – delivery vector and its attributes  
+* **`region()`** – chromosomal region / repeat block  
+* **`governance()`** – ethical compliance flags  
+* **`risk()`** – quantitative risk annotation  
+* **`repeat_edit()`** – interruption of pathogenic triplet repeats
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🔄 Mayo 2025 – Nuevas construcciones
+* **ector()** – vector de entrega y atributos  
+* **egion()** – región cromosómica / repetidos  
+* **governance()** – flags de cumplimiento ético  
+* **isk()** – anotación cuantitativa de riesgo  
+* **epeat_edit()** – interrupción de repeticiones patogénicas
+### 🆕 Mayo 2025 – Soporte de aminoácidos no canónicos
+* `residue(ncAA:<name>)` — declara la inserción o edición de un residuo no estándar
