### 2025-05-30 f7811e8
feat: generic parser supports multi‑attribute statements
