+++ gfl/parser.py
@@ -1,7 +1,12 @@
+﻿"""GFL parser  — now supports comma‑separated multi‑attributes in one line.
+Any statement of the form  name(key:val, key2=val2, singleval)  is converted
+into a GFLNode whose attrs dict may contain >1 keys.
+"""
 import re
-from typing import Any, Dict, List
+from typing import Dict, Any, List
 
-class GFLParseError(Exception): pass
+class GFLParseError(Exception):
+    pass
 
 class GFLNode:
     def __init__(self, type_: str, attrs: Dict[str, Any] | None = None):
@@ -9,46 +14,59 @@ class GFLNode:
         self.attrs = attrs or {}
         self.children: List["GFLNode"] = []
 
-    def to_dict(self): return {
-        "type": self.type, "attrs": self.attrs,
-        "children": [c.to_dict() for c in self.children]
-    }
+    def to_dict(self):
+        return {
+            "type": self.type,
+            "attrs": self.attrs,
+            "children": [c.to_dict() for c in self.children],
+        }
 
     def __repr__(self):
         import json; return json.dumps(self.to_dict(), indent=2)
 
 class GFLParser:
-    """Parser with Vector, Region, Governance, Risk and Repeat-Edit support."""
-    def __init__(self):
-        self.rx = {
-            "edit":     re.compile(r"^edit\\(([^)]+)\\)$"),
-            "target":   re.compile(r"^target\\(([^:]+):([^)]+)\\)$"),
-            "effect":   re.compile(r"^effect\\(([^:]+):([^)]+)\\)$"),
-            "pathway":  re.compile(r"^pathway\\(([^:]+):([^)]+)\\)$"),
-            "link":     re.compile(r"^link\\(([^-]+)->([^)]+)\\)$"),
-            "simulate": re.compile(r"^simulate\\(([^)]+)\\)$"),
-            # extensions
-            "vector":      re.compile(r"^vector\\((.+)\\)$"),
-            "region":      re.compile(r"^region\\((.+)\\)$"),
-            "governance":  re.compile(r"^governance\\((.+)\\)$"),
-            "risk":        re.compile(r"^risk\\((.+)\\)$"),
-            "repeat_edit": re.compile(r"^repeat_edit\\((.+)\\)$"),
-        }
+    """Single‑regex parser with flexible attribute splitting."""
 
+    _LINK_RX = re.compile(r"^link\(([^-]+)->([^)]+)\)$")
+    _GEN_RX  = re.compile(r"^(\w+)\((.*)\)$")
+
+    def _kvdict(self, body: str) -> Dict[str, Any]:
+        """Split comma‑separated items into a dict.
+        Accepts  key:val  or  key=val  or lonevalue  (stored as {'value':...})."""
+        if not body.strip():
+            return {}
+        items = [p.strip() for p in body.split(',') if p.strip()]
+        # single bare value
+        if len(items) == 1 and all(c not in items[0] for c in (':','=')):
+            return {"value": items[0]}
+        out: Dict[str, Any] = {}
+        for it in items:
+            if ':' in it:
+                k, v = it.split(':', 1)
+            elif '=' in it:
+                k, v = it.split('=', 1)
+            else:  # bare value as fall‑back
+                k, v = "value", it
+            out[k.strip()] = v.strip()
+        return out
+
+    # ------------------------------------------------------------------
     def _parse_line(self, line: str):
-        for typ, rx in self.rx.items():
-            m = rx.match(line.strip())
-            if not m: continue
-            if typ in ("target", "effect", "pathway"):
-                return GFLNode(typ, {"key": m.group(1), "val": m.group(2)})
-            if typ == "link":
-                return GFLNode(typ, {"from": m.group(1), "to": m.group(2)})
-            return GFLNode(typ, {"val": m.group(1)})
[…]
