+++ gfl/grammar.md
@@ -42,3 +42,10 @@
 * **epeat_edit()** – interrupción de repeticiones patogénicas
 ### 🆕 Mayo 2025 – Soporte de aminoácidos no canónicos
 * `residue(ncAA:<name>)` — declara la inserción o edición de un residuo no estándar
+**Nota (Mayo 2025)**  Una llamada puede incluir **varios atributos separados por comas**:
+
+```gfl
+vector(AAV9, serotype=9, tropism=retina, payload_kb=4.7)
+target(gene:ADE2, organism:Saccharomyces_cerevisiae, locus:ChrXV, allele:wildtype)
+```
+Cada par `key:val` o `key=val` se guarda en `attrs`.
