### 2025-05-30 7396810
feat: add immunity() construct & free‑form attributes
