### 2025-05-31 930d6a0
Actualización de identidad, núcleo y manifiesto de Amaia
