### 2025-06-01 b6f12ee
feat: extend grammar with UTR/translation features & add TranslateLSTM stub
