### 2025-06-01 d6d25e1
feat: update parser to handle multi-line with balanced parentheses (Jun 5, 2025)
