+++ gfl/parser.py
@@ -1,35 +1,75 @@
 ﻿import re, json
 from typing import Any, Dict, List
-class GFLParseError(Exception): pass
+
+class GFLParseError(Exception):
+    pass
+
 class GFLNode:
-    def __init__(self, t:str, attrs:Dict[str,Any]|None=None):
-        self.type=t; self.attrs=attrs or {}; self.children:List["GFLNode"]=[]
-    def to_dict(self): return {"type":self.type,"attrs":self.attrs,"children":[c.to_dict() for c in self.children]}
-    def __repr__(self): return json.dumps(self.to_dict(),indent=2)
+    def __init__(self, type_: str, attrs: Dict[str, Any] | None = None):
+        self.type = type_
+        self.attrs = attrs or {}
+        self.children: List["GFLNode"] = []
+
+    def to_dict(self):
+        return {
+            "type": self.type,
+            "attrs": self.attrs,
+            "children": [c.to_dict() for c in self.children]
+        }
+
+    def __repr__(self):
+        return json.dumps(self.to_dict(), indent=2)
 
 class GFLParser:
-    _GEN = re.compile(r'^(\w+)\((.*)\)$')
-    _LINK= re.compile(r'^link\(([^-]+)->([^)]+)\)$')
-    def _split_attrs(self,s):
-        if not s.strip(): return {}
-        d={} ; parts=[p.strip() for p in s.split(',') if p.strip()]
-        if len(parts)==1 and all(c not in parts[0] for c in ':='): return {"value":parts[0]}
-        for p in parts:
-            if ':' in p: k,v=p.split(':',1)
-            elif '=' in p: k,v=p.split('=',1)
-            else: k,v="value",p
-            d[k.strip()]=v.strip()
-        return d
-    def _parse_line(self,l):
-        l=l.strip();
-        m=self._LINK.match(l)
-        if m: return GFLNode('link',{'from':m.group(1),'to':m.group(2)})
-        m=self._GEN.match(l)
-        if m: return GFLNode(m.group(1),self._split_attrs(m.group(2)))
-        if l: raise GFLParseError(f"Syntax error: {l}")
-    def parse(self,text:str):
-        root=GFLNode('program')
-        for raw in text.strip().splitlines():
-            n=self._parse_line(raw)
-            if n: root.children.append(n)
+    """Parser that combines multi-line statements based on balanced parentheses and strips BOM."""
+
+    _LINK_RX = re.compile(r"^link\(([^-]+)->([^)]+)\)$")
+    _GEN_RX  = re.compile(r"^(\w+)\((.*)\)$")
+
+    def _split_attrs(self, body: str) -> Dict[str, Any]:
+        if not body.strip():
+            return {}
+        # Split on commas not inside brackets
+        items = [p.strip() for p in re.split(r',(?=(?:[^\[\]]*\[[^\[\]]*\])*[^\[\]]*$)', body) if p.strip()]
+        if len(items) == 1 and all(c not in items[0] for c in (":","=")):
+            return {"value": items[0]}
+        out: Dict[str, Any] = {}
+        for it in items:
+            if ":" in it:
+                k, v = it.split(":", 1)
+            elif "=" in it:
+                k, v = it.split("=", 1)
+            else:
+                k, v = "value", it
+            out[k.strip()] = v.strip()
+        return out
+
+    def _parse_line(self, line: str):
+        line = line.lstrip("\ufeff").strip()
+        if not line:
+            return None
+        m_link = self._LINK_RX.match(line)
+        if m_link:
+            return GFLNode("link", {"from": m_link.group(1), "to": m_link.group(2)})
+        m_gen = self._GEN_RX.match(line)
+        if m_gen:
+            typ, body = m_gen.group(1), m_gen.group(2)
+            return GFLNode(typ, self._split_attrs(body))
[…]
