### 2025-06-01 2df2763
feat: soporte léxico y script main para ejecutar archivos GFL
