+++ gfl/__init__.py
@@ -1,0 +1,4 @@
+
+from .lexer import lexer
+
+from .lexer import lexer
