### 2025-06-01 fcc3dc6
feat: parser extendido con soporte para peg(...), strings, números, signos
