+++ gfl/lexer.py
@@ -1,13 +1,14 @@
 ﻿import ply.lex as lex
 
 tokens = [
-    'ID', 'NUMBER', 'LPAREN', 'RPAREN', 'COMMA', 'EQUALS'
+    'ID', 'STRING', 'NUMBER', 'LPAREN', 'RPAREN', 'COMMA', 'EQUALS', 'PLUS', 'MINUS'
 ]
 
 reserved = {
     'prime_edit': 'PRIME_EDIT',
     'base_edit': 'BASE_EDIT',
-    'prime_del': 'PRIME_DEL'
+    'prime_del': 'PRIME_DEL',
+    'peg': 'PEG'
 }
 
 tokens += list(reserved.values())
@@ -16,7 +17,14 @@ t_LPAREN = r'\('
 t_RPAREN = r'\)'
 t_COMMA  = r','
 t_EQUALS = r'='
+t_PLUS   = r'\+'
+t_MINUS  = r'-'
 t_ignore = ' \t'
+
+def t_STRING(t):
+    r'"[^"]*"'
+    t.value = t.value[1:-1]  # Remove quotes
+    return t
 
 def t_ID(t):
     r'[a-zA-Z_][a-zA-Z0-9_]*'
@@ -24,7 +32,7 @@ def t_ID(t):
     return t
 
 def t_NUMBER(t):
-    r'\d+(\.\d+)?'
+    r'[+-]?\d+(\.\d+)?'
     t.value = float(t.value) if '.' in t.value else int(t.value)
     return t
 
@@ -33,7 +41,7 @@ def t_newline(t):
     t.lexer.lineno += len(t.value)
 
 def t_error(t):
-    print("Illegal character '%s' at line %d" % (t.value[0], t.lineno))
+    print(f"Illegal character '{t.value[0]}' at line {t.lineno}")
     t.lexer.skip(1)
 
 lexer = lex.lex()
