### 2025-06-02 c838929
feat: integración de validador semántico en main.py
