+++ main.py
@@ -1,5 +1,6 @@
 ﻿import sys
 from gfl import parser
+from gfl.semantic_validator import validate_node
 
 def main():
     if len(sys.argv) != 2:
@@ -19,6 +20,15 @@ def main():
 
     print("✅ AST generado:")
     print(result)
+
+    print("\n🔎 Validación semántica:")
+    errors = validate_node(result)
+    if errors:
+        print("❌ Errores encontrados:")
+        for e in errors:
+            print(" -", e)
+    else:
+        print("✅ Validación semántica correcta.")
 
 if __name__ == "__main__":
     main()
