### 2025-06-02 4b9ebbf
fix: limpieza final de lexer y compatibilidad con llamadas anidadas
