### 2025-06-03 7107e8c
✅ Finalizado flujo test_semantics con validación avanzada OK
