+++ gfl/parsetab.py
@@ -1,0 +1,55 @@
+
+# parsetab.py
+# This file is automatically generated. Do not edit.
+# pylint: disable=W,C,R
+_tabversion = '3.10'
+
+_lr_method = 'LALR'
+
+_lr_signature = 'statementsBASE_EDIT COMMA ENDOGENOUS_RETROVIRUS EQUALS ID LPAREN MITOCHONDRIAL_GENE NICK_SITE NUMBER PEG PIRNA PRIME_DEL PRIME_EDIT REVERSE_TRANSCRIPTASE RPAREN STRING TRANSPOSONstatements : statement\n| statement statementsstatement : prime_edit\n| base_edit\n| prime_del\n| peg\n| nick_site\n| reverse_transcriptase\n| pirna\n| transposon\n| endogenous_retrovirus\n| mitochondrial_geneprime_edit : PRIME_EDIT LPAREN arguments RPARENbase_edit : BASE_EDIT LPAREN arguments RPARENprime_del : PRIME_DEL LPAREN arguments RPARENpeg : PEG LPAREN arguments RPARENnick_site : NICK_SITE LPAREN arguments RPARENreverse_transcriptase : REVERSE_TRANSCRIPTASE LPAREN arguments RPARENpirna : PIRNA LPAREN arguments RPARENtransposon : TRANSPOSON LPAREN arguments RPARENendogenous_retrovirus : ENDOGENOUS_RETROVIRUS LPAREN arguments RPARENmitochondrial_gene : MITOCHONDRIAL_GENE LPAREN arguments RPARENarguments : argument\n| argument COMMA argumentsargument : ID EQUALS ID'
+    
[…]
