+++ main.py
@@ -1,6 +1,6 @@
 ﻿import sys
 from gfl import parser
-from gfl.semantic_validator import validate_node
+from gfl.semantic_validator import validate
 
 def main():
     if len(sys.argv) != 2:
@@ -21,14 +21,11 @@ def main():
     print("✅ AST generado:")
     print(result)
 
-    print("\n🔎 Validación semántica:")
-    errors = validate_node(result)
-    if errors:
-        print("❌ Errores encontrados:")
-        for e in errors:
-            print(" -", e)
-    else:
-        print("✅ Validación semántica correcta.")
+    print("\n🔍 Validación semántica:")
+    valid = validate(result)
+    if not valid:
+        print("❌ Validación fallida.")
+        sys.exit(2)
 
 if __name__ == "__main__":
     main()
