### 2025-06-07 0b15b68
chore: add external_integrations folder and README
