+++ external_integrations/README.md
@@ -1,0 +1,8 @@
+﻿# External Integrations
+
+This folder will contain integration modules for BioPython, CRISPOR, and Benchling.
+
+- BioPython: Sequence manipulation and analysis
+- CRISPOR: Guide RNA design
+- Benchling: Data management APIs
+
