### 2025-06-07 2b7140f
feat: add initial FastAPI app for REST API
