+++ api/main.py
@@ -1,0 +1,7 @@
+﻿from fastapi import FastAPI
+
+app = FastAPI()
+
+@app.get('/')
+def read_root():
+    return { 'message': 'Welcome to GeneForge API v1.0.0' }
