### 2025-06-07 afca105
chore: add external integrations, API, SDK, and Dockerfile base
