+++ sdk_python/__init__.py
@@ -1,0 +1,4 @@
+﻿# GeneForge Python SDK v0.1
+
+def hello():
+    return 'Hello from GeneForge Python SDK!'
