+++ Dockerfile
@@ -1,0 +1,10 @@
+﻿FROM python:3.11-slim
+
+WORKDIR /app
+
+COPY . /app
+
+RUN pip install --upgrade pip && \
+    pip install -r requirements.txt
+
+CMD [\"python\", \"-m\", \"gf_core.cli\"]
