### 2025-06-07 2486384
docs: update README with roadmap phase 2
