+++ README.md
@@ -1,140 +1,10 @@
-# 🧬 GeneForgeLang: Symbolic-to-Sequence & Cross-Modality Biomolecular Design Toolkit
+﻿# GeneForgeLang
 
-**GeneForgeLang (GFL)** is a symbolic, generative language that allows scientists to **design, analyze and simulate** DNA, RNA, and protein structures with unified syntax and AI-compatible logic.
+This repository contains the core language features for GeneForge.
 
----
+## Roadmap Phase 2
 
-## 🌐 What It Enables
-
-- ✅ Symbolic design of gene therapies, edits, enhancers, and protein domains
-- 🔁 Cross-modality conversion: DNA ↔ RNA ↔ Protein
-- 🧠 Causal and conditional logic: `if A then B`, `EDIT(...)`, `EFFECT(...)`
-- 🧬 Multi-level modeling with time, dosage, pathway, and simulation
-- 📡 Integration with AI models for code generation and interpretation
-
----
-
-## 🚀 Core Features
-
-| Module                  | Description                                                             |
-|-------------------------|-------------------------------------------------------------------------|
-| 🧠 Phrase → Protein      | Generate realistic protein sequences from symbolic phrases              |
-| 🔁 Transcode Molecules   | Translate GeneForgeLang phrases between DNA, RNA, and Protein           |
-| 📖 Phrase → English      | Interpret symbolic code into natural scientific language                |
-| 🧬 Enhancer Generator    | Create regulatory sequences with TF logic and conditionals              |
-| 🔮 Simulate Edits        | Predict outcomes via `SIMULATE:` and `HYPOTHESIS:` logic                |
-| ⏳ Timeline & Dosing     | Represent `DOSE(n)` and `TIME(n):` blocks for therapeutic modeling      |
-| ⚙️ AI-Ready Interface     | Compatible with transformers like ProtGPT2, GeneForgeTransformer         |
-
----
-
-## 🧪 Example Input Phrases
-
-### DNA → RNA
-
-```gfl
-~d:[TATA]-ATG-[EX]-[IN]-[EX2]
-↓
-:r:Cap5'-Ex1-Ex2-UTR3'
-```
-
-### RNA → Protein
-
-```gfl
-:r:Ex1-Ex2
-↓
-^p:Dom(Kin)-Mot(NLS)*AcK@147
-```
-
----
-
-## 🔬 In Vivo Gene Editing Example (CRISPR 2.0)
-
-```gfl
-~d:[TATA]CPS1[MUT:PAT:A>G@1001]
-EDIT:Base(A→G@1001){efficacy=partial, cells=liver}
-DELIV(mRNA+LNP@IV)
-DOSE(1):EDIT:Base(A→G@1001)
-TIME(0d):DELIV(...)
-EFFECT(restore function=urea cycle)
-HYPOTHESIS: if MUT(Q335X) → Loss(CPS1)
-SIMULATE: {EDIT:Base(...), OUTCOME:↓ammonia}
-MACRO:FIX1 = {DELIV(...) - EDIT:Base(...)}
-USE:FIX1
-```
-
----
-
-## 🧬 Enhancer Design Example: ARE Activation
-
-```yaml
-enhancer:
-  name: "ARE_Synthetic_Enhancer"
-  target_gene: "NQO1"
-  cell_type: "hepatocyte"
-  species: "Homo sapiens"
-  factors:
-    - NFE2L2
-    - MAFG
-    - ATF4
-  goal: "activate"
-  model: "GeneForgeEnhancerGen-v1"
-```
-
-📂 See: `examples/enhancer_ARE_example.ipynb`
-
----
-
-## ▶️ How to Use this Example
-
[…]
