### 2025-06-07 1494e77
chore: add external_integrations folder and README
