+++ external_integrations/README.md
@@ -2,7 +2,6 @@
 
 This folder will contain integration modules for BioPython, CRISPOR, and Benchling.
 
-- BioPython: Sequence manipulation and analysis
-- CRISPOR: Guide RNA design
-- Benchling: Data management APIs
-
+- BioPython: Sequence manipulation and analysis  
+- CRISPOR: Guide RNA design  
+- Benchling: Data management APIs  
