### 2025-06-07 0dbb2f5
chore: add base Dockerfile
