### 2025-06-07 bd8b477
feat: add base Python SDK module and basic tests
