+++ gfl_sdk/__init__.py
@@ -1,0 +1,13 @@
+﻿# GeneForgeLang Python SDK base module
+
+def parse_phrase(phrase: str) -> dict:
+    '''
+    Dummy parser function - replace with actual implementation
+    '''
+    return {'input': phrase, 'parsed': True}
+
+def simulate_edit(edit: dict) -> dict:
+    '''
+    Dummy simulate function - extend for real simulation
+    '''
+    return {'edit': edit, 'result': 'simulated'}
