### 2025-06-07 bcb4100
test: add basic simulation function unit test
