### 2025-06-07 ad71041
chore: add external integrations folder and README
