### 2025-06-07 4070211
docs: add external integrations README and modules for CRISPOR and Benchling with tests
