+++ gfl_sdk/external_integrations/README.md
@@ -1,0 +1,11 @@
+﻿# README para integraciones externas en gfl_sdk
+This module provides basic integrations with external bioinformatics tools like CRISPOR and Benchling.
+
+## Usage
+Import functions from gfl_sdk.external_integrations:
+from gfl_sdk.external_integrations import get_crispor_guides, get_projects
+
+## Tests
+Run tests with:
+python -m unittest test_crispor.py
+python -m unittest test_benchling.py
