### 2025-06-07 81bdd08
feat: integrate CRISPOR and Benchling in SDK; extend tests and update README
