### 2025-06-07 dbc1520
docs: update README with SDK and external integrations; sync code and tests
