### 2025-06-07 32d6589
chore: add pending tests, grammar, parser, and integration folder; clean caches
