### 2025-06-07 4e11afe
feat: add enhancer and AAV vector modeling with basic simulation and tests
