### 2025-06-08 c7ef3f3
docs: add fundamental principle for LLM and GFL integration synergy
