+++ docs/LLM_GFL_Fundamental_Principle.md
@@ -1,0 +1,12 @@
+﻿# Fundamental Principle for LLM + GFL Integration
+
+The combination of Large Language Models (LLM) and GeneForgeLang (GFL) is not a mere hybridization or a simple translator between languages, but a profound synergy aimed at:
+
+- Exploiting and combining the unique expressive and reasoning capabilities of both systems.
+- Creating a cognitive power much greater than the sum of its parts, a "power squared" effect.
+- Building a robust cognitive foundation, not just a technical one, to enable a qualitative leap in problem-solving and knowledge generation in genetics.
+- Avoiding superficial translation or parallels; the goal is joint thinking and conceptual co-evolution.
+- Prioritizing deep understanding, abstraction, and complex reasoning capabilities that neither system could achieve independently.
+
+This principle will guide all future designs, proposals, and developments in the project, ensuring a groundbreaking approach to cognitive enhancement in genetic engineering.
+
