### 2025-06-08 bcebf84
feat: add roadmap phase 3 structure and placeholders
