+++ roadmap_phase_3/api/main.py
@@ -1,0 +1,14 @@
+﻿from fastapi import FastAPI
+from pydantic import BaseModel
+from typing import List, Dict
+
+app = FastAPI()
+
+class Edit(BaseModel):
+    type: str
+    details: Dict
+
+@app.post('/simulate')
+def simulate_edits(edits: List[Edit]):
+    # Placeholder simulation endpoint
+    return {'result': 'Simulation received', 'num_edits': len(edits)}
