+++ roadmap_phase_3/simulation_engine/dynamic_simulation.py
@@ -1,0 +1,8 @@
+﻿def run_dynamic_simulation(edit_sequence):
+    '''
+    Placeholder function for dynamic probabilistic simulation of gene edits and AAV effects.
+    Input: edit_sequence (list of dicts describing edits)
+    Output: dict with simulation results and probabilistic outcomes.
+    '''
+    results = {'simulated_edits': edit_sequence, 'outcome_probability': 0.75, 'notes': 'Simulation placeholder'}
+    return results
