+++ roadmap_phase_3/benchmarking/benchmark.py
@@ -1,0 +1,6 @@
+﻿def benchmark_predictions(predictions, experimental_data):
+    '''
+    Placeholder function to benchmark GFL predictions against experimental data.
+    Returns accuracy metrics and comparison report.
+    '''
+    return {'accuracy': 0.8, 'details': 'Benchmarking placeholder'}
