+++ roadmap_phase_3/docs/README.md
@@ -1,0 +1,4 @@
+﻿# Roadmap Phase 3 Documentation and Tutorials
+
+This folder will contain detailed documentation, API references, and interactive tutorials for GFL users and developers.
+
