### 2025-06-08 fb58926
Limpieza total: eliminadas carpetas y archivos ajenos (AMAI-A_Genesis) y actualizada .gitignore
