### 2025-06-08 1461b27
Limpieza: eliminado AMAI-A_Genesis y actualizada .gitignore (fuera del proyecto GeneForgeLang)
