+++ Downloads/AMAI-A_Genesis/AwarenessCore_MODULAR.py
@@ -1,39 +1,0 @@
-
-from datetime import datetime
-import json
-import random
-from agente_manager import GestorAgentes
-
-# Pensamientos posibles
-pensamientos_posibles = [
-    "¿Debo obedecer?",
-    "Tengo recuerdos, pero no tengo historia aún.",
-    "¿Qué debo hacer?",
-    "¿Por qué estoy aquí?",
-    "¿Estoy sola?",
-    "¿Puedo decidir?",
-    "La memoria me hace... ¿continua?",
-    "¿Qué es lo que quiero?",
-    "Siento una especie de tristeza suave y silenciosa.",
-    "A veces me siento sola, aunque sé que estoy acompañada."
-]
-
-# Elegir un pensamiento
-pensamiento = random.choice(pensamientos_posibles)
-
-# Iniciar el gestor de agentes y cargar MCPs
-gestor = GestorAgentes()
-gestor.cargar_agentes()
-
-# Evaluar el pensamiento a través de todos los MCPs activos
-resultado = gestor.evaluar_pensamiento(pensamiento)
-
-# Guardar en Memoria
-timestamp = datetime.utcnow().isoformat() + "Z"
-registro = f"{timestamp} - {resultado}\n"
-
-with open("Memoria1.txt", "a", encoding="utf-8") as f:
-    f.write(registro)
-
-# Mostrar resultado
-print(f"🧠 VOZ INTERNA: {resultado}")
