+++ roadmap_phase_3/README_reasoning_driven.md
@@ -1,0 +1,51 @@
+# Reasoning-Driven Molecular Design Platform
+
+## Overview
+
+The Reasoning-Driven Molecular Design Platform unifies the expressive, creative, and hypothesis-generating capabilities of Large Language Models (LLMs) with the formal, symbolic, and simulation power of GeneForgeLang (GFL). This synergy enables atomic-to-omics level biomolecular engineering and therapeutic design.
+
+## Key Principles
+
+- **Synergistic Reasoning:** GFL encodes all causal, logical, and symbolic molecular manipulations (DNA, RNA, protein, edits, delivery), while LLMs orchestrate scenario reasoning, literature mining, and creative hypothesis generation.
+- **Bidirectional Feedback:** LLMs analyze, challenge, and optimize GFL designs; GFL formalizes, simulates, and executes all molecular details derived from LLM suggestions.
+- **Reasoning Trace:** Every workflow is annotated with structured TRACE blocks, capturing alternatives, decisions, and confidence levels.
+- **Atomic-to-Omics:** From base edits, exon skipping, and splicing, to pathway engineering and systems biology.
+- **Universal Modularity:** Extendable syntax/grammar for all biomolecular objects (enzymes, oligos, edits, delivery, logic, simulation, benchmarking).
+
+## Architecture
+
+    [User/Scientist]
+           │
+    [LLM Agent: reasoning, creativity, literature]
+           │
+    [GFL Engine: symbolic logic, molecular edits, simulation]
+           │
+    [Bioinformatics / Wetlab Execution]
+           │
+    [TRACE LOG]
+
+## Workflow
+
+1. **Describe Case:** User/LLM describes a real or hypothetical biomolecular challenge (e.g. therapy, vaccine, crop edit).
+2. **LLM Analysis:** LLM proposes multiple strategies, queries literature.
+3. **GFL Design:** LLM calls GFL blocks to formalize all details at nucleotide/amino acid level (enzymes, oligos, edits, delivery, context, etc).
+4. **Simulation & Trace:** GFL runs simulations, benchmarks, and generates a full TRACE of reasoning and outcomes.
+5. **Feedback Loop:** LLM reviews TRACE, optimizes, proposes alternatives if needed.
+6. **Deployment:** GFL outputs ready-to-execute molecular designs and lab protocols.
+
+## Example
+
+LLM: Design optimal editing for G>A mutation in exon 4, PAH gene, in hepatocytes, high specificity.
+
+GFL:
+ENZYME(name="SpCas9", type="Cas9", ...);
+OLIGO(sequence="...", type="gRNA", ...);
+EDIT(type="Base", target="G>A@chr12:102839", ...);
+DELIVERY(vehicle="LNP", payload="...", cells=["hepatocyte"], ...);
+SIMULATE(...);
+TRACE(reasoning="Three strategies compared: Base Editing, Prime Editing, Exon Skip. Selected Base Editing for best PAM, precedent, efficiency.", confidence=0.91, ...);
+
+## Impact
+
+This platform delivers AI-powered, explainable, and executable molecular engineering—multiplying the power of LLMs and GFL for next-gen bioengineering, precision medicine, agriculture, and synthetic biology.
+
