### 2025-06-08 1271917
ADD: Extended GFL Syntax for Reasoning-Driven Molecular Design (syntax_extended.md)
