### 2025-06-08 5c9ed4b
ADD: GeneForgeLang v2 syntax formal specification (syntax_v2.md)
