### 2025-06-08 b05adec
ADD: GeneForgeLang v2 formal grammar definition (geneforge_grammar_v2.json) and updated syntax.md
