### 2025-06-09 a93d074
Update README: current project status and usage (English only)
