### 2025-06-09 cd503e5
Update GFL core: lexer, parser, reasoner, docs
