+++ gfl/lexer.py
@@ -1,48 +1,64 @@
-﻿import ply.lex as lex
+import ply.lex as lex
 
 tokens = (
-    'PRIME_EDIT',
-    'BASE_EDIT',
-    'PRIME_DEL',
-    'PEG',
-    'NICK_SITE',
-    'REVERSE_TRANSCRIPTASE',
-    'PIRNA',
-    'TRANSPOSON',
-    'ENDOGENOUS_RETROVIRUS',
-    'MITOCHONDRIAL_GENE',
-    'LPAREN',
-    'RPAREN',
+    'SIMULATE',
+    'EXPERIMENT',
+    'TYPE',
+    'TOOL',
+    'PARAMS',
+    'IDENTIFIER',
+    'NUMBER',
+    'LBRACE',
+    'RBRACE',
+    'COLON',
     'COMMA',
-    'EQUALS',
-    'ID',
-    'STRING',
-    'NUMBER'
+    'TRUE',
+    'FALSE'
 )
 
-t_LPAREN = r'\('
-t_RPAREN = r'\)'
+t_SIMULATE = r'simulate'
+t_EXPERIMENT = r'experiment'
+t_TYPE = r'type'
+t_TOOL = r'tool'
+t_PARAMS = r'params'
+
+t_LBRACE = r'\{'
+t_RBRACE = r'\}'
+t_COLON = r':'
 t_COMMA = r','
-t_EQUALS = r'='
-t_ignore = ' \t\r\n'
 
-def t_PRIME_EDIT(t): r'prime_edit'; return t
-def t_BASE_EDIT(t): r'base_edit'; return t
-def t_PRIME_DEL(t): r'prime_del'; return t
-def t_PEG(t): r'peg'; return t
-def t_NICK_SITE(t): r'nick_site'; return t
-def t_REVERSE_TRANSCRIPTASE(t): r'reverse_transcriptase'; return t
-def t_PIRNA(t): r'pirna'; return t
-def t_TRANSPOSON(t): r'transposon'; return t
-def t_ENDOGENOUS_RETROVIRUS(t): r'endogenous_retrovirus'; return t
-def t_MITOCHONDRIAL_GENE(t): r'mitochondrial_gene'; return t
-
-def t_ID(t):
-    r'[a-zA-Z_][a-zA-Z_0-9\-]*'
+def t_TRUE(t):
+    r'true'
+    t.value = 'true'
     return t
 
+def t_FALSE(t):
+    r'false'
+    t.value = 'false'
+    return t
+
+def t_NUMBER(t):
+    r'\d+\.\d+|\d+'
+    if '.' in t.value:
+        t.value = float(t.value)
+    else:
+        t.value = int(t.value)
+    return t
+
+def t_IDENTIFIER(t):
+    r'[A-Za-z_][A-Za-z_0-9]*'
+    if t.value in {'simulate', 'experiment', 'type', 'tool', 'params', 'true', 'false'}:
+        t.type = t.value.upper()
+    return t
+
+t_ignore = ' \t'
+
+def t_newline(t):
+    r'\n+'
+    t.lexer.lineno += len(t.value)
+
 def t_error(t):
-    print(f"Illegal character {t.value[0]!r}")
+    print(f"Carácter ilegal '{t.value[0]}'")
     t.lexer.skip(1)
 
 lexer = lex.lex()
