### 2025-06-09 dfccaa6
Robust DSL engine, adaptive logic, Scanpy integration, and documentation update (June 2025)
