+++ gfl/parsetab.py
@@ -6,9 +6,9 @@ _tabversion = '3.10'
 
 _lr_method = 'LALR'
 
-_lr_signature = 'statementsBASE_EDIT COMMA ENDOGENOUS_RETROVIRUS EQUALS ID LPAREN MITOCHONDRIAL_GENE NICK_SITE NUMBER PEG PIRNA PRIME_DEL PRIME_EDIT REVERSE_TRANSCRIPTASE RPAREN STRING TRANSPOSONstatements : statement\n| statement statementsstatement : prime_edit\n| base_edit\n| prime_del\n| peg\n| nick_site\n| reverse_transcriptase\n| pirna\n| transposon\n| endogenous_retrovirus\n| mitochondrial_geneprime_edit : PRIME_EDIT LPAREN arguments RPARENbase_edit : BASE_EDIT LPAREN arguments RPARENprime_del : PRIME_DEL LPAREN arguments RPARENpeg : PEG LPAREN arguments RPARENnick_site : NICK_SITE LPAREN arguments RPARENreverse_transcriptase : REVERSE_TRANSCRIPTASE LPAREN arguments RPARENpirna : PIRNA LPAREN arguments RPARENtransposon : TRANSPOSON LPAREN arguments RPARENendogenous_retrovirus : ENDOGENOUS_RETROVIRUS LPAREN arguments RPARENmitochondrial_gene : MITOCHONDRIAL_GENE LPAREN arguments RPARENarguments : argument\n| argument COMMA argumentsargument : ID EQUALS ID'
+_lr_signature = 'COLON COMMA EXPERIMENT FALSE IDENTIFIER LBRACE NUMBER PARAMS RBRACE SIMULATE TOOL TRUE TYPEprogram : statementsstatements : statements statementstatements : statementstatement : SIMULATE IDENTIFIERstatement : EXPERIMENT LBRACE experiment_body RBRACEexperiment_body : experiment_fieldsexperiment_fields : experiment_fields experiment_fieldexperiment_fields : experiment_fieldexperiment_field : TYPE COLON IDENTIFIERexperiment_field : TOOL COLON IDENTIFIERexperiment_field : PARAMS COLON param_dictparam_dict : LBRACE param_pairs RBRACEparam_pairs : param_pairs COMMA param_pairparam_pairs : param_pairparam_pair : IDENTIFIER COLON param_valueparam_value : TRUE\n                   | FALSEparam_value : NUMBERparam_value : IDENTIFIER'
     
[…]
