+++ gfl/__init__.py
@@ -1,4 +1,0 @@
-
-from .lexer import lexer
-
-from .lexer import lexer
