+++ gfl/execution/scanpy_interface.py
@@ -1,0 +1,33 @@
+import json
+
+def run_scrna_experiment(params):
+    """
+    Simula la ejecución de un experimento de scRNA-seq con Scanpy.
+    """
+    print(f"\n[SCANPY INTERFACE] Simulando experimento scRNA-seq con parámetros: {params}")
+    # Simulación de resultados
+    simulated_results = {
+        "analysis_type": "scRNA_seq",
+        "parameters_used": params,
+        "metrics": {
+            "n_cells": 10000,
+            "n_genes": 20000,
+            "mean_reads_per_cell": 50000,
+            "dropout_rate": 0.25, # Tasa de dropout inicial
+            "pca_components": params.get('pca_dims', 50)
+        },
+        "clusters": [
+            {"name": "Cluster 1", "n_cells": 3000, "marker_genes": ["GeneA", "GeneB"]},
+            {"name": "Cluster 2", "n_cells": 2500, "marker_genes": ["GeneC", "GeneD"]}
+        ],
+        "recommendations": [] # Aquí podemos añadir recomendaciones futuras basadas en el análisis
+    }
+    print("[SCANPY INTERFACE] Resultados simulados generados.")
+    return simulated_results
+
+if __name__ == '__main__':
+    # Ejemplo de uso directo para pruebas
+    test_params = {'normalize': True, 'pca_dims': 50}
+    results = run_scrna_experiment(test_params)
+    print("Resultados de la simulación:")
+    print(json.dumps(results, indent=2))
