### 2025-06-09 0db8df2
Consolidated: robust engine, adaptive logic, Scanpy integration, and improved documentation (June 2025)
