### 2025-06-12 4c6f9ac
Docs: Overwrote README with complete English content, including improved messages and roadmap.
