+++ README.md
@@ -1,190 +1,70 @@
-# 🧬 GeneForgeLang: Symbolic-to-Sequence & Cross-Modality Biomolecular Design Toolkit
+# GeneForgeLang: A Domain-Specific Language for Computational Biology
 
-**GeneForgeLang (GFL)** is a symbolic, generative language that allows scientists to **design, analyze and simulate** DNA, RNA, and protein structures with unified syntax and AI-compatible logic.
+## Overview
+GeneForgeLang is a Domain-Specific Language (DSL) designed for biological researchers to describe complex data analysis and simulation workflows in an intuitive, readable, and reproducible manner. Our goal is to simplify interaction with bioinformatics tools and simulation engines, allowing users to focus on biology rather than complex syntax or pipeline management.
 
----
+## Key Features
+* **Intuitive Syntax:** Language designed to be close to the natural language of the biological domain.
+* **Robust Lexical and Syntactic Analysis:** Processing GFL code into an Abstract Syntax Tree (AST).
+* **Basic Workflow Evaluation:** Simulation of the execution of `analyze`, `experiment`, `simulate`, and `branch` operations.
+* **AI-Generated GFL Post-processing:** Ability to standardize and validate GFL code originating from language models.
+* **Enhanced Error and Warning Messages:** Clear, informative, and user-friendly feedback with specific suggestions.
 
-## 🌐 What It Enables
+## Roadmap (Next Steps)
+Currently, the project is in **Phase 0 (Foundations and Proof of Concept)**. The next steps include:
+* **Phase 1: Language Enrichment and Semantic Analysis:** Implementation of type validations, valid values, and contextual parameter compatibility.
+* **Phase 2: Integration with Real Bioinformatics Tools:** Development of wrappers for tools like DESeq2, Scanpy, and connection with simulation engines.
+* **Phase 3: Optimization and Development Tools:** Performance improvements, robust CLI, comprehensive documentation, and IDE support.
 
-- ✅ Symbolic design of gene therapies, edits, enhancers, and protein domains
-- 🔁 Cross-modality conversion: DNA ↔ RNA ↔ Protein
-- 🧠 Causal and conditional logic: `if A then B`, `EDIT(...)`, `EFFECT(...)`
-- 🧬 Multi-level modeling with time, dosage, pathway, and simulation
-- 📡 Integration with AI models for code generation and interpretation
+## Installation
+# * (Add your installation instructions here, e.g.:)
+# 1. Clone the repository: `git clone <REPO_URL>`
+# 2. Navigate into the project directory: `cd GeneForgeLang`
+# 3. Create and activate a virtual environment:
+#    ```bash
+#    python3 -m venv venv
+#    source venv/bin/activate
+#    ```
+# 4. Install dependencies (if you have a `requirements.txt`):
+#    ```bash
+#    pip install -r requirements.txt
+#    # Or install ply directly if it's the only one:
+#    pip install ply
+#    ```
 
----
+## Usage
+# * (Add examples of how to run the parser, evaluator, or demo here, e.g.:)
+# To run the current demo:
+# ```bash
+# (venv) user@DESKTOP-MKJRNF0:~/GeneForgeLang$ PYTHONPATH=. python3 scripts/fix_and_demo.py
+# ```
+# To test the parser with a GFL file:
+# ```bash
+# (venv) user@DESKTOP-MKJRNF0:~/GeneForgeLang$ PYTHONPATH=. python3 gfl/parser.py gfl/gfl_example.gfl
+# ```
 
-## 🚀 Core Features
-
-| Module                  | Description                                                             |
[…]
