### 2025-06-12 158d7f0
Fix: Addressed identified security vulnerabilities by updating dependencies as per Dependabot.
