### 2025-06-12 f05de51
Refactor: Comprehensive cleanup and git re-indexing. Added new training data and scripts. Final attempt to resolve Dependabot warnings.
