+++ scripts/fix_and_demo.py
@@ -1,66 +1,356 @@
-import sys
 import os
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import sys
+import logging
+import inspect
+import re
 
-from gfl.lexer import lexer
-from gfl.parser_rules import parser
-from gfl.adaptive_reasoner import apply_adaptive_rules
-from gfl.execution.experiment_runner import run_experiment
-from pprint import pprint
+# Configuración básica del logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logger = logging.getLogger(__name__)
 
-def feedback_cycle(ast, max_cycles=3):
-    cycle = 0
-    ast_current = ast
-    while cycle < max_cycles:
-        print(f"\n== CICLO DE FEEDBACK #{cycle+1} ==")
-        new_ast = []
-        feedback_applied = False
-        idx = 0
-        while idx < len(ast_current):
-            node = ast_current[idx]
-            if (
-                node.get("node_type") == "experiment"
-                and not node.get("fixed")
-                and not node.get("params", {}).get("imputation")
-            ):
-                print(f"  > Ejecutando experimento '{node.get('type')}' con {node.get('params')}")
-                result = run_experiment(node)
-                print(f"    ...resultado: {result}")
-                if result["status"] == "bad_quality" and not feedback_applied:
-                    imput_node = {
-                        "node_type": "experiment",
-                        "tool": node.get("tool", "scanpy"),
-                        "type": node.get("type", "scRNA"),
-                        "params": {"imputation": True},
-                        "fixed": True
-                    }
-                    node = dict(node)
-                    node["fixed"] = True
-                    new_ast.append(node)
-                    new_ast.append(imput_node)
-                    print("    ...[IMPUTACIÓN INYECTADA EN PIPELINE]")
-                    feedback_applied = True
-                    idx += 1
-                    continue
-            new_ast.append(node)
-            idx += 1
-        if not feedback_applied:
-            print("  ✔️  No se requiere más feedback adaptativo.")
-            break
-        ast_current = new_ast
-        cycle += 1
-    return ast_current
+# Asegura que el directorio padre del paquete 'gfl' esté en sys.path.
+current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+parent_dir = os.path.dirname(current_dir)
+if parent_dir not in sys.path:
+    sys.path.insert(0, parent_dir)
+
+from gfl.parser import parse_gfl_code
+
+def post_process_gfl_output(gfl_code):
+    """
+    Realiza un post-procesamiento básico para corregir errores comunes de formato
+    en la salida GFL generada por la IA, antes de pasarla al parser.
+    """
+    lines = gfl_code.splitlines()
+    processed_lines = []
+    in_block = 0  # Contador para detectar la profundidad del bloque {}
+    
+    # Patrón para identificar una propiedad: IDENTIFIER: value
+    property_pattern = re.compile(r'^\s*([a-zA-Z_][a-zA-Z_0-9]*)\s*:\s*(.*)')
+
+    for i, line in enumerate(lines):
+        stripped_line = line.strip()
+        original_line_indent = len(line) - len(line.lstrip())
+
+        # Eliminar comas extra al inicio de una línea o después de ':'
+        stripped_line = re.sub(r'^\s*,\s*', '', stripped_line) # Eliminar coma al inicio
+        stripped_line = re.sub(r':\s*,', ':', stripped_line) # Eliminar coma después de ':'
+
+        # Normalizar True/False sin comillas si la IA las añade
+        stripped_line = stripped_line.replace('"true"', 'true').replace('"false"', 'false')
+        
+        # Ajustar espacios después de ':' y ','
[…]
