+++ gfl/parser_rules.py
@@ -1,90 +1,124 @@
 import ply.yacc as yacc
-from gfl.lexer import tokens
+import logging
+from gfl.lexer import GFLLexer # Importación necesaria para acceder a los tokens
 
-def p_program(p):
-    'program : statements'
-    p[0] = p[1]
+class GFLParser:
+    # Define los tokens que este parser reconocerá.
+    tokens = GFLLexer.tokens
 
-def p_statements_multiple(p):
-    'statements : statements statement'
-    p[0] = p[1] + [p[2]]
+    # Define el símbolo de inicio de tu gramática GFL.
+    start = 'gfl_code'
 
-def p_statements_single(p):
-    'statements : statement'
-    p[0] = [p[1]]
+    # Importante: No hay un método __init__ aquí que cree lexer o parser de PLY.
+    # Esos se crean centralmente en gfl/parser.py.
 
-def p_statement_simulate(p):
-    'statement : SIMULATE IDENTIFIER'
-    p[0] = {'node_type': 'simulate', 'type': 'simulate', 'target': p[2]}
+    # --- Reglas de Producción de la Gramática GFL ---
 
-def p_statement_experiment_block(p):
-    'statement : EXPERIMENT LBRACE experiment_body RBRACE'
-    body = p[3]
-    body['node_type'] = 'experiment'
-    p[0] = body
+    # Regla principal: un programa GFL es una secuencia de sentencias.
+    def p_gfl_code(self, p):
+        '''gfl_code : statements'''
+        p[0] = p[1]
 
-def p_experiment_body(p):
-    'experiment_body : experiment_fields'
-    exp = {}
-    for d in p[1]:
-        exp.update(d)
-    p[0] = exp
+    # Una secuencia de sentencias.
+    def p_statements(self, p):
+        '''statements : statement
+                      | statements statement'''
+        if len(p) == 2:
+            p[0] = [p[1]]
+        else:
+            p[0] = p[1] + [p[2]]
 
-def p_experiment_fields_multiple(p):
-    'experiment_fields : experiment_fields experiment_field'
-    p[0] = p[1] + [p[2]]
+    # Una sentencia puede ser de varios tipos.
+    def p_statement(self, p):
+        '''statement : analyze_statement
+                     | experiment_statement
+                     | simulate_statement
+                     | branch_statement'''
+        p[0] = p[1]
 
-def p_experiment_fields_single(p):
-    'experiment_fields : experiment_field'
-    p[0] = [p[1]]
+    # Sentencia ANALYZE
+    def p_analyze_statement(self, p):
+        '''analyze_statement : ANALYZE LBRACE properties RBRACE'''
+        p[0] = {'node_type': 'analyze', 'properties': p[3]}
 
-def p_experiment_field_type(p):
-    'experiment_field : TYPE COLON IDENTIFIER'
-    p[0] = {'type': p[3]}
+    # Sentencia EXPERIMENT
+    def p_experiment_statement(self, p):
+        '''experiment_statement : EXPERIMENT LBRACE properties RBRACE'''
+        p[0] = {'node_type': 'experiment', 'properties': p[3]}
 
-def p_experiment_field_tool(p):
-    'experiment_field : TOOL COLON IDENTIFIER'
-    p[0] = {'tool': p[3]}
+    # Sentencia SIMULATE
+    def p_simulate_statement(self, p):
+        '''simulate_statement : SIMULATE IDENTIFIER
+                              | SIMULATE IDENTIFIER properties'''
+        if len(p) == 3:
+            p[0] = {'node_type': 'simulate', 'target': p[2]}
+        else:
+            p[0] = {'node_type': 'simulate', 'target': p[2], 'properties': p[3]}
 
-def p_experiment_field_params(p):
[…]

[…]
