### 2025-06-12 b7b5e38
Feat: Added gfl/validation_registry.py for semantic validation of simulation targets, tools, and experiment types.
