### 2025-06-12 e8a4d33
Feat: Implemented semantic validation for simulation targets, analysis tools, experiment types, and analysis strategies in gfl/evaluator.py.
