+++ gfl/evaluator.py
@@ -1,0 +1,228 @@
+# gfl/evaluator.py
+
+import sys
+from collections import deque
+from .validation_registry import VALID_SIMULATION_TARGETS, VALID_ANALYSIS_TOOLS, VALID_EXPERIMENT_TYPES, VALID_ANALYSIS_STRATEGIES
+
+class Evaluator:
+    def __init__(self):
+        # Aquí puedes inicializar el estado de la simulación o análisis
+        self.simulation_state = {}
+        self.analysis_results = []
+        self.output_buffer = deque() # Usamos una cola para mensajes de output/error
+
+    def _log_error(self, message, line=None, column=None, suggestion=None):
+        """Genera un mensaje de error formateado y lo añade al buffer."""
+        location = ""
+        if line is not None:
+            location += f" (línea {line}"
+            if column is not None:
+                location += f", columna {column}"
+            location += ")"
+        
+        full_message = f"ERROR: {message}{location}."
+        if suggestion:
+            full_message += f" ¿Quizás quiso decir: {suggestion}?"
+        self.output_buffer.append(full_message)
+        # Aquí podrías decidir si el error es fatal y detener la ejecución
+        # raise ValueError(full_message) # Para detener la ejecución inmediatamente
+
+    def _log_warning(self, message, line=None, column=None, suggestion=None):
+        """Genera un mensaje de advertencia formateado y lo añade al buffer."""
+        location = ""
+        if line is not None:
+            location += f" (línea {line}"
+            if column is not None:
+                location += f", columna {column}"
+            location += ")"
+        
+        full_message = f"ADVERTENCIA: {message}{location}."
+        if suggestion:
+            full_message += f" {suggestion}"
+        self.output_buffer.append(full_message)
+
+    def _validate_target(self, target_name, node_info):
+        """Valida que el objetivo de simulación sea reconocido."""
+        if target_name not in VALID_SIMULATION_TARGETS:
+            valid_targets_str = ", ".join(VALID_SIMULATION_TARGETS.keys())
+            suggestion = None
+            if "death" in target_name.lower():
+                suggestion = "'apoptosis'"
+            elif "growth" in target_name.lower():
+                suggestion = "'cell_growth'"
+            elif "division" in target_name.lower():
+                suggestion = "'cell_division'"
+            
+            self._log_error(
+                f"El objetivo de simulación '{target_name}' no es reconocido. Los objetivos válidos son: {valid_targets_str}",
+                line=node_info.get('line'),
+                column=node_info.get('column'),
+                suggestion=suggestion
+            )
+            return False
+        return True
+
+    def _validate_tool(self, tool_name, node_info):
+        """Valida que la herramienta de análisis sea reconocida."""
+        if tool_name not in VALID_ANALYSIS_TOOLS:
+            valid_tools_str = ", ".join(VALID_ANALYSIS_TOOLS.keys())
+            suggestion = None
+            if "deseq" in tool_name.lower():
+                suggestion = "'DESeq2'"
+            elif "scan" in tool_name.lower():
+                suggestion = "'Scanpy'"
+
+            self._log_error(
+                f"La herramienta de análisis '{tool_name}' no es reconocida. Las herramientas válidas son: {valid_tools_str}",
+                line=node_info.get('line'),
+                column=node_info.get('column'),
+                suggestion=suggestion
+            )
+            return False
+        return True
+
+    def _validate_experiment_type(self, exp_type, node_info):
+        """Valida que el tipo de experimento sea reconocido."""
+        if exp_type not in VALID_EXPERIMENT_TYPES:
+            valid_types_str = ", ".join(VALID_EXPERIMENT_TYPES)
+            
+            self._log_error(
[…]
