### 2025-06-12 bd835cc
Feat: Implemented detailed parameter validation in gfl/evaluator.py using expanded VALID_PARAMS_BY_TOOL_STRATEGY in gfl/validation_registry.py.
