+++ gfl/evaluator.py
@@ -2,17 +2,18 @@
 
 import sys
 from collections import deque
-from .validation_registry import VALID_SIMULATION_TARGETS, VALID_ANALYSIS_TOOLS, VALID_EXPERIMENT_TYPES, VALID_ANALYSIS_STRATEGIES
+from .validation_registry import (
+    VALID_SIMULATION_TARGETS, VALID_ANALYSIS_TOOLS, VALID_EXPERIMENT_TYPES,
+    VALID_ANALYSIS_STRATEGIES, VALID_PARAMS_BY_TOOL_STRATEGY
+)
 
 class Evaluator:
     def __init__(self):
-        # Aquí puedes inicializar el estado de la simulación o análisis
         self.simulation_state = {}
         self.analysis_results = []
-        self.output_buffer = deque() # Usamos una cola para mensajes de output/error
+        self.output_buffer = deque()
 
     def _log_error(self, message, line=None, column=None, suggestion=None):
-        """Genera un mensaje de error formateado y lo añade al buffer."""
         location = ""
         if line is not None:
             location += f" (línea {line}"
@@ -24,11 +25,8 @@ class Evaluator:
         if suggestion:
             full_message += f" ¿Quizás quiso decir: {suggestion}?"
         self.output_buffer.append(full_message)
-        # Aquí podrías decidir si el error es fatal y detener la ejecución
-        # raise ValueError(full_message) # Para detener la ejecución inmediatamente
 
     def _log_warning(self, message, line=None, column=None, suggestion=None):
-        """Genera un mensaje de advertencia formateado y lo añade al buffer."""
         location = ""
         if line is not None:
             location += f" (línea {line}"
@@ -42,7 +40,6 @@ class Evaluator:
         self.output_buffer.append(full_message)
 
     def _validate_target(self, target_name, node_info):
-        """Valida que el objetivo de simulación sea reconocido."""
         if target_name not in VALID_SIMULATION_TARGETS:
             valid_targets_str = ", ".join(VALID_SIMULATION_TARGETS.keys())
             suggestion = None
@@ -63,7 +60,6 @@ class Evaluator:
         return True
 
     def _validate_tool(self, tool_name, node_info):
-        """Valida que la herramienta de análisis sea reconocida."""
         if tool_name not in VALID_ANALYSIS_TOOLS:
             valid_tools_str = ", ".join(VALID_ANALYSIS_TOOLS.keys())
             suggestion = None
@@ -82,7 +78,6 @@ class Evaluator:
         return True
 
     def _validate_experiment_type(self, exp_type, node_info):
-        """Valida que el tipo de experimento sea reconocido."""
         if exp_type not in VALID_EXPERIMENT_TYPES:
             valid_types_str = ", ".join(VALID_EXPERIMENT_TYPES)
             
@@ -95,7 +90,6 @@ class Evaluator:
         return True
 
     def _validate_analysis_strategy(self, strategy_name, node_info):
-        """Valida que la estrategia de análisis sea reconocida."""
         if strategy_name not in VALID_ANALYSIS_STRATEGIES:
             valid_strategies_str = ", ".join(VALID_ANALYSIS_STRATEGIES)
             self._log_error(
@@ -105,14 +99,53 @@ class Evaluator:
             )
             return False
         return True
+    
+    def _validate_analysis_params(self, tool, strategy, params, node_info):
+        """
+        Valida que los parámetros proporcionados sean válidos para la herramienta y estrategia dadas.
+        """
+        tool_params = VALID_PARAMS_BY_TOOL_STRATEGY.get(tool)
+        if not tool_params:
+            # Si la herramienta no tiene parámetros definidos en el registro,
[…]
