### 2025-06-12 6787cb3
Fix: Configured lexer to ignore line comments (#) and removed LINE_COMMENT from tokens list to resolve parsing errors.
