+++ README.md
@@ -1,4 +1,52 @@
-GeneForgeLang: A Domain-Specific Language for Bio-simulation WorkflowsGeneForgeLang (GFL) is a custom Domain-Specific Language (DSL) designed to simplify and automate complex bio-simulation and analysis workflows. It provides a human-readable syntax to describe multi-step processes involving experimental design, data analysis, and biological simulations.Table of ContentsFeaturesProject StructureGetting StartedPrerequisitesSetupRunning the DemoGFL Language SyntaxAnalyze StatementExperiment StatementSimulate StatementBranch StatementCommentsError Handling and ValidationRoadmap (Next Steps)ContributingLicenseContactFeaturesGeneForgeLang's compiler (lexer and parser) is built using PLY (Python Lex-Yacc) and features:Intuitive Syntax: Language designed to be close to the natural language of the biological domain.Robust Lexical Analysis (Lexer):Tokenizes GFL source code, breaking it down into meaningful units (keywords, identifiers, literals, operators, etc.).Handles and ignores whitespace.Dedicated Pre-processor:A component (gfl/preprocessor.py) that specifically removes both single-line (//, #) and multi-line (/* ... */) comments from the GFL code before it is passed to the lexer. This simplifies the lexer's responsibilities and maintains a clean token stream.Syntactic Analysis (Parser):Constructs an Abstract Syntax Tree (AST) from the token stream, rigorously validating the code against the defined GFL grammar.Supports complex nested structures for analyze, experiment, and branch statements.Basic Workflow Evaluation & Validation: The parser implicitly performs initial workflow evaluation by logging actions and implements basic validation rules for parameters and simulation targets (e.g., checking for valid simulation targets or unknown parameters for specific tools). It also flags some contextual parameter compatibility issues.Structured Output: The parser generates a list of operations (representing the AST) that can then be interpreted or executed by a backend system.Enhanced Error and Warning Messages: Provides clear, informative, and user-friendly feedback with specific suggestions, aiding in debugging GFL code.AI-Generated GFL Post-processing (Implicit): The compiler's design supports the standardization and validation of GFL code, making it suitable for integration with language models that generate GFL.Project StructureGeneForgeLang/
+# GeneForgeLang: A Domain-Specific Language for Bio-simulation Workflows
+
+GeneForgeLang (GFL) is a custom Domain-Specific Language (DSL) designed to simplify and automate complex bio-simulation and analysis workflows. It provides a human-readable syntax to describe multi-step processes involving experimental design, data analysis, and biological simulations.
+
+## Table of Contents
+* [Features](#features)
+* [Project Structure](#project-structure)
+* [Getting Started](#getting-started)
+  * [Prerequisites](#prerequisites)
+  * [Setup](#setup)
+  * [Running the Demo](#running-the-demo)
+* [GFL Language Syntax](#gfl-language-syntax)
+  * [Analyze Statement](#analyze-statement)
+  * [Experiment Statement](#experiment-statement)
+  * [Simulate Statement](#simulate-statement)
+  * [Branch Statement](#branch-statement)
+  * [Comments](#comments)
+* [Error Handling and Validation](#error-handling-and-validation)
+* [Roadmap (Next Steps)](#roadmap-next-steps)
+* [Contributing](#contributing)
+* [License](#license)
+* [Contact](#contact)
+
+---
+
+## Features
+
+GeneForgeLang's compiler (lexer and parser) is built using **PLY (Python Lex-Yacc)** and features:
+
+* **Intuitive Syntax:** Language designed to be close to the natural language of the biological domain.
+* **Robust Lexical Analysis (Lexer):**
[…]
