### 2025-07-05 7cba320
Feature: Initial interpreter setup and parser fixes; Added new plugin scaffolding
