+++ gfl/parser.py
@@ -1,328 +1,208 @@
-import ply.lex as lex
 import ply.yacc as yacc
+from lexer import tokens
 import logging
-import os
-import sys
 
-# Configuración del logger
-logging.basicConfig(level=logging.INFO,
-                    format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
 
-# Desactivar mensajes de depuración de PLY si no estamos depurando
-if not os.environ.get('GFL_DEBUG_PARSER'):
-    logging.getLogger('ply.lex').setLevel(logging.WARNING)
-    logging.getLogger('ply.yacc').setLevel(logging.WARNING)
-else:
-    logging.getLogger('ply.lex').setLevel(logging.INFO)
-    logging.getLogger('ply.yacc').setLevel(logging.INFO)
-# --- LEXER ---
-
-# Lista de nombres de tokens
-tokens = (
-    'ANALYZE', 'EXPERIMENT', 'SIMULATE', 'BRANCH',
-    'IF', 'THEN', 'ELSE',
-    'COLON', 'LBRACE', 'RBRACE',
-    'STRING', 'NUMBER', 'IDENTIFIER',
-    'TRUE', 'FALSE',
-    'EQUALS', 'NOT_EQUALS', 'LT', 'GT', 'LTE', 'GTE',
-    'AND', 'OR', 'LPAREN', 'RPAREN', 'COMMA',
-    'COMMENT', 'LINE_COMMENT'
-)
-
-# Palabras reservadas (keywords)
-reserved = {
-    'analyze': 'ANALYZE',
-    'experiment': 'EXPERIMENT',
-    'simulate': 'SIMULATE',
-    'branch': 'BRANCH',
-    'if': 'IF',
-    'then': 'THEN',
-    'else': 'ELSE',
-    'true': 'TRUE',
-    'false': 'FALSE',
-    'AND': 'AND', # Operador lógico AND
-    'OR': 'OR',   # Operador lógico OR
-}
-
-# Expresiones regulares para tokens simples
-t_COLON = r':'
-t_LBRACE = r'{'
-t_RBRACE = r'}'
-t_COMMA = r','
-t_LPAREN = r'\('
-t_RPAREN = r'\)'
-
-# Operadores de comparación
-t_EQUALS = r'=='
-t_NOT_EQUALS = r'!='
-t_LTE = r'<='
-t_GTE = r'>='
-t_LT = r'<'
-t_GT = r'>'
-
-# Un token para identificadores y palabras reservadas
-def t_IDENTIFIER(t):
-    r'[a-zA-Z_][a-zA-Z_0-9]*'
-    t.type = reserved.get(t.value, 'IDENTIFIER')  # Check for reserved words
-    return t
-
-# Cadenas (strings)
-t_STRING = r'\"([^\\\n]|(\\.))*?\"|\'([^\\\n]|(\\.))*?\''
-
-# Números (integers y floats)
-def t_NUMBER(t):
-    r'\d+(\.\d*)?([eE][+-]?\d+)?'
-    t.value = float(t.value) if '.' in t.value or 'e' in t.value or 'E' in t.value else int(t.value)
-    return t
-
-# Comentarios de una sola línea (// o #)
-t_LINE_COMMENT = r'//.*|\#.*'
-
-# Comentarios de varias líneas (/* ... */)
-t_COMMENT = r'/\*[^*]*\*+(?:[^/*][^*]*\*+)*/'
-
-# Ignorar espacios y tabulaciones
-t_ignore = ' \t'
-
-# Contar saltos de línea
-def t_newline(t):
-    r'\n+'
-    t.lexer.lineno += len(t.value)
-
-# Manejo de errores de lexer
-def t_error(t):
[…]
