+++ gfl/interpreter.py
@@ -1,0 +1,209 @@
+import logging
+from gfl.plugins import plugin_registry # Importa el registro de plugins
+
+logger = logging.getLogger(__name__)
+
+class Interpreter:
+    def __init__(self):
+        self.symbol_table = {} # Para almacenar variables y sus valores
+
+    def interpret(self, ast):
+        """
+        Interpreta el Abstract Syntax Tree (AST).
+        """
+        self.visit(ast)
+
+    def visit(self, node):
+        """
+        Método genérico para visitar nodos del AST.
+        """
+        if not isinstance(node, dict) or 'type' not in node:
+            logger.warning(f"Invalid AST node encountered: {node}")
+            return
+
+        method_name = f"visit_{node['type']}"
+        visitor_method = getattr(self, method_name, self.generic_visit)
+        return visitor_method(node)
+
+    def generic_visit(self, node):
+        """
+        Visita recursivamente los hijos de un nodo.
+        """
+        for key, value in node.items():
+            if isinstance(value, dict) and 'type' in value:
+                self.visit(value)
+            elif isinstance(value, list):
+                for item in value:
+                    if isinstance(item, dict) and 'type' in item:
+                        self.visit(item)
+
+    def visit_program(self, node):
+        """
+        Visita el nodo raíz del programa.
+        """
+        for statement in node.get('body', []):
+            self.visit(statement)
+
+    def visit_define_statement(self, node):
+        """
+        Interpreta una declaración DEFINE.
+        """
+        var_name = node['name']
+        value = self.evaluate_expression(node.get('value'))
+        self.symbol_table[var_name] = value
+        logger.info(f"DEFINED: {var_name} = {value}")
+
+    def visit_invoke_statement(self, node):
+        """
+        Interpreta una declaración INVOKE.
+        """
+        plugin_name = node['plugin']
+        method_name = node['method']
+        params = {}
+        for param_name, param_node in node.get('params', {}).items():
+            params[param_name] = self.evaluate_expression(param_node)
+
+        as_var = node.get('as_var')
+
+        try:
+            plugin = plugin_registry.get(plugin_name)
+            result = plugin.execute(method_name, params, self.symbol_table)
+            if as_var:
+                self.symbol_table[as_var] = result
+                logger.info(f"INVOKED: {plugin_name}.{method_name} (result stored in {as_var})")
+            else:
+                logger.info(f"INVOKED: {plugin_name}.{method_name}")
+            return result # Devuelve el resultado por si se necesita encadenar llamadas
+        except ValueError as e:
+            logger.error(f"Plugin Error: {e}")
+            raise # Re-lanza para que el TRY/CATCH de GFL pueda capturarlo
+        except NotImplementedError as e:
+            logger.error(f"Plugin method not implemented: {e}")
+            raise
+        except Exception as e:
+            logger.error(f"Error executing plugin {plugin_name}.{method_name}: {e}", exc_info=True)
+            raise
+
+    def visit_message_statement(self, node):
+        """
+        Interpreta una declaración MESSAGE (actualmente solo para mensajes simples).
+        """
+        message_value = self.evaluate_expression(node['message'])
+        logger.info(f"MESSAGE: {message_value}")
+
+    def visit_branch_statement(self, node):
+        """
+        Interpreta una declaración BRANCH (IF/THEN/ELSE).
+        """
+        condition_result = self.evaluate_expression(node['condition'])
+        if condition_result:
+            logger.info("BRANCH: Condition true, executing THEN branch.")
+            self.visit(node['then_branch'])
+        elif 'else_branch' in node:
+            logger.info("BRANCH: Condition false, executing ELSE branch.")
+            self.visit(node['else_branch'])
[…]

[…]
