### 2025-07-05 bd3e9b4
Fix: Addressed Dependabot alerts by upgrading urllib3 and torch
