+++ gfl/requirements.txt
@@ -1,0 +1,31 @@
+certifi==2025.6.15
+charset-normalizer==3.4.2
+filelock==3.18.0
+fsspec==2025.5.1
+idna==3.10
+Jinja2==3.1.6
+MarkupSafe==3.0.2
+mpmath==1.3.0
+networkx==3.5
+nvidia-cublas-cu12==12.6.4.1
+nvidia-cuda-cupti-cu12==12.6.80
+nvidia-cuda-nvrtc-cu12==12.6.77
+nvidia-cuda-runtime-cu12==12.6.77
+nvidia-cudnn-cu12==9.5.1.17
+nvidia-cufft-cu12==11.3.0.4
+nvidia-cufile-cu12==1.11.1.6
+nvidia-curand-cu12==10.3.7.77
+nvidia-cusolver-cu12==11.7.1.2
+nvidia-cusparse-cu12==12.5.4.2
+nvidia-cusparselt-cu12==0.6.3
+nvidia-nccl-cu12==2.26.2
+nvidia-nvjitlink-cu12==12.6.85
+nvidia-nvtx-cu12==12.6.77
+ply==3.11
+requests==2.32.4
+setuptools==80.9.0
+sympy==1.14.0
+torch==2.7.1
+triton==3.3.1
+typing_extensions==4.14.1
+urllib3==2.5.0
