### 2025-07-06 82f105a
Add initial README.md for GeneForgeLang project
