+++ alphagenome_plugin.py
@@ -1,0 +1,51 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+class AlphaGenomePlugin:
+    def execute(self, method_name: str, params: dict, symbol_table: dict):
+        """
+        Simula la ejecución de métodos de AlphaGenome.
+        Cuando haya una API real, este método se conectará a ella.
+        """
+        logger.info(f"AlphaGenomePlugin: Simulating method '{method_name}' with params: {params}")
+
+        if method_name == "predict_functional_tracks":
+            sequence = params.get("sequence")
+            tracks = params.get("tracks", [])
+            variants = params.get("variants", [])
+
+            # Validación básica de parámetros
+            if not sequence:
+                raise ValueError("Parameter 'sequence' is required for predict_functional_tracks.")
+            if not isinstance(tracks, list) or not tracks:
+                raise ValueError("Parameter 'tracks' must be a non-empty list for predict_functional_tracks.")
+
+            logger.info(f"AlphaGenomePlugin: Simulating prediction for sequence '{sequence}' with tracks {tracks} and variants {variants}.")
+
+            # Simular resultados (esto sería la salida real de AlphaGenome)
+            simulated_results = {
+                "sequence": sequence,
+                "predicted_tracks": {},
+                "variant_effects": {}
+            }
+
+            for track in tracks:
+                # Simular datos de tracks (ejemplo muy simple)
+                simulated_results["predicted_tracks"][track] = f"Simulated data for {track} on {sequence}"
+
+            for var in variants:
+                # Simular efectos de variantes (ejemplo muy simple)
+                pos = var.get("pos")
+                ref = var.get("ref")
+                alt = var.get("alt")
+                simulated_results["variant_effects"][f"{pos}_{ref}>{alt}"] = {
+                    "gene_expression_lfc": 0.5,
+                    "chromatin_accessibility_score": 0.8,
+                    "splicing_impact": "high"
+                }
+
+            logger.info("AlphaGenomePlugin: Simulated prediction complete.")
+            return simulated_results
+        else:
+            raise NotImplementedError(f"AlphaGenome method '{method_name}' not implemented in simulation.")
