### 2025-07-06 a3ea80a
Add dummy genetic risk predictor model for testing and update GFL parser and lexer.
