+++ gfl/parser.py
@@ -1,187 +1,297 @@
-import ply.yacc as yacc
-from lexer import tokens
-import logging
+# gfl/parser.py
 
-logger = logging.getLogger(__name__)
+import ply.yacc as yacc
+from .lexer import lexer # Importa la instancia del lexer desde lexer.py
+
+# Helper function to get line number and column from token
+def _get_line_col(p, idx=1):
+    if hasattr(p[idx], 'lineno') and hasattr(p[idx], 'lexpos') and hasattr(p.lexer, 'lexdata'):
+        # Calculate column based on lexpos and the start of the line
+        line_start = p.lexer.lexdata.rfind('\n', 0, p[idx].lexpos)
+        if line_start == -1: # If no newline before, it's the first line (column starts from 1)
+            line_start = 0
+        column = (p[idx].lexpos - line_start) + 1 # +1 for 1-based column number
+        return p[idx].lineno, column
+    return 'N/A', 'N/N' # Fallback for tokens without lineno or lexpos
 
 # Precedencia de operadores (de menor a mayor)
-# Esto es crucial para que las expresiones se evalúen correctamente
 precedence = (
     ('left', 'OR'),
     ('left', 'AND'),
     ('right', 'NOT'),
     ('left', 'EQUALS_EQUALS', 'NOT_EQUALS', 'GREATER_THAN', 'LESS_THAN', 'LTE', 'GTE'), # Comparaciones
-    ('left', 'PLUS', 'MINUS', 'TIMES', 'DIVIDE', 'AMPERSAND'), # Aritméticos y concatenación
-    ('right', 'UNARY_MINUS'), # Menos unario (para números negativos)
-    ('left', 'DOT'), # Acceso a miembros
+    ('left', 'PLUS', 'MINUS'),
+    ('left', 'TIMES', 'DIVIDE'),
 )
 
-# Reglas de la gramática
+# --- Parser Definition ---
 
-def p_program(p):
-    '''program : statements'''
-    p[0] = {'type': 'program', 'body': p[1]}
+class GFLParser:
+    """
+    GFL (GeneForge Language) Parser using PLY.
+    """
+    # Define tokens directamente dentro de la clase del parser
+    # PLY los encontrará automáticamente cuando uses yacc.yacc(module=self)
+    tokens = [
+        'DEFINE', 'INVOKE', 'MESSAGE', 'IF', 'THEN', 'ELSE', 'END', 'BRANCH',
+        'TRY', 'CATCH', 'AS', 'BASED_ON', 'NOT',
+        'EQUALS_EQUALS', 'NOT_EQUALS', 'LTE', 'GTE', 'GREATER_THAN', 'LESS_THAN', # Operadores de comparación
+        'PLUS', 'MINUS', 'TIMES', 'DIVIDE', 'EQUALS', 'DOT', 'AMPERSAND',
+        'LPAREN', 'RPAREN', 'COMMA', 'COLON',
+        'LBRACKET', 'RBRACKET', 'LCURLY', 'RCURLY',
+        'AND', 'OR',
+        'TRUE', 'FALSE',
+        'ANALYZE', 'USING', 'WITH', 'STRATEGY', 'PARAMS', # New tokens for analyze statement
+        'EXPERIMENT', 'SIMULATE',
+        'LINE_COMMENT',
+        'BLOCK_COMMENT',
+        'ID', 'STRING', 'NUMBER' # ID, STRING, NUMBER deben ser los últimos
+    ]
 
-def p_statements(p):
-    '''
-    statements : statement
-               | statements statement
-    '''
-    if len(p) == 2:
-        p[0] = [p[1]]
-    else:
-        p[0] = p[1] + [p[2]]
[…]
