+++ scripts/fix_and_demo.py
@@ -1,356 +1,121 @@
-import os
 import sys
-import logging
-import inspect
-import re
+import os
 
-# Configuración básica del logging
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
+# Add the parent directory to the Python path to allow importing 'gfl'
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-# Asegura que el directorio padre del paquete 'gfl' esté en sys.path.
-current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-parent_dir = os.path.dirname(current_dir)
-if parent_dir not in sys.path:
-    sys.path.insert(0, parent_dir)
+from gfl.lexer import lexer # Import the lexer instance directly
+from gfl.parser import GFLParser
 
-from gfl.parser import parse_gfl_code
+def run_parser_demo(parser, code_snippets):
+    """Runs the parser on a list of GFL code snippets."""
+    print("\n--- GFL Parser Demo ---")
+    for i, (name, gfl_code) in enumerate(code_snippets):
+        print(f"\nParsing {name} (Snippet {i+1}):")
+        print("----------------------------------------")
+        print(gfl_code)
+        print("----------------------------------------")
 
-def post_process_gfl_output(gfl_code):
-    """
-    Realiza un post-procesamiento básico para corregir errores comunes de formato
-    en la salida GFL generada por la IA, antes de pasarla al parser.
-    """
-    lines = gfl_code.splitlines()
-    processed_lines = []
-    in_block = 0  # Contador para detectar la profundidad del bloque {}
-    
-    # Patrón para identificar una propiedad: IDENTIFIER: value
-    property_pattern = re.compile(r'^\s*([a-zA-Z_][a-zA-Z_0-9]*)\s*:\s*(.*)')
+        # Parse the code directly without preprocessing for now
+        operations = parser.parse(gfl_code)
 
-    for i, line in enumerate(lines):
-        stripped_line = line.strip()
-        original_line_indent = len(line) - len(line.lstrip())
-
-        # Eliminar comas extra al inicio de una línea o después de ':'
-        stripped_line = re.sub(r'^\s*,\s*', '', stripped_line) # Eliminar coma al inicio
-        stripped_line = re.sub(r':\s*,', ':', stripped_line) # Eliminar coma después de ':'
-
-        # Normalizar True/False sin comillas si la IA las añade
-        stripped_line = stripped_line.replace('"true"', 'true').replace('"false"', 'false')
-        
-        # Ajustar espacios después de ':' y ','
-        stripped_line = re.sub(r':\s*', ':', stripped_line)
-        stripped_line = re.sub(r',\s*', ',', stripped_line)
-
-        # Insertar comas entre propiedades si faltan, solo si estamos en un bloque
-        if in_block > 0 and processed_lines:
-            last_processed_line = processed_lines[-1].strip()
-            current_is_property = property_pattern.match(stripped_line)
-            
-            # Heurística: Si la línea anterior es una propiedad completa (no solo el inicio de un bloque)
-            # y no termina en '{' o '}', y la línea actual es otra propiedad,
-            # y no hay una coma explícita ya en la línea anterior.
-            if not last_processed_line.endswith('{') and \
-               not last_processed_line.endswith('}') and \
-               not last_processed_line.endswith(',') and \
-               current_is_property and \
-               property_pattern.match(last_processed_line): # Asegurarse que la línea anterior también era una propiedad
-                
-                # Añadir la coma solo si la línea anterior realmente contiene un valor (no solo "key:")
[…]
