+++ gfl/parser.py
@@ -1,90 +1,35 @@
-# gfl/parser.py
-
 import ply.yacc as yacc
-from .lexer import lexer # Importa la instancia del lexer desde lexer.py
-
-# Helper function to get line number and column from token
-def _get_line_col(p, idx=1):
-    if hasattr(p[idx], 'lineno') and hasattr(p[idx], 'lexpos') and hasattr(p.lexer, 'lexdata'):
-        # Calculate column based on lexpos and the start of the line
-        line_start = p.lexer.lexdata.rfind('\n', 0, p[idx].lexpos)
-        if line_start == -1: # If no newline before, it's the first line (column starts from 1)
-            line_start = 0
-        column = (p[idx].lexpos - line_start) + 1 # +1 for 1-based column number
-        return p[idx].lineno, column
-    return 'N/A', 'N/N' # Fallback for tokens without lineno or lexpos
-
-# Precedencia de operadores (de menor a mayor)
-precedence = (
-    ('left', 'OR'),
-    ('left', 'AND'),
-    ('right', 'NOT'),
-    ('left', 'EQUALS_EQUALS', 'NOT_EQUALS', 'GREATER_THAN', 'LESS_THAN', 'LTE', 'GTE'), # Comparaciones
-    ('left', 'PLUS', 'MINUS'),
-    ('left', 'TIMES', 'DIVIDE'),
-)
-
-# --- Parser Definition ---
+import ply.lex as lex
+from gfl.lexer import GFLLexer
 
 class GFLParser:
     """
-    GFL (GeneForge Language) Parser using PLY.
+    Clase principal para el parser de GeneForgeLang (GFL).
+    Utiliza PLY (Python Lex-Yacc) para construir el analizador sintáctico.
     """
-    # Define tokens directamente dentro de la clase del parser
-    # PLY los encontrará automáticamente cuando uses yacc.yacc(module=self)
-    tokens = [
-        'DEFINE', 'INVOKE', 'MESSAGE', 'IF', 'THEN', 'ELSE', 'END', 'BRANCH',
-        'TRY', 'CATCH', 'AS', 'BASED_ON', 'NOT',
-        'EQUALS_EQUALS', 'NOT_EQUALS', 'LTE', 'GTE', 'GREATER_THAN', 'LESS_THAN', # Operadores de comparación
-        'PLUS', 'MINUS', 'TIMES', 'DIVIDE', 'EQUALS', 'DOT', 'AMPERSAND',
-        'LPAREN', 'RPAREN', 'COMMA', 'COLON',
-        'LBRACKET', 'RBRACKET', 'LCURLY', 'RCURLY',
-        'AND', 'OR',
-        'TRUE', 'FALSE',
-        'ANALYZE', 'USING', 'WITH', 'STRATEGY', 'PARAMS', # New tokens for analyze statement
-        'EXPERIMENT', 'SIMULATE',
-        'LINE_COMMENT',
-        'BLOCK_COMMENT',
-        'ID', 'STRING', 'NUMBER' # ID, STRING, NUMBER deben ser los últimos
-    ]
-
-    def __init__(self, lexer_instance):
-        self.lexer = lexer_instance
-        # PLY ahora encontrará GFLParser.tokens automáticamente
-        self.parser = yacc.yacc(module=self, start='program', debug=True)
-        self.operations = [] # Stores the parsed operations/AST
-        self.errors = []     # Stores parsing errors
-
-        # Basic validation data (can be expanded)
-        self.valid_sim_targets = ["cell_growth", "apoptosis", "cell_division", "mutation_rate", "gene_expression"]
-        self.valid_analyze_strategies = {
-            "pathway_enrichment": ["FDR", "threshold"],
-            "clustering": ["resolution"],
[…]
