+++ scripts/fix_and_demo.py
@@ -1,121 +1,183 @@
-import sys
 import os
 
-# Add the parent directory to the Python path to allow importing 'gfl'
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+# Asegúrate de que el PYTHONPATH esté configurado correctamente para importar desde gfl
+# Esto es manejado por el comando de ejecución 'PYTHONPATH=. python3 scripts/fix_and_demo.py'
+# así que no necesitamos modificar sys.path aquí.
 
-from gfl.lexer import lexer # Import the lexer instance directly
+# Importa las clases GFLLexer y GFLParser
+from gfl.lexer import GFLLexer
 from gfl.parser import GFLParser
 
-def run_parser_demo(parser, code_snippets):
-    """Runs the parser on a list of GFL code snippets."""
-    print("\n--- GFL Parser Demo ---")
-    for i, (name, gfl_code) in enumerate(code_snippets):
-        print(f"\nParsing {name} (Snippet {i+1}):")
-        print("----------------------------------------")
-        print(gfl_code)
-        print("----------------------------------------")
+def run_demo():
+    print("--- GFL Parser Demo ---")
 
-        # Parse the code directly without preprocessing for now
-        operations = parser.parse(gfl_code)
+    # Instancia el lexer y el parser
+    gfl_lexer = GFLLexer()
+    gfl_parser = GFLParser()
 
-        if parser.errors:
-            print("\nParsing Errors/Warnings:")
-            for error in parser.errors:
-                print(f"  - {error}")
-            parser.errors = [] # Clear errors for next snippet
-        else:
-            print("\nSuccessfully Parsed! Generated AST (Operations):")
-            if operations:
-                for op in operations:
-                    print(f"  {op}")
+    # Define los snippets de código GFL para parsear
+    snippets = {
+        "gfl/gfl_example.gfl": """
+analyze {
+    strategy: "pathway_enrichment",
+    thresholds: {
+        FDR: 0.05
+    }
+}
+
+experiment {
+    tool: "DESeq2",
+    type: "bulkRNA",
+    params: {
+        condition_group: "disease",
+        control_group: "healthy"
+    }
+}
+
+simulate cell_growth
+
+branch {
+    if: tumor_size_increased AND cell_death_rate_high
+    then: {
+        simulate apoptosis
+        analyze {
+            strategy: "clustering",
+            thresholds: {
+                resolution: 0.8
+            }
+        }
+    }
+    else: {
+        simulate cell_division
+    }
+}
+
+# Este bloque debería generar un ERROR de validación de parámetros
+analyze using DESeq2 with strategy differential_expression params {
+    threshold: 0.05,
+    log2FC: 1.0,
+    unknown_param: "esto_es_invalido"
+}
+
+# Este bloque debería generar un ERROR de validación de parámetros
+analyze using Scanpy with strategy clustering params {
+    resolution: 0.8,
+    invalid_cluster_method: "leiden"
+}
+
+# Este bloque debería generar un ERROR de validación de parámetros
+analyze using GSEA with strategy pathway_enrichment params {
+    gene_set_library: "GO_BP",
+    extra_option: true
+}
+""",
+        "AI-Generated Scenario 1 (Simple Analyze)": """
+analyze {
+    strategy: "pathway_enrichment",
+    thresholds: {
+        FDR: 0.05
+    }
+}
+""",
+        "AI-Generated Scenario 2 (Complex Branch)": """
+branch {
+    if: tumor_size_increased AND cell_death_rate_high
+    then: {
+        simulate apoptosis
+        analyze {
+            strategy: "clustering",
+            thresholds: {
+                resolution: 0.8
+            }
+        }
+    }
+    else: {
+        simulate cell_division
+    }
+}
+""",
[…]

[…]
