### 2025-07-06 312fb9f
Update README: Add Genetic Risk Predictor CLI tool documentation
