### 2025-07-06 44704be
Revert "Update README: Add Genetic Risk Predictor CLI tool documentation"
This reverts commit 312fb9f40fe1abc454462938241d7c1301e5d473.
