### 2025-07-26 #8281 @Kevin-Mattheus-Moerman
@manuelmenendezgonzalez

- [ ] Can you set your ORCID, if yours is 0000-0002-5218-0774 then set that one.
- [ ] Can you check your references are now valid? You can use `@editorialbot check references` to check them again.
