### 2025-07-27 62d27be
fix(parser): Replace legacy parser with YAML implementation
