### 2025-07-28 80a9601
Update gfl_core submodule
