### 2025-08-21 33571af
feat: Sincronización de cambios locales en el submódulo de GFL Core.
