### 2025-08-24 #8281 @abhishektiwari
@manuelmenendezgonzalez Following up my last comment.

> [@manuelmenendezgonzalez](https://github.com/manuelmenendezgonzalez)
>
> * Citation for "Eugene – A Domain Specific Language for Specifying and Constraining Synthetic Biological Parts, Devices, and Systems". I am not sure if entry with key `Eugene2009` in your `paper.bib` file is correct. Can you please review above missing DOI error and check if [this article](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0018882) is correct citation? If so please update your `paper.bib`.
> * I was unable to find the `CONTRIBUTING.md` mentioned in [contributing section](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/tree/main?tab=readme-ov-file#contributing) of README. If says _"Please see CONTRIBUTING.md (if available) for guidelines on how to submit pull requests, report issues, and contribute to the project."_ Please add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).
>
> Once above two are done, can you suggest a few potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers.
