### 2025-08-28 c3185c7
Update paper.bib
