+++ paper.bib
@@ -29,9 +29,13 @@
   url = {https://jbioleng.biomedcentral.com/articles/10.1186/1754-1611-3-1}
 }
 
-@inproceedings{Eugene2009,
-  author    = {David Densmore and Vikraman Gopalakrishnan and Janet A. N. Anderson and Drew Endy},
+@inproceedings{Eugene2011,
+  author    = {Lesia Bilitchenko,Adam Liu,Sherine Cheung,Emma Weeding,Bing Xia,Mariana Leguia,J. Christopher Anderson,Douglas Densmore},
   title     = {Eugene – A Domain Specific Language for Specifying and Constraining Synthetic Biological Parts, Devices, and Systems},
-  booktitle = {Bioinformatics Workshop},
-  year      = {2009}
+  booktitle = {PLOS ONE},
+  year    = {2011},
+  volume  = {6},
+  number  = {4},
+  pages   = {e18882},
+  doi     = {10.1371/journal.pone.0018882}
 }
