### 2025-08-29 dc0ea06
Add translator app with validation + inference; public API; dummy model; tests; pre-commit and CI; integrations layout.
