### 2025-08-29 1e18077
feat(app): add Inference Summary panel; tests: round-trip validation for examples/
