### 2025-08-29 72b4a60
chore(docs): add CI/PyPI badges to README feat(app): add validation status banner and model selector (Dummy/SimpleHeuristic)
