### 2025-08-29 2aa173f
feat(docs,app): add compact badges and Try the app section; persist model selection via state; add downloadable outputs for GFL and inference JSON
