+++ docs/images/README.md
@@ -1,0 +1,7 @@
+Place UI screenshots here for the README references:
+
+- translator_home.png
+- translator_inference.png
+
+These files are not included in the repository by default.
+
