### 2025-08-29 1509c97
fix(tests): make validator tolerant to missing plugins; add prime_edit param check; fix plugin registry imports; remove stray root __init__ causing pytest import
