### 2025-08-29 63e9665
ci: add environment diagnostics and run pytest -vv for clearer failures
