### 2025-08-29 733a929
ci: capture pytest output and upload artifacts; explicit failure on non-zero
