### 2025-08-29 7e345af
docs: clarify screenshots note (retrigger CI)
