+++ docs/images/README.md
@@ -3,5 +3,4 @@ Place UI screenshots here for the README references:
 - translator_home.png
 - translator_inference.png
 
-These files are not included in the repository by default.
-
+These files are not included in the repository by default. Add your screenshots to improve the README visuals.
