### 2025-08-29 79a2f5d
docs: remove screenshot links from README
