### 2025-08-29 edff3b6
tests: relax smoke test to accept legacy DSL strings in examples/
