### 2025-08-30 b593bd4
﻿feat: add YAML parser with PyYAML, package metadata v0.2.0, migrate GFL YAML spec/examples, conditional public API, and NL translator module
