+++ README.md
@@ -1,279 +1,170 @@
-# GeneForgeLang
+# GeneForge: Applications and Pipelines for GeneForgeLang
 
-GeneForgeLang (GFL) is a domain-specific language designed for specifying and executing complex genomic workflows. It provides a high-level, human-readable syntax for defining genetic experiments, data processing pipelines, and analytical tasks.
+## Overview
+
+This repository hosts the **applications, pipelines, examples, and utility scripts** that extend and leverage the **GeneForgeLang (GFL) DSL** (Domain-Specific Language). It serves as the practical implementation layer, demonstrating how GFL can be used to build and execute complex biological workflows.
+
+## Relationship with GeneForgeLang
+
+GeneForge relies on the **core GeneForgeLang engine** (available in the [GeneForgeLang repository](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang.git)) for parsing, evaluating, and reasoning about biological workflows. This repository focuses on the **applications and pipelines** that utilize this engine.
+
+## Recent Updates
+
+As of July 2025, the **GeneForgeLang (GFL) engine** has received significant improvements in the clarity of its error and warning messages, making workflow development more intuitive. This project (GeneForge applications) now includes **enhanced GFL parser and inference engine logic**, resulting in more robust handling of GFL workflows. For details on the DSL's roadmap, please refer to the [GeneForgeLang repository](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang.git).
 
 ## Project Structure
 
-```
-GeneForgeLang/
-├── __init__.py
-├── .gitattributes
-├── .gitignore
-├── advanced_validator.py
-├── alphagenenome_plugin.py
-├── app_gradio_full_graph.py
-├── app.py
-├── bfg.jar
-├── CITATION.cff
-├── Ejecutando
-├── example1.gfl
-├── export_ast.py
-├── geneforge_grammar_v1.2.json
-├── geneforge_grammar_v1.3.json
-├── geneforge_syntax_v1.2.md
-├── geneforgegrammar.json
-├── generar_desde_frase_input_v2.py
-├── generar_desde_frase_json.py
-├── generar_desde_frase_v2.py
-├── generar_interactivo.py
-├── gfl_benchmark_tasks.md
-├── gfl_example.gfl
-├── gfl_examples.gfl
-├── gfl_to_vcf.py
-├── grammar.md
-├── license
-├── main.py
-├── old_parser_root.py
-├── ontology.md
-├── output_ast.json
-├── paper.bib
-├── paper.md
-├── parselog.txt
-├── parser_notebook.ipynb
-├── pytest test_parser.py
-├── README.md
-├── requirements.txt
-├── rules.json
-├── run_axiom_demo.py
-├── sanitize_identifiers.py
-├── semillas.json
-├── summarize_ast.py
-├── syntax_v1.2.md
-├── syntax.md
-├── test_gfl_logic.gfl
-├── test_semantics.py
-├── translate_to_geneforgelang.py
-├── variant_simulation_plugin.py
-├── visualize_ast.py
-├── whitepaper.md
-├── ..bfg-report/
-│   └── 2025-06-08/
-│       ├── 23-48-11/
-│       │   ├── cache-stats.txt
-│       │   ├── deleted-files.txt
-│       │   └── object-id-map.old-new.txt
-│       ├── 23-48-12/
-│       │   ├── cache-stats.txt
-│       │   ├── deleted-files.txt
-│       │   └── object-id-map.old-new.txt
-│       ├── 23-48-13/
-│       │   ├── cache-stats.txt
-│       │   ├── deleted-files.txt
-│       │   └── object-id-map.old-new.txt
[…]
