### 2025-08-30 064ea5e
chore: governance templates + CI; add YAML parser and packaging v0.2.0
