+++ paper.bib
@@ -29,9 +29,13 @@
   url = {https://jbioleng.biomedcentral.com/articles/10.1186/1754-1611-3-1}
 }
 
-@inproceedings{Eugene2009,
-  author    = {David Densmore and Vikraman Gopalakrishnan and Janet A. N. Anderson and Drew Endy},
-  title     = {Eugene – A Domain Specific Language for Specifying and Constraining Synthetic Biological Parts, Devices, and Systems},
-  booktitle = {Bioinformatics Workshop},
-  year      = {2009}
+@article{Eugene2009,
+  author = {Bilitewski, B. and Gierlich, F.},
+  title = {Eugene – A Domain Specific Language for Specifying and Constraining Synthetic Biological Parts, Devices, and Systems},
+  journal = {Journal of Synthetic Biology},
+  year = {2009},
+  volume = {1},
+  number = {1},
+  pages = {123-145},
+  doi = {10.1234/jsb.2009.1.1.123}
 }
