### 2025-08-30 3f813c0
fix: valid pyproject.toml; bump 0.2.3
