+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "geneforgelang"
-version = "0.2.0"
+version = "0.2.3"
 description = "GeneForgeLang (GFL): core language with YAML DSL and optional NL translator"
 readme = "README.md"
 requires-python = ">=3.9"
@@ -12,23 +12,15 @@ authors = [{ name = "Fundación de Neurociencias", email = "admin@fneurociencias
 license = { file = "license" }
 keywords = ["bioinformatics", "language", "GFL", "workflow"]
 classifiers = [
-  "Programming Language :: Python :: 3",
-  "License :: OSI Approved :: MIT License",
-  "Operating System :: OS Independent"
+"Programming Language :: Python :: 3",
+"License :: OSI Approved :: MIT License",
+"Operating System :: OS Independent"
 ]
-dependencies = [
-  "pyyaml>=6.0"
-]
+dependencies = ["pyyaml>=6.0"]
 
 [project.optional-dependencies]
-nl = [
-  "transformers>=4.35",
-  "datasets>=2.14"
-]
-ply = [
-  "ply>=3.11"
-]
-
+nl = ["transformers>=4.35", "datasets>=2.14"]
+ply = ["ply>=3.11"]
 dev = ["pytest"]
 
 [tool.setuptools.packages.find]
