### 2025-08-30 c768e94
lang: packaging fixes; add nl extras (transformers,datasets,torch); prob_rules: init rule_scores
