### 2025-08-30 f42798e
chore: normalize line endings for CONTRIBUTING.MD
