### 2025-08-30 1fbb7ce
feat: add GFL JSON schema, CLI entry points, and .gitattributes
