### 2025-08-30 c36ba93
build: add console scripts (gfl-parse, gfl-validate)
