### 2025-08-30 b7402c4
build: fix [project.scripts] entry points
