### 2025-08-30 d5bcb34
fix(cli): correct indentation and bodies for CLI commands
