### 2025-08-30 541dda4
chore(ci): add ruff+mypy+pytest and pre-commit to CI matrix
