+++ .github/workflows/ci.yml
@@ -1,52 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches: [ main, master ]
-  pull_request:
-    branches: [ main, master ]
-
-jobs:
-  tests:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.11'
-      - name: Install package (core)
-        run: |
-          python -m pip install --upgrade pip
-          pip install -e .
-          pip install pytest
-      - name: Debug environment
-        run: |
-          echo "Python:"; python -V
-          echo "Which python:"; which python || where python || true
-          echo "Pip freeze:"; pip freeze
-          echo "Repo tree (top-level):"; ls -la
-          echo "Python path:"; python - << 'PY'
-          import sys, pprint
-          pprint.pprint(sys.path)
-          PY
-      - name: List repository files (depth 2)
-        run: |
-          find . -maxdepth 2 -type f -print | sort | sed -n '1,400p'
-      - name: Run tests (verbose, capture)
-        id: run_pytest
-        shell: bash
-        run: |
-          set -o pipefail
-          pytest -vv --maxfail=1 --junitxml=pytest.xml | tee pytest.out
-        continue-on-error: true
-      - name: Upload test logs
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: pytest-logs
-          path: |
-            pytest.out
-            pytest.xml
-      - name: Fail if tests failed
-        if: steps.run_pytest.outcome == 'failure'
-        run: exit 1
