### 2025-08-30 36d983a
chore(ci): add ruff+mypy+pytest and pre-commit to CI matrix
