+++ .github/workflows/ci.yml
@@ -1,0 +1,37 @@
+﻿name: CI
+
+on:
+  push:
+    branches: [ "main", "master" ]
+  pull_request:
+    branches: [ "main", "master" ]
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest]
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Upgrade pip
+        run: python -m pip install -U pip
+      - name: Install package (editable) + tools
+        run: |
+          python -m pip install -e .
+          python -m pip install pytest mypy ruff pre-commit
+      - name: Ruff lint
+        run: ruff check .
+      - name: Ruff format (check only)
+        run: ruff format --check .
+      - name: Mypy type check
+        run: mypy .
+      - name: Run tests (if any)
+        if: ${{ hashFiles('tests/**') != '' }}
+        run: pytest -q
+      - name: Run pre-commit (all files)
+        uses: pre-commit/action@v3.0.1
