### 2025-08-30 253fee2
chore: apply repository-wide formatting and normalization (pre-commit)
