+++ gfl/cli.py
@@ -1,30 +1,264 @@
-﻿import sys
+﻿"""Command-line interface for GeneForgeLang.
+
+Provides console commands for parsing, validation, inference, and utilities.
+"""
+
 import argparse
-from gfl.api import parse, validate
+import json
+import sys
+
+from gfl.api import get_api_info, infer, parse, validate
+
+try:
+    from gfl.schema_validator import comprehensive_validate, load_schema
+
+    HAS_SCHEMA = True
+except ImportError:
+    HAS_SCHEMA = False
+    comprehensive_validate = None
+    load_schema = None
+
+try:
+    from gfl.models.dummy import DummyModel
+
+    HAS_DUMMY_MODEL = True
+except ImportError:
+    HAS_DUMMY_MODEL = False
+    DummyModel = None
 
 
 def cmd_parse():
-    ap = argparse.ArgumentParser(
-        description="Parse GFL (YAML) to AST and print it."
-    )
-    ap.add_argument("file", help="Path to GFL YAML file")
-    args = ap.parse_args()
-    text = open(args.file, encoding="utf-8").read()
-    ast = parse(text)
-    print(ast)
+    """Parse GFL file and output AST."""
+    parser = argparse.ArgumentParser(description="Parse GFL file to AST")
+    parser.add_argument("file", help="Path to GFL file")
+    parser.add_argument("--typed", action="store_true", help="Use typed AST")
+    parser.add_argument("--format", choices=["json", "yaml", "repr"], default="repr")
+    parser.add_argument("--output", "-o", help="Output file")
+
+    args = parser.parse_args()
+
+    try:
+        with open(args.file, encoding="utf-8") as f:
+            text = f.read()
+
+        ast = parse(text, typed=args.typed)
+
+        if args.format == "json":
+            if args.typed:
+                output = json.dumps(ast.to_dict(), indent=2)
+            else:
+                output = json.dumps(ast, indent=2)
+        elif args.format == "yaml":
+            import yaml
+
+            if args.typed:
+                output = yaml.dump(ast.to_dict(), default_flow_style=False)
+            else:
+                output = yaml.dump(ast, default_flow_style=False)
+        else:  # repr
+            output = repr(ast)
+
+        if args.output:
+            with open(args.output, "w", encoding="utf-8") as f:
+                f.write(output)
+            print(f"AST written to {args.output}")
+        else:
+            print(output)
+
+    except Exception as e:
+        print(f"Error: {e}", file=sys.stderr)
+        sys.exit(1)
 
 
 def cmd_validate():
-    ap = argparse.ArgumentParser(
-        description="Validate GFL (YAML) semantics."
-    )
-    ap.add_argument("file", help="Path to GFL YAML file")
-    args = ap.parse_args()
-    text = open(args.file, encoding="utf-8").read()
-    ast = parse(text)
-    errs = validate(ast)
-    if errs:
-        for e in errs:
-            print(f"- {e}")
+    """Validate GFL file semantics and schema."""
+    parser = argparse.ArgumentParser(description="Validate GFL file")
+    parser.add_argument("file", help="Path to GFL file")
+    parser.add_argument("--detailed", action="store_true", help="Detailed validation")
+    parser.add_argument("--schema", action="store_true", help="Schema validation")
+    parser.add_argument("--format", choices=["text", "json"], default="text")
+    parser.add_argument("--output", "-o", help="Output file")
+
+    args = parser.parse_args()
+
+    try:
+        with open(args.file, encoding="utf-8") as f:
+            text = f.read()
+
+        ast = parse(text)
+
+        if args.schema and HAS_SCHEMA:
+            result = comprehensive_validate(ast)
+        else:
+            result = validate(ast, detailed=args.detailed)
+
[…]
