+++ gfl/parser.py
@@ -1,7 +1,9 @@
-import yaml
 import logging
 
+import yaml
+
 logger = logging.getLogger(__name__)
+
 
 def parse_gfl(gfl_string):
     """
