+++ run_axiom_demo.py
@@ -1,7 +1,8 @@
+import os
+
+from gf.axioms.axiom_utils import load_axiom_store
 from gfl import parser
 from gfl.axiom_hooks import register_axiom_candidates
-from gf.axioms.axiom_utils import load_axiom_store
-import os
 
 example_path = os.path.join("gfl", "examples", "example1.gfl")
 with open(example_path, "r", encoding="utf-8") as f:
