### 2025-08-30 b13253b
feat: comprehensive GeneForgeLang enhancement to v0.2.0
 Major Features Added:
- Stable & Typed API with dataclasses and TypedDict
- Comprehensive JSON Schema validation
- PyPI-ready packaging with entry points
- Advanced CLI tools (gfl-parse, gfl-validate, gfl-infer, gfl-schema, gfl-info)
- Plugin discovery system with entry points
- Multi-platform CI matrix (Linux/Windows/macOS, Python 3.10-3.13)
- Pre-commit hooks with quality tools (ruff, black, mypy, bandit)
- MCP integration for cross-platform AI context sharing
- Complete documentation with API reference and migration guide

 Technical 
[…]
