+++ gfl/types.py
@@ -1,0 +1,279 @@
+"""Type definitions for GeneForgeLang AST structures.
+
+This module provides stable, typed representations of GFL AST nodes using
+dataclasses and TypedDict. These types ensure API stability and enable
+better IDE support, static analysis, and documentation.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import Any, Dict, List, Literal, Optional, Union
+
+# Type aliases for clarity
+GFLValue = Union[str, int, float, bool, None]
+GFLDict = Dict[str, Any]
+
+
+class ExperimentType(str, Enum):
+    """Valid experiment types in GFL."""
+
+    GENE_EDITING = "gene_editing"
+    SEQUENCING = "sequencing"
+    ANALYSIS = "analysis"
+    SIMULATION = "simulation"
+    VALIDATION = "validation"
+
+
+class AnalysisStrategy(str, Enum):
+    """Valid analysis strategies."""
+
+    DIFFERENTIAL = "differential"
+    PATHWAY = "pathway"
+    VARIANT = "variant"
+    EXPRESSION = "expression"
+    STRUCTURAL = "structural"
+
+
+# Validation result types
+@dataclass
+class ValidationError:
+    """Represents a validation error."""
+
+    message: str
+    location: Optional[str] = None
+    severity: Literal["error", "warning", "info"] = "error"
+    code: Optional[str] = None
+
+    def __str__(self) -> str:
+        """String representation of the error."""
+        parts = []
+        if self.severity != "error":
+            parts.append(f"[{self.severity.upper()}]")
+        if self.location:
+            parts.append(f"{self.location}:")
+        parts.append(self.message)
+        if self.code:
+            parts.append(f"({self.code})")
+        return " ".join(parts)
+
+
+@dataclass
+class ValidationResult:
+    """Result of AST validation."""
+
+    errors: List[ValidationError] = field(default_factory=list)
+    warnings: List[ValidationError] = field(default_factory=list)
+    info: List[ValidationError] = field(default_factory=list)
+
+    @property
+    def is_valid(self) -> bool:
+        """True if no errors (warnings are allowed)."""
+        return len(self.errors) == 0
+
+    @property
+    def all_messages(self) -> List[ValidationError]:
+        """All validation messages combined."""
+        return self.errors + self.warnings + self.info
+
+    def to_string_list(self) -> List[str]:
+        """Convert to list of strings for backward compatibility."""
+        return [str(msg) for msg in self.all_messages]
+
+
+# Inference result types
+@dataclass
+class InferenceResult:
+    """Result of probabilistic inference."""
+
+    predictions: Dict[str, Any]
+    confidence: float
+    metadata: Dict[str, Any] = field(default_factory=dict)
+
+    def to_dict(self) -> Dict[str, Any]:
+        """Convert to dictionary representation."""
+        return {
+            "predictions": self.predictions,
+            "confidence": self.confidence,
+            "metadata": self.metadata,
+        }
+
+
+# Simple dataclass versions for core structures
+@dataclass
+class ExperimentParams:
+    """Parameters for experiment configuration."""
+
+    target_gene: Optional[str] = None
+    sequence: Optional[str] = None
+    guide_rna: Optional[str] = None
+    vector: Optional[str] = None
+    concentration: Optional[float] = None
+    temperature: Optional[float] = None
+    duration: Optional[str] = None
+    replicates: Optional[int] = None
+
+    # Allow arbitrary additional parameters
+    extra: Dict[str, GFLValue] = field(default_factory=dict)
+
+    def to_dict(self) -> Dict[str, Any]:
+        """Convert to dictionary representation."""
+        result = {}
[…]
