### 2025-08-30 a5abc17
Sync local changes: Add security summary, update project configuration, and normalize line endings
