+++ sanitize_identifiers.py
@@ -1,4 +1,4 @@
-﻿import re
+import re
 import sys
 
 with open(sys.argv[1], encoding="utf-8") as f:
