+++ test_semantics.py
@@ -1,4 +1,4 @@
-﻿import json
+import json
 import subprocess
 import sys
 
