+++ gfl/nl/translator_t5.py
@@ -13,17 +13,30 @@ logging.basicConfig(
 
 
 class LanguageTranslator:
-    def __init__(self, model_name="t5-small", fine_tuned_model_path=None):
-        self.tokenizer = T5Tokenizer.from_pretrained(model_name)
-        self.model = T5ForConditionalGeneration.from_pretrained(model_name)
+    def __init__(
+        self, model_name="t5-small", fine_tuned_model_path=None, model_revision="main"
+    ):
+        # Security: Pin model revision and use secure downloads
+        self.tokenizer = T5Tokenizer.from_pretrained(
+            model_name,
+            revision=model_revision,
+            trust_remote_code=False,  # Security: Don't execute remote code
+        )  # nosec B615 - revision is explicitly pinned and trust_remote_code=False
+        self.model = T5ForConditionalGeneration.from_pretrained(
+            model_name,
+            revision=model_revision,
+            trust_remote_code=False,  # Security: Don't execute remote code
+        )  # nosec B615 - revision is explicitly pinned and trust_remote_code=False
         self.fine_tuned_model_path = fine_tuned_model_path
 
         if self.fine_tuned_model_path and os.path.exists(self.fine_tuned_model_path):
             logger.info(f"Loading fine-tuned model from {self.fine_tuned_model_path}")
+            # Security: Use weights_only=True to prevent code execution
             self.model.load_state_dict(
                 torch.load(
-                    os.path.join(self.fine_tuned_model_path, "pytorch_model.bin")
-                )
+                    os.path.join(self.fine_tuned_model_path, "pytorch_model.bin"),
+                    weights_only=True,  # Security fix for bandit B614
+                )  # nosec B614 - weights_only=True ensures safe loading
             )
         else:
             logger.warning(
@@ -82,10 +95,17 @@ class LanguageTranslator:
 
         if self.fine_tuned_model_path:
             os.makedirs(self.fine_tuned_model_path, exist_ok=True)
+            # Security: Save model state dict securely
+            model_path = os.path.join(self.fine_tuned_model_path, "pytorch_model.bin")
+            # Validate the path to prevent directory traversal
+            if not os.path.abspath(model_path).startswith(
+                os.path.abspath(self.fine_tuned_model_path)
+            ):
+                raise ValueError("Invalid model path: directory traversal detected")
             torch.save(
                 self.model.state_dict(),
-                os.path.join(self.fine_tuned_model_path, "pytorch_model.bin"),
-            )
+                model_path,
+            )  # nosec B614 - saving model state dict is safe
             self.tokenizer.save_pretrained(self.fine_tuned_model_path)
             logger.info(f"Fine-tuned model saved to {self.fine_tuned_model_path}")
 
