+++ gfl/yaml_lang/parser.py
@@ -1,4 +1,4 @@
-﻿from typing import Any, Dict
+from typing import Any, Dict
 
 import yaml
 
