+++ gfl/translate_lstm.py
@@ -1,4 +1,4 @@
-﻿import torch
+import torch
 import torch.nn as nn
 
 
