### 2025-08-30 c6cdd95
feat: Complete Phase 3 implementation with web interface and API server
- Implement comprehensive FastAPI REST API server with 12+ endpoints
- Add interactive Gradio web interface with 5 tabs (editor, inference, comparison, management, batch)
- Create unified server launcher for process management
- Build type-safe client SDK with sync/async support
- Add comprehensive testing suite with mock-based validation
- Integrate enhanced inference engine with ML model support
- Include advanced models (protein generation, genomic classification, multimodal)
- Implement security best practices (rate li
[…]
