+++ tests/unit/test_types.py
@@ -1,0 +1,520 @@
+"""Unit tests for the GFL types module.
+
+Tests cover:
+- Dataclass creation and validation
+- Type conversion and serialization
+- AST construction from dictionaries
+- Backward compatibility with dict-based API
+"""
+
+import pytest
+from typing import Dict, Any
+
+from gfl.types import (
+    GFLAST,
+    Analysis,
+    AnalysisStrategy,
+    Experiment,
+    ExperimentParams,
+    ExperimentType,
+    InferenceResult,
+    ValidationError,
+    ValidationResult,
+)
+
+
+class TestExperimentParams:
+    """Test ExperimentParams dataclass."""
+
+    def test_create_empty_params(self):
+        """Test creating empty experiment parameters."""
+        params = ExperimentParams()
+        assert params.target_gene is None
+        assert params.concentration is None
+        assert params.extra == {}
+
+    def test_create_params_with_values(self):
+        """Test creating parameters with specific values."""
+        params = ExperimentParams(
+            target_gene="TP53",
+            concentration=50.0,
+            temperature=37.0,
+            replicates=3,
+        )
+
+        assert params.target_gene == "TP53"
+        assert params.concentration == 50.0
+        assert params.temperature == 37.0
+        assert params.replicates == 3
+
+    def test_params_extra_dict(self):
+        """Test extra parameters dictionary."""
+        params = ExperimentParams(
+            target_gene="TP53", extra={"custom_param": "custom_value", "buffer": "PBS"}
+        )
+
+        assert params.extra["custom_param"] == "custom_value"
+        assert params.extra["buffer"] == "PBS"
+
+    def test_params_to_dict(self):
+        """Test conversion of parameters to dictionary."""
+        params = ExperimentParams(
+            target_gene="TP53", concentration=50.0, extra={"custom_param": "value"}
+        )
+
+        result = params.to_dict()
+
+        assert result["target_gene"] == "TP53"
+        assert result["concentration"] == 50.0
+        assert result["custom_param"] == "value"
+        assert "extra" not in result  # Should be flattened
+
+    def test_params_to_dict_excludes_none(self):
+        """Test that to_dict excludes None values."""
+        params = ExperimentParams(target_gene="TP53")
+        result = params.to_dict()
+
+        assert "target_gene" in result
+        assert "concentration" not in result
+        assert "temperature" not in result
+
+
+class TestExperiment:
+    """Test Experiment dataclass."""
+
+    def test_create_minimal_experiment(self):
+        """Test creating minimal experiment."""
+        exp = Experiment(tool="CRISPR_cas9", type="gene_editing")
+
+        assert exp.tool == "CRISPR_cas9"
+        assert exp.type == "gene_editing"
+        assert isinstance(exp.params, ExperimentParams)
+        assert exp.strategy is None
+
+    def test_create_experiment_with_params(self):
+        """Test creating experiment with parameters."""
+        params = ExperimentParams(target_gene="TP53", concentration=50.0)
+        exp = Experiment(
+            tool="CRISPR_cas9", type="gene_editing", params=params, strategy="knockout"
+        )
+
+        assert exp.params.target_gene == "TP53"
+        assert exp.strategy == "knockout"
+
+    def test_experiment_to_dict(self):
+        """Test experiment conversion to dictionary."""
+        params = ExperimentParams(target_gene="TP53")
+        exp = Experiment(
+            tool="CRISPR_cas9", type="gene_editing", params=params, strategy="knockout"
+        )
+
+        result = exp.to_dict()
+
+        assert result["tool"] == "CRISPR_cas9"
+        assert result["type"] == "gene_editing"
+        assert result["strategy"] == "knockout"
+        assert isinstance(result["params"], dict)
[…]
