+++ tests/unit/test_enhanced_plugins.py
@@ -1,0 +1,350 @@
+"""Tests for enhanced plugin system with dependencies and lifecycle hooks."""
+
+import pytest
+from typing import Dict, Any, List
+
+from gfl.plugins.plugin_registry import (
+    plugin_registry,
+    BaseGFLPlugin,
+    PluginInfo,
+    PluginState,
+    PluginPriority,
+    PluginDependency,
+    get_plugin,
+    activate_plugin,
+    process_with_plugins,
+    validate_plugin_dependencies,
+    add_lifecycle_hook,
+)
+
+
+class MockPlugin(BaseGFLPlugin):
+    """Mock plugin for testing."""
+
+    def __init__(
+        self,
+        name: str = "mock",
+        version: str = "1.0.0",
+        priority: PluginPriority = PluginPriority.NORMAL,
+        dependencies: List[PluginDependency] = None,
+    ):
+        super().__init__()
+        self._name = name
+        self._version = version
+        self._priority = priority
+        self._dependencies = dependencies or []
+        self.load_called = False
+        self.unload_called = False
+        self.activate_called = False
+        self.deactivate_called = False
+
+    @property
+    def name(self) -> str:
+        return self._name
+
+    @property
+    def version(self) -> str:
+        return self._version
+
+    @property
+    def priority(self) -> PluginPriority:
+        return self._priority
+
+    @property
+    def dependencies(self) -> List[PluginDependency]:
+        return self._dependencies
+
+    def on_load(self) -> None:
+        self.load_called = True
+
+    def on_unload(self) -> None:
+        self.unload_called = True
+
+    def on_activate(self) -> None:
+        self.activate_called = True
+
+    def on_deactivate(self) -> None:
+        self.deactivate_called = True
+
+    def process(self, data: Dict[str, Any]) -> Dict[str, Any]:
+        result = data.copy()
+        result[f"processed_by_{self.name}"] = True
+        return result
+
+
+class TestPluginDependency:
+    """Test plugin dependency checking."""
+
+    def test_dependency_satisfied_import_available(self):
+        """Test dependency satisfaction when import is available."""
+        dep = PluginDependency("sys")  # sys is always available
+        assert dep.is_satisfied()
+
+    def test_dependency_not_satisfied_import_unavailable(self):
+        """Test dependency not satisfied when import unavailable."""
+        dep = PluginDependency("nonexistent_package_12345")
+        assert not dep.is_satisfied()
+
+    def test_optional_dependency_not_satisfied(self):
+        """Test optional dependency handling."""
+        dep = PluginDependency("nonexistent_package_12345", optional=True)
+        assert dep.is_satisfied()  # Optional deps are always "satisfied"
+
+    def test_version_spec_satisfied(self):
+        """Test version specification checking."""
+        # This is a basic test - in practice you'd need packaging library
+        PluginDependency("sys", ">=1.0.0")
+        # We can't easily test version checking without mocking
+        # In a real test, you'd mock the version() function
+
+
+class TestPluginInfo:
+    """Test enhanced PluginInfo functionality."""
+
+    def setup_method(self):
+        """Set up test fixtures."""
+        self.mock_plugin = MockPlugin()
+        self.plugin_info = PluginInfo(
+            name="test_plugin",
+            version="1.0.0",
+            plugin_class=MockPlugin,
+        )
+
+    def test_plugin_info_initialization(self):
+        """Test PluginInfo initialization."""
+        assert self.plugin_info.name == "test_plugin"
+        assert self.plugin_info.version == "1.0.0"
+        assert self.plugin_info.state == PluginState.UNLOADED
+        assert not self.plugin_info.is_loaded
+
+    def test_plugin_load_success(self):
+        """Test successful plugin loading."""
[…]
