### 2025-08-30 e33e591
fix: Address linting issues, security vulnerabilities, and test platform functionality
