+++ gfl/grammar_parser.py
@@ -22,6 +22,7 @@ from typing import Any, List, Optional
 try:
     import ply.lex as lex
     import ply.yacc as yacc
+
     HAS_PLY = True
 except ImportError:
     HAS_PLY = False
@@ -57,6 +58,7 @@ except ImportError:
     def cached(cache_name=None, ttl=None, max_size=None):
         def decorator(func):
             return func
+
         return decorator
 
     class MockMonitor:
@@ -64,12 +66,15 @@ except ImportError:
             class MockContext:
                 def __enter__(self):
                     return self
+
                 def __exit__(self, *args):
                     pass
+
             return MockContext()
 
     def get_monitor():
         return MockMonitor()
+
 
 logger = logging.getLogger(__name__)
 
@@ -89,111 +94,129 @@ class AdvancedGFLLexer:
     # Reserved words
     reserved = {
         # Core blocks
-        'experiment': 'EXPERIMENT',
-        'analyze': 'ANALYZE',
-        'analyse': 'ANALYZE',  # British spelling
-        'simulate': 'SIMULATE',
-        'branch': 'BRANCH',
-        'metadata': 'METADATA',
-
+        "experiment": "EXPERIMENT",
+        "analyze": "ANALYZE",
+        "analyse": "ANALYZE",  # British spelling
+        "simulate": "SIMULATE",
+        "branch": "BRANCH",
+        "metadata": "METADATA",
         # Control flow
-        'if': 'IF',
-        'then': 'THEN',
-        'else': 'ELSE',
-        'elif': 'ELIF',
-        'while': 'WHILE',
-        'for': 'FOR',
-        'in': 'IN',
-
+        "if": "IF",
+        "then": "THEN",
+        "else": "ELSE",
+        "elif": "ELIF",
+        "while": "WHILE",
+        "for": "FOR",
+        "in": "IN",
         # Logical operators
-        'and': 'AND',
-        'or': 'OR',
-        'not': 'NOT',
-
+        "and": "AND",
+        "or": "OR",
+        "not": "NOT",
         # Keywords
-        'tool': 'TOOL',
-        'type': 'TYPE',
-        'strategy': 'STRATEGY',
-        'params': 'PARAMS',
-        'parameters': 'PARAMS',  # Alias
-        'data': 'DATA',
-        'output': 'OUTPUT',
-        'input': 'INPUT',
-        'using': 'USING',
-        'with': 'WITH',
-        'as': 'AS',
-        'from': 'FROM',
-        'import': 'IMPORT',
-
+        "tool": "TOOL",
+        "type": "TYPE",
+        "strategy": "STRATEGY",
+        "params": "PARAMS",
+        "parameters": "PARAMS",  # Alias
+        "data": "DATA",
+        "output": "OUTPUT",
+        "input": "INPUT",
+        "using": "USING",
+        "with": "WITH",
+        "as": "AS",
+        "from": "FROM",
+        "import": "IMPORT",
         # Boolean values
-        'true': 'BOOLEAN',
-        'false': 'BOOLEAN',
-        'yes': 'BOOLEAN',
-        'no': 'BOOLEAN',
-        'on': 'BOOLEAN',
-        'off': 'BOOLEAN',
-
+        "true": "BOOLEAN",
+        "false": "BOOLEAN",
+        "yes": "BOOLEAN",
+        "no": "BOOLEAN",
+        "on": "BOOLEAN",
+        "off": "BOOLEAN",
         # Null values
-        'null': 'NULL',
-        'none': 'NULL',
-        'nil': 'NULL',
+        "null": "NULL",
+        "none": "NULL",
+        "nil": "NULL",
     }
 
     # List of token names
     tokens = [
         # Literals
[…]
