+++ gfl/enhanced_cli.py
@@ -28,11 +28,18 @@ from typing import Any, Dict, List, Optional
 try:
     from rich.console import Console
     from rich.table import Table
-    from rich.progress import Progress, SpinnerColumn, TextColumn, BarColumn, TaskProgressColumn
+    from rich.progress import (
+        Progress,
+        SpinnerColumn,
+        TextColumn,
+        BarColumn,
+        TaskProgressColumn,
+    )
     from rich.panel import Panel
     from rich.syntax import Syntax
     from rich.tree import Tree
     from rich.text import Text
+
     HAS_RICH = True
 except ImportError:
     HAS_RICH = False
@@ -51,28 +58,30 @@ try:
         list_plugins,
         activate_plugin,
         deactivate_plugin,
-        validate_plugin_dependencies
+        validate_plugin_dependencies,
     )
+
     HAS_PLUGINS = True
 except ImportError:
     HAS_PLUGINS = False
 
 try:
     from gfl.enhanced_schema_validator import EnhancedSchemaValidator
+
     HAS_ENHANCED_SCHEMA = True
 except ImportError:
     HAS_ENHANCED_SCHEMA = False
 
 # Configure logging
 logging.basicConfig(
-    level=logging.WARNING,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+    level=logging.WARNING, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 )
 logger = logging.getLogger(__name__)
 
 
 class CLIError(Exception):
     """Base exception for CLI errors."""
+
     pass
 
 
@@ -106,7 +115,9 @@ class OutputFormatter:
                 print(f"=== {title} ===")
             print(json_str)
 
-    def print_table(self, data: List[Dict[str, Any]], headers: List[str], title: str = None):
+    def print_table(
+        self, data: List[Dict[str, Any]], headers: List[str], title: str = None
+    ):
         """Print data as a table."""
         if self.use_rich:
             table = Table(title=title)
@@ -138,7 +149,11 @@ class OutputFormatter:
 
             # Print rows
             for row in data:
-                print(" | ".join(str(row.get(h, "")).ljust(w) for h, w in zip(headers, widths)))
+                print(
+                    " | ".join(
+                        str(row.get(h, "")).ljust(w) for h, w in zip(headers, widths)
+                    )
+                )
 
     def print_error(self, message: str, error: Exception = None):
         """Print error message."""
@@ -146,11 +161,13 @@ class OutputFormatter:
             self.console.print(f"[red]Error:[/red] {message}")
             if error and logger.isEnabledFor(logging.DEBUG):
                 import traceback
+
                 self.console.print(f"[dim]{traceback.format_exc()}[/dim]")
         else:
             print(f"Error: {message}", file=sys.stderr)
             if error and logger.isEnabledFor(logging.DEBUG):
                 import traceback
+
                 print(traceback.format_exc(), file=sys.stderr)
 
     def print_success(self, message: str):
@@ -176,7 +193,7 @@ class OutputFormatter:
                 BarColumn(),
                 TaskProgressColumn(),
                 console=self.console,
-                expand=True
+                expand=True,
             )
         else:
             # Simple fallback
@@ -184,12 +201,16 @@ class OutputFormatter:
                 def __enter__(self):
                     print(f"{description}...")
                     return self
+
                 def __exit__(self, *args):
                     pass
+
                 def add_task(self, description, total=None):
                     return 0
+
                 def update(self, task_id, completed=None, advance=None):
                     pass
+
             return SimpleProgress()
 
 
@@ -197,14 +218,14 @@ class GFLConfig:
     """Configuration management for GFL CLI."""
 
[…]
