### 2025-08-31 cac332f
Add comprehensive documentation structure with MkDocs
- Add complete documentation website with MkDocs Material theme
- Include installation guide, tutorial, and API reference
- Add export guide for documentation deployment
- Update existing documentation summaries
- Set up professional documentation structure
- Ready for GitHub Pages deployment
- Add GitHub Actions workflow for automatic docs deployment
