+++ docs/index.md
@@ -1,0 +1,129 @@
+# GeneForgeLang Documentation
+
+Welcome to the comprehensive documentation for GeneForgeLang (GFL), a domain-specific language for genomic workflow specification, validation, and AI-powered analysis.
+
+## 🚀 Quick Start
+
+- **[Installation Guide](installation.md)** - Get started with GeneForgeLang
+- **[Getting Started Tutorial](tutorial.md)** - Your first GFL workflow
+- **[API Reference](API_REFERENCE.md)** - Complete API documentation
+- **[CLI Tools Guide](cli.md)** - Command-line interface usage
+
+## 📖 Core Documentation
+
+### User Guides
+- **[Language Specification](language/)**
+  - [GFL Syntax Reference](language/syntax.md)
+  - [Workflow Examples](language/examples.md)
+  - [Best Practices](language/best-practices.md)
+
+- **[Web Interface](web-interface/)**
+  - [Getting Started](web-interface/getting-started.md)
+  - [Advanced Features](web-interface/advanced-features.md)
+  - [Troubleshooting](web-interface/troubleshooting.md)
+
+### Developer Documentation
+- **[API Documentation](api/)**
+  - [REST API Reference](API_REFERENCE.md)
+  - [Client SDK](api/client-sdk.md)
+  - [Authentication](api/authentication.md)
+
+- **[Architecture](architecture/)**
+  - [System Overview](architecture/overview.md)
+  - [Plugin System](architecture/plugins.md)
+  - [Inference Engine](architecture/inference.md)
+
+## 🔬 Advanced Features
+
+### AI & Machine Learning
+- **[Enhanced Inference Engine](inference/)**
+  - [Model Integration](ENHANCED_INFERENCE_SUMMARY.md)
+  - [Custom Models](inference/custom-models.md)
+  - [Performance Tuning](inference/performance.md)
+
+### Web & API Platform
+- **[Platform Features](platform/)**
+  - [Web Interface Overview](WEB_API_IMPLEMENTATION_SUMMARY.md)
+  - [Server Deployment](platform/deployment.md)
+  - [Monitoring & Analytics](platform/monitoring.md)
+
+## 🔒 Security & Operations
+
+- **[Security](security/)**
+  - [Security Advisory](SECURITY_ADVISORY.md)
+  - [Best Practices](security/best-practices.md)
+  - [Vulnerability Management](security/vulnerability-management.md)
+
+- **[Deployment](deployment/)**
+  - [Production Setup](deployment/production.md)
+  - [Docker & Kubernetes](deployment/containers.md)
+  - [Cloud Deployment](deployment/cloud.md)
+
+## 🛣️ Roadmap & Development
+
+- **[Development](development/)**
+  - [Contributing Guide](../CONTRIBUTING.md)
+  - [Development Setup](development/setup.md)
+  - [Testing Guide](development/testing.md)
+
+- **[Roadmap](roadmap/)**
+  - [Phase 4 Planning](PHASE_4_PLANNING.md)
+  - [Feature Requests](roadmap/feature-requests.md)
+  - [Release Notes](../CHANGELOG.md)
+
+## 📚 Reference Materials
+
+### Academic & Research
+- **[Research Paper](../paper.md)** - Academic publication
+- **[Citations](../CITATION.cff)** - How to cite GeneForgeLang
+- **[Bibliography](../paper.bib)** - Related research
+
+### Technical Specifications
+- **[Schema Reference](../schema/)** - JSON Schema definitions
+- **[Examples Collection](../examples/)** - Code examples and demos
+- **[Testing Suite](../tests/)** - Comprehensive test examples
+
+## 🎯 Use Cases & Applications
+
[…]

[…]
