### 2025-08-31 b5460c2
Fix MkDocs YAML configuration
