+++ mypy.ini
@@ -1,0 +1,45 @@
+[tool.mypy]
+python_version = "3.9"
+warn_return_any = false
+warn_unused_configs = true
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
+check_untyped_defs = false
+disallow_untyped_decorators = false
+no_implicit_optional = false
+warn_redundant_casts = false
+warn_unused_ignores = false
+warn_no_return = false
+warn_unreachable = false
+ignore_missing_imports = true
+implicit_reexport = true
+show_error_codes = true
+
+# Exclude problematic directories
+exclude = [
+    "^site/",
+    "^build/",
+    "^dist/",
+    "^docs/",
+    "^examples/",
+    "^scripts/",
+    "^tools/",
+    "^tests/",
+]
+
+# Per-module options
+[[tool.mypy.overrides]]
+module = [
+    "gfl.models.*",
+    "gfl.plugins.*", 
+    "gfl.nl.*",
+    "gfl.execution.*",
+    "applications.*",
+    "*.test_*",
+]
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "tests.*"
+ignore_missing_imports = true
+disallow_untyped_defs = false
