### 2025-08-31 c1d8842
Trigger GitHub Pages deployment
- Update documentation index to trigger Pages workflow
- Test GitHub Pages automatic deployment
