### 2025-08-31 670cce6
Enhance README with comprehensive project overview
- Add modern badges including GitHub Pages documentation link
- Include detailed feature overview with emojis and clear sections
- Add comprehensive quick start guide with code examples
- Expand use cases for genomics, clinical, and agricultural applications
- Include detailed CLI tools documentation
- Add repository structure and web applications guide
- Enhance security, performance, and community sections
- Update roadmap with current achievements and future plans
- Add proper citation format and improved contributing guidelines
- Link all 
[…]
