+++ README.md
@@ -1,100 +1,312 @@
-GeneForgeLang (GFL)
+# GeneForgeLang (GFL) 🧬
 
 [![CI](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/actions/workflows/ci.yml/badge.svg)](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/actions/workflows/ci.yml)
-[![Documentation](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/actions/workflows/docs.yml/badge.svg)](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/actions/workflows/docs.yml)
+[![Documentation](https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/actions/workflows/docs.yml/badge.svg)](https://fundacion-de-neurociencias.github.io/GeneForgeLang/)
+[![GitHub Pages](https://img.shields.io/badge/docs-GitHub%20Pages-blue)](https://fundacion-de-neurociencias.github.io/GeneForgeLang/)
+[![Python](https://img.shields.io/badge/python-3.9%2B-blue)](https://www.python.org/downloads/)
 [![Code Style: Black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
-[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](license)
-[![PyPI](https://img.shields.io/badge/PyPI-coming%20soon-lightgrey)](https://pypi.org/project/geneforgelang/)
+[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](LICENSE)
+[![Security](https://img.shields.io/badge/security-bandit-yellow)](https://github.com/PyCQA/bandit)
+[![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
 
-GFL is a domain-specific language for specifying, validating, and reasoning about genomic workflows and experiments. It is intended primarily as a library for other applications to embed, with a few small demo apps included.
+> **A powerful Domain-Specific Language (DSL) for genomic workflows and bioinformatics applications with AI-powered analysis capabilities.**
+
+GeneForgeLang (GFL) is a comprehensive framework for specifying, validating, and executing genomic workflows. It combines the simplicity of YAML-like syntax with advanced features like AI-powered inference, plugin extensibility, and web-based interfaces.
+
+## ✨ Key Features
+
+🔬 **Genomic Workflow Specification** - Declarative YAML-like syntax for complex genomic experiments
+🤖 **AI-Powered Analysis** - Built-in inference engine with machine learning capabilities
+🌐 **Web Interface** - Modern web platform for interactive workflow creation and execution
+🔌 **Plugin System** - Extensible architecture for custom tools and integrations
+⚡ **High Performance** - Optimized for large-scale genomic data processing
+🔒 **Secure & Robust** - Comprehensive security features and error handling
+
+## 🚀 Quick Start
+
+### Installation
+
+```bash
+# Basic installation
+pip install -e .
+
+# With all features
+pip install -e .[full]
+
+# Optional extras
+pip install -e .[apps]     # Demo applications with Gradio
+pip install -e .[ml]       # Machine learning capabilities
+pip install -e .[server]   # Web server and API
+```
+
+### Your First GFL Workflow
+
+```python
+from gfl.api import parse, validate, infer
+
+# Define a CRISPR gene editing experiment
+workflow = """
+experiment:
+  tool: CRISPR_cas9
+  type: gene_editing
+  params:
+    target_gene: TP53
[…]
