### 2025-08-31 c73c7ab
Phase 3: Complete Plugin Ecosystem Implementation
Major Features Added:
 Comprehensive Plugin Interface System
- GeneratorPlugin, OptimizerPlugin, PriorsPlugin abstract base classes
- Specialized interfaces: SequenceGeneratorPlugin, MoleculeGeneratorPlugin, BayesianOptimizerPlugin
- Plugin registry with automatic discovery and dependency management
- Example implementations: ProteinVAEGenerator, MoleculeTransformerGenerator, BayesianOptimizer

 Workflow Execution Engine
- Intelligent dispatch system for design and optimize blocks
- Parameter injection mechanism with \ syntax support
- Comprehe
[…]
