+++ test_plugin_ecosystem.py
@@ -1,0 +1,305 @@
+#!/usr/bin/env python3
+"""
+Test the complete plugin ecosystem with real GFL execution.
+"""
+
+from gfl.api import parse, validate, execute, list_available_plugins, get_api_info
+from gfl.plugins.example_implementations import register_example_plugins
+
+
+def test_plugin_ecosystem():
+    """Test the complete plugin ecosystem functionality."""
+    
+    print("🚀 Testing GeneForgeLang Plugin Ecosystem")
+    print("=" * 60)
+    
+    # 1. Register example plugins
+    print("\n1. Registering example plugins...")
+    try:
+        register_example_plugins()
+        print("✓ Successfully registered example plugins")
+    except Exception as e:
+        print(f"❌ Failed to register plugins: {e}")
+        return False
+    
+    # 2. Check API info and available plugins
+    print("\n2. Checking API capabilities...")
+    api_info = get_api_info()
+    print(f"✓ API version: {api_info['api_version']}")
+    print(f"✓ Workflow execution: {api_info['features']['workflow_execution']}")
+    print(f"✓ Plugin system: {api_info['features']['plugin_system']}")
+    
+    available_plugins = list_available_plugins()
+    print(f"✓ Available generators: {available_plugins['generators']}")
+    print(f"✓ Available optimizers: {available_plugins['optimizers']}")
+    
+    # 3. Test design block execution
+    print("\n3. Testing design block execution...")
+    design_gfl = """
+metadata:
+  experiment_id: DESIGN_TEST_001
+  researcher: Dr. Test User
+  project: plugin_ecosystem_test
+  description: Testing protein design with VAE generator
+
+design:
+  entity: ProteinSequence
+  model: ProteinVAEGenerator
+  objective:
+    maximize: stability
+    target: therapeutic_protein
+  constraints:
+    - length(50, 150)
+    - synthesizability > 0.8
+    - stability_score > 0.7
+  count: 5
+  output: designed_proteins
+"""
+    
+    try:
+        # Parse GFL
+        ast = parse(design_gfl)
+        print("✓ Successfully parsed design GFL")
+        
+        # Validate AST
+        errors = validate(ast)
+        if errors:
+            print(f"❌ Validation errors: {errors}")
+            return False
+        print("✓ AST validation passed")
+        
+        # Execute design block
+        result = execute(ast)
+        print("✓ Design block executed successfully")
+        
+        # Check results
+        design_result = result['design']
+        candidates = design_result['candidates']
+        
+        print(f"✓ Generated {len(candidates)} protein candidates")
+        print(f"✓ First candidate: {candidates[0].sequence[:20]}...")
+        print(f"✓ First candidate properties: {candidates[0].properties}")
+        print(f"✓ First candidate confidence: {candidates[0].confidence:.3f}")
+        
+    except Exception as e:
+        print(f"❌ Design block execution failed: {e}")
+        return False
+    
+    # 4. Test optimize block execution
+    print("\n4. Testing optimize block execution...")
+    optimize_gfl = """
+metadata:
+  experiment_id: OPTIMIZE_TEST_001
+  researcher: Dr. Test User
+  project: plugin_ecosystem_test
+  description: Testing CRISPR parameter optimization
+
+optimize:
+  search_space:
+    temperature: range(25, 42)
+    guide_concentration: range(10, 100)
+    duration: choice([6, 12, 24, 48])
+    cas9_ratio: range(0.5, 2.0)
+
+  strategy:
+    name: BayesianOptimization
+    uncertainty_metric: entropy
+    initial_samples: 3
+    batch_size: 1
+    exploration_weight: 0.1
+
+  objective:
+    maximize: editing_efficiency
+
+  budget:
+    max_experiments: 10
+    convergence_threshold: 0.01
+
+  run:
[…]
