+++ tests/test_new_features.py
@@ -1,0 +1,710 @@
+"""
+Comprehensive regression test suite for GeneForgeLang new features.
+
+This module contains pytest test cases to validate the correct parsing and 
+semantic validation of the recently implemented design and optimize blocks.
+These tests prevent future updates from breaking existing functionality.
+"""
+
+import pytest
+
+from gfl.api import parse, validate
+
+
+class TestNewFeaturesRegression:
+    """Regression tests for new GFL features."""
+
+    def test_valid_gfl_script_with_all_new_features(self):
+        """Test that a valid GFL script using all new features parses and validates correctly."""
+        
+        gfl_script = """
+        metadata:
+          experiment_id: COMPREHENSIVE_TEST_001
+          researcher: Dr. Test Suite
+          project: feature_validation
+          description: Comprehensive test of all new GFL features
+
+        design:
+          entity: ProteinSequence
+          model: ProteinGeneratorVAE
+          objective:
+            maximize: binding_affinity
+            target: ACE2_receptor
+          constraints:
+            - length(120, 150)
+            - synthesizability > 0.7
+          count: 25
+          output: designed_sequences
+
+        analyze:
+          strategy: functional
+          data: designed_sequences
+          thresholds:
+            binding_score: 0.8
+            stability_score: 0.7
+          operations:
+            - type: sort
+              params:
+                by: binding_affinity
+                order: descending
+
+        optimize:
+          search_space:
+            temperature: range(25, 42)
+            concentration: range(10, 100)
+            duration: choice([6, 12, 24, 48])
+            buffer_ph: range(6.5, 8.5)
+          strategy:
+            name: ActiveLearning
+            uncertainty_metric: entropy
+            initial_samples: 8
+          objective:
+            maximize: reaction_efficiency
+          budget:
+            max_experiments: 100
+            max_time: 48h
+            convergence_threshold: 0.01
+          run:
+            experiment:
+              tool: PCR
+              type: validation
+              params:
+                temp: ${temperature}
+                conc: ${concentration}
+                time: ${duration}h
+                ph: ${buffer_ph}
+                target_gene: GFP
+                replicates: 3
+        """
+        
+        # Test parsing
+        ast = parse(gfl_script)
+        assert ast is not None, "Valid GFL script should parse successfully"
+        
+        # Verify all blocks are present
+        assert "metadata" in ast, "Metadata block should be present"
+        assert "design" in ast, "Design block should be present"
+        assert "analyze" in ast, "Analyze block should be present"
+        assert "optimize" in ast, "Optimize block should be present"
+        
+        # Test validation
+        errors = validate(ast)
+        assert not errors, f"Valid GFL script should validate without errors, got: {errors}"
+
+
+class TestDesignBlockValidation:
+    """Regression tests for design block validation."""
+
+    def test_design_block_missing_objective_field(self):
+        """Test that design block without objective field fails validation."""
+        
+        gfl_script = """
+        design:
+          entity: ProteinSequence
+          model: ProteinGeneratorVAE
+          # Missing objective field
+          count: 10
+          output: designed_proteins
+        """
+        
+        ast = parse(gfl_script)
+        assert ast is not None, "Script should parse even with missing fields"
+        
+        errors = validate(ast)
+        assert len(errors) > 0, "Missing objective field should cause validation error"
+        
+        # Check that error mentions objective
+        error_text = " ".join(str(error) for error in errors).lower()
+        assert "objective" in error_text, "Error should mention missing objective field"
+
+    def test_design_block_missing_entity_field(self):
+        """Test that design block without entity field fails validation."""
+        
+        gfl_script = """
+        design:
+          # Missing entity field
+          model: ProteinGeneratorVAE
+          objective:
+            maximize: binding_affinity
[…]
