### 2025-08-31 12c43fe
Fix type checking issues and improve mypy configuration
- Fixed mypy.ini format from TOML to proper INI syntax
- Added type: ignore annotations for conditional imports in api.py and inference_engine.py
- Fixed SourceLocation constructor calls in grammar_parser.py (file_path instead of filename)
- Fixed EnhancedValidationResult constructor calls in api.py and grammar_parser.py
- Added ast, syntax_errors, semantic_errors properties to EnhancedValidationResult
- Enhanced mypy configuration with better module-specific overrides
- Improved type safety while maintaining backward compatibility

These
[…]
