+++ tests/test_new_features.py
@@ -1,7 +1,7 @@
 """
 Comprehensive regression test suite for GeneForgeLang new features.
 
-This module contains pytest test cases to validate the correct parsing and 
+This module contains pytest test cases to validate the correct parsing and
 semantic validation of the recently implemented design and optimize blocks.
 These tests prevent future updates from breaking existing functionality.
 """
@@ -16,7 +16,7 @@ class TestNewFeaturesRegression:
 
     def test_valid_gfl_script_with_all_new_features(self):
         """Test that a valid GFL script using all new features parses and validates correctly."""
-        
+
         gfl_script = """
         metadata:
           experiment_id: COMPREHENSIVE_TEST_001
@@ -76,17 +76,17 @@ class TestNewFeaturesRegression:
                 target_gene: GFP
                 replicates: 3
         """
-        
+
         # Test parsing
         ast = parse(gfl_script)
         assert ast is not None, "Valid GFL script should parse successfully"
-        
+
         # Verify all blocks are present
         assert "metadata" in ast, "Metadata block should be present"
         assert "design" in ast, "Design block should be present"
         assert "analyze" in ast, "Analyze block should be present"
         assert "optimize" in ast, "Optimize block should be present"
-        
+
         # Test validation
         errors = validate(ast)
         assert not errors, f"Valid GFL script should validate without errors, got: {errors}"
@@ -97,7 +97,7 @@ class TestDesignBlockValidation:
 
     def test_design_block_missing_objective_field(self):
         """Test that design block without objective field fails validation."""
-        
+
         gfl_script = """
         design:
           entity: ProteinSequence
@@ -106,20 +106,20 @@ class TestDesignBlockValidation:
           count: 10
           output: designed_proteins
         """
-        
+
         ast = parse(gfl_script)
         assert ast is not None, "Script should parse even with missing fields"
-        
+
         errors = validate(ast)
         assert len(errors) > 0, "Missing objective field should cause validation error"
-        
+
         # Check that error mentions objective
         error_text = " ".join(str(error) for error in errors).lower()
         assert "objective" in error_text, "Error should mention missing objective field"
 
     def test_design_block_missing_entity_field(self):
         """Test that design block without entity field fails validation."""
-        
+
         gfl_script = """
         design:
           # Missing entity field
@@ -129,17 +129,17 @@ class TestDesignBlockValidation:
           count: 10
           output: designed_proteins
         """
-        
+
         ast = parse(gfl_script)
         errors = validate(ast)
         assert len(errors) > 0, "Missing entity field should cause validation error"
-        
+
         error_text = " ".join(str(error) for error in errors).lower()
         assert "entity" in error_text, "Error should mention missing entity field"
 
     def test_design_block_missing_model_field(self):
         """Test that design block without model field fails validation."""
-        
+
         gfl_script = """
         design:
           entity: ProteinSequence
@@ -149,17 +149,17 @@ class TestDesignBlockValidation:
           count: 10
           output: designed_proteins
         """
-        
+
         ast = parse(gfl_script)
         errors = validate(ast)
         assert len(errors) > 0, "Missing model field should cause validation error"
-        
+
         error_text = " ".join(str(error) for error in errors).lower()
         assert "model" in error_text, "Error should mention missing model field"
 
     def test_design_block_invalid_count_negative(self):
         """Test that design block with negative count fails validation."""
-        
+
         gfl_script = """
         design:
           entity: ProteinSequence
@@ -169,17 +169,17 @@ class TestDesignBlockValidation:
           count: -1
           output: designed_proteins
         """
-        
+
         ast = parse(gfl_script)
         errors = validate(ast)
[…]
