+++ gfl/execution_engine.py
@@ -6,7 +6,7 @@ types and configurations.
 
 The engine integrates:
 1. Design block execution via GeneratorPlugin instances
-2. Optimize block execution via OptimizerPlugin instances  
+2. Optimize block execution via OptimizerPlugin instances
 3. Result aggregation and workflow state management
 4. Error handling and recovery mechanisms
 """
@@ -33,7 +33,7 @@ logger = logging.getLogger(__name__)
 
 class ExecutionError(Exception):
     """Exception raised during GFL block execution."""
-    
+
     def __init__(self, message: str, block_type: str, plugin_name: Optional[str] = None):
         super().__init__(message)
         self.block_type = block_type
@@ -42,20 +42,20 @@ class ExecutionError(Exception):
 
 class WorkflowState:
     """Manages state across workflow execution."""
-    
+
     def __init__(self):
         self.variables: Dict[str, Any] = {}
         self.execution_history: List[Dict[str, Any]] = []
         self.current_block: Optional[str] = None
-        
+
     def set_variable(self, name: str, value: Any) -> None:
         """Set a workflow variable."""
         self.variables[name] = value
-        
+
     def get_variable(self, name: str) -> Any:
         """Get a workflow variable."""
         return self.variables.get(name)
-        
+
     def add_execution_record(self, block_type: str, block_config: Dict[str, Any], result: Any) -> None:
         """Add execution record to history."""
         record = {
@@ -69,19 +69,19 @@ class WorkflowState:
 
 class GFLExecutionEngine:
     """Core execution engine for GFL workflow blocks."""
-    
+
     def __init__(self):
         self.state = WorkflowState()
-        
+
     def execute_design_block(self, design_config: Dict[str, Any]) -> List[DesignCandidate]:
         """Execute a design block using appropriate GeneratorPlugin.
-        
+
         Args:
             design_config: Design block configuration from GFL AST
-            
+
         Returns:
             List of generated DesignCandidate objects
-            
+
         Raises:
             ExecutionError: If design execution fails
         """
@@ -93,10 +93,10 @@ class GFLExecutionEngine:
             count = design_config["count"]
             output_var = design_config["output"]
             constraints = design_config.get("constraints", [])
-            
+
             # Find appropriate generator plugin
             generators = get_available_generators()
-            
+
             if model_name not in generators:
                 available = ", ".join(generators.keys())
                 raise ExecutionError(
@@ -104,9 +104,9 @@ class GFLExecutionEngine:
                     "design",
                     model_name
                 )
-            
+
             generator = generators[model_name]
-            
+
             # Validate plugin supports the requested entity type
             if not self._validate_entity_support(generator, entity):
                 supported = [e.value for e in generator.supported_entities]
@@ -116,17 +116,17 @@ class GFLExecutionEngine:
                     "design",
                     model_name
                 )
-            
+
             # Validate objective and constraints
             obj_errors = generator.validate_objective(objective)
             constraint_errors = generator.validate_constraints(constraints)
-            
+
             if obj_errors or constraint_errors:
                 error_msg = "Validation failed: " + "; ".join(obj_errors + constraint_errors)
                 raise ExecutionError(error_msg, "design", model_name)
-            
+
             logger.info(f"Executing design block with {model_name} for {count} {entity} candidates")
-            
+
             # Execute generation
             start_time = time.time()
             candidates = generator.generate(
@@ -136,35 +136,35 @@ class GFLExecutionEngine:
                 count=count
             )
             execution_time = time.time() - start_time
-            
+
             # Store results in workflow state
             self.state.set_variable(output_var, candidates)
-            
+
             # Record execution
             self.state.add_execution_record("design", design_config, {
                 "candidate_count": len(candidates),
                 "execution_time": execution_time,
[…]
