### 2025-08-31 1706659
Final type checking improvements - Fixed EnhancedValidationResult syntax_errors and semantic_errors properties - Added proper setters for backward compatibility - Improved property handling for both assignment and dynamic computation - Basic functionality tests passing - Reduced MyPy errors from 80+ to manageable library stub issues
