### 2025-08-31 80ee682
Add GFL service implementation
