+++ roadmap_phase_3/gfl-genesis/plugins/gfl-plugin-ontarget-scorer/gfl_plugin_ontarget_scorer/__init__.py
@@ -1,0 +1,3 @@
+"""
+GFL Plugin for On-Target Scoring
+"""
