+++ roadmap_phase_3/gfl-genesis/plugins/gfl-crispr-evaluator/gfl_crispr_evaluator/__init__.py
@@ -1,0 +1,3 @@
+"""
+GFL Plugin for CRISPR Evaluation Orchestration
+"""
