+++ gfl/error_handling.py
@@ -363,6 +363,16 @@ class ErrorCodes:
     SEMANTIC_MISSING_EXPERIMENT_BLOCK = "SEMANTIC006"
     SEMANTIC_VARIABLE_REDEFINITION = "SEMANTIC007"
     SEMANTIC_UNDEFINED_VARIABLE = "SEMANTIC008"
+    
+    # IO Contract errors (SEMANTIC100-SEMANTIC199)
+    SEMANTIC_INVALID_CONTRACT = "SEMANTIC100"
+    SEMANTIC_CONTRACT_MISMATCH = "SEMANTIC101"
+    SEMANTIC_MISSING_CONTRACT = "SEMANTIC102"
+    
+    # Schema Registry errors (SEMANTIC200-SEMANTIC299)
+    SEMANTIC_INVALID_SCHEMA_FILE = "SEMANTIC200"
+    SEMANTIC_SCHEMA_NOT_FOUND = "SEMANTIC201"
+    SEMANTIC_INVALID_SCHEMA_DEFINITION = "SEMANTIC202"
 
     # Type errors (TYPE001-TYPE099)
     TYPE_INVALID_TYPE = "TYPE001"
