### 2025-08-31 #8281 @abhishektiwari
@metelkin @shahmoradi @gchure - would any of you be willing to review this submission for Journal of Open Source Software (JOSS)?

The submission under consideration is "GeneForgeLang: A Symbolic Language for Causal Reasoning in Genetic Engineering and Molecular Bioinformatics".

Thanks
Abhi
