### 2025-09-01 #8281 @metelkin
@abhishektiwari Thank you for the invitation to review. I’ve skimmed the submission, but a Python environment is not my usual stack. I can assess the scientific content and formal criteria, but I’m not confident evaluating Python-specific conventions or running the package. If a co-reviewer can cover the Python/installation/testing aspects, I’m happy to proceed, otherwise, I should decline to avoid a review that may not do the submission justice.
