### 2025-09-02 3662c7b
Add Publications section to README with scientific references
