### 2025-09-02 e444af7
Add symbolic reasoning capabilities and update documentation
