+++ gfl/schema_loader.py
@@ -24,10 +24,10 @@ logger = logging.getLogger(__name__)
 
 class SchemaDefinition:
     """Represents a schema definition loaded from a YAML file."""
-    
+
     def __init__(self, name: str, type_info: Dict[str, Any]):
         """Initialize a schema definition.
-        
+
         Args:
             name: The name of the schema type
             type_info: Dictionary containing type and attributes information
@@ -36,18 +36,18 @@ class SchemaDefinition:
         self.base_type = type_info.get("type")
         self.attributes = type_info.get("attributes", {})
         self.description = type_info.get("description", "")
-        
+
     def validate_definition(self, result: EnhancedValidationResult) -> bool:
         """Validate the schema definition structure.
-        
+
         Args:
             result: Validation result to add errors to
-            
+
         Returns:
             True if valid, False otherwise
         """
         is_valid = True
-        
+
         if not self.base_type:
             error = result.add_error(
                 f"Schema '{self.name}' missing required 'type' field",
@@ -56,7 +56,7 @@ class SchemaDefinition:
             )
             error.add_fix(f"Add 'type: <base_type>' to schema '{self.name}'")
             is_valid = False
-            
+
         if not isinstance(self.base_type, str):
             error = result.add_error(
                 f"Schema '{self.name}' type must be a string",
@@ -65,7 +65,7 @@ class SchemaDefinition:
             )
             error.add_fix(f"Change type to a string value in schema '{self.name}'")
             is_valid = False
-            
+
         if not isinstance(self.attributes, dict):
             error = result.add_error(
                 f"Schema '{self.name}' attributes must be a dictionary",
@@ -74,7 +74,7 @@ class SchemaDefinition:
             )
             error.add_fix(f"Format attributes as a dictionary in schema '{self.name}'")
             is_valid = False
-            
+
         # Validate each attribute definition
         for attr_name, attr_def in self.attributes.items():
             if not isinstance(attr_def, dict):
@@ -86,7 +86,7 @@ class SchemaDefinition:
                 error.add_fix(f"Format attribute '{attr_name}' as a dictionary with type and required fields")
                 is_valid = False
                 continue
-                
+
             # Check required fields in attribute definition
             if "type" not in attr_def:
                 error = result.add_error(
@@ -96,30 +96,30 @@ class SchemaDefinition:
                 )
                 error.add_fix(f"Add 'type: <attr_type>' to attribute '{attr_name}' in schema '{self.name}'")
                 is_valid = False
-                
+
         return is_valid
 
 
 class SchemaLoader:
     """Loader for external schema definition files."""
-    
+
     def __init__(self):
         """Initialize the schema loader."""
         self.schemas: Dict[str, SchemaDefinition] = {}
         self.loaded_files: List[str] = []
-        
+
     def load_schema_file(self, file_path: Union[str, Path], result: EnhancedValidationResult) -> bool:
         """Load and parse a schema definition file.
-        
+
         Args:
             file_path: Path to the schema definition file
             result: Validation result to add errors to
-            
+
         Returns:
             True if loaded successfully, False otherwise
         """
         file_path = Path(file_path)
-        
+
         # Check if file exists
         if not file_path.exists():
             error = result.add_error(
@@ -129,17 +129,17 @@ class SchemaLoader:
             )
             error.add_fix(f"Ensure schema file exists at {file_path}")
             return False
-            
+
         # Check if already loaded
         if str(file_path) in self.loaded_files:
             logger.debug(f"Schema file already loaded: {file_path}")
             return True
-            
+
         try:
             # Load YAML content
             with open(file_path, "r", encoding="utf-8") as f:
                 content = yaml.safe_load(f)
-                
+
             if not isinstance(content, dict):
[…]
