+++ gfl/semantic_validator.py
@@ -55,7 +55,7 @@ class EnhancedSemanticValidator:
         try:
             # Load schema imports first
             self._load_schema_imports(ast)
-            
+
             self._validate_root_structure(ast)
             self._validate_blocks(ast)
         except Exception as e:
@@ -70,13 +70,13 @@ class EnhancedSemanticValidator:
 
     def _load_schema_imports(self, ast: Dict[str, Any]) -> None:
         """Load schema imports from the AST.
-        
+
         Args:
             ast: The AST dictionary to process.
         """
         if "import_schemas" not in ast:
             return
-            
+
         import_schemas = ast["import_schemas"]
         if not isinstance(import_schemas, list):
             error = self.result.add_error(
@@ -85,10 +85,10 @@ class EnhancedSemanticValidator:
             )
             error.add_fix("Format import_schemas as a list of schema file paths")
             return
-            
+
         # Resolve file paths relative to the GFL file location
         base_path = os.path.dirname(self.result.file_path) if self.result.file_path else "."
-        
+
         schema_files = []
         for schema_path in import_schemas:
             if not isinstance(schema_path, str):
@@ -98,15 +98,15 @@ class EnhancedSemanticValidator:
                 )
                 error.add_fix("Use string values for schema file paths")
                 continue
-                
+
             # Resolve relative paths
             if not os.path.isabs(schema_path):
                 resolved_path = os.path.join(base_path, schema_path)
             else:
                 resolved_path = schema_path
-                
+
             schema_files.append(resolved_path)
-            
+
         # Load all schema files
         load_schemas_from_files(schema_files, self.result)
 
@@ -121,7 +121,7 @@ class EnhancedSemanticValidator:
             return
 
         # Check for at least one main block
-        main_blocks = {"experiment", "analyze", "simulate", "branch", "design", "optimize", "refine_data", "guided_discovery"}
+        main_blocks = {"experiment", "analyze", "simulate", "branch", "design", "optimize", "refine_data", "guided_discovery", "rules", "hypothesis", "timeline", "pathways", "complexes"}
         found_blocks = set(ast.keys()) & main_blocks
 
         if not found_blocks:
@@ -131,7 +131,7 @@ class EnhancedSemanticValidator:
                 ErrorSeverity.ERROR,
             )
             error.add_fix(
-                "Add an 'experiment', 'analyze', 'simulate', 'design', 'optimize', 'refine_data', or 'guided_discovery' block"
+                "Add an 'experiment', 'analyze', 'simulate', 'design', 'optimize', 'refine_data', 'guided_discovery', 'rules', 'hypothesis', 'timeline', 'pathways', or 'complexes' block"
             )
             error.add_context("available_blocks", list(main_blocks))
 
@@ -150,6 +150,12 @@ class EnhancedSemanticValidator:
 
     def _validate_blocks(self, ast: Dict[str, Any]) -> None:
         """Validate individual blocks in the AST."""
+        # First, collect entity definitions for reference validation
+        self._collect_entity_definitions(ast)
+        
+        # Collect hypothesis definitions for reference validation
+        self._collect_hypothesis_definitions(ast)
+        
         for block_name, block_content in ast.items():
             self.current_block = block_name
 
@@ -177,6 +183,292 @@ class EnhancedSemanticValidator:
                 self._validate_guided_discovery_block(block_content)
             elif block_name == "metadata":
                 self._validate_metadata_block(block_content)
+            elif block_name == "rules":
+                self._validate_rules_block(block_content)
+            elif block_name == "hypothesis":
+                self._validate_hypothesis_block(block_content)
[…]
