### 2025-09-02 64344a5
Add symbolic reasoning capabilities and update documentation
