### 2025-09-02 0b86474
Update GFL specification to include symbolic reasoning capabilities
