### 2025-09-03 #8281 @abhishektiwari
👋👋👋👋 Hi @josecediel25 @jeskowagner @jnarayan81 @telatin, would any of you be open to review this submission for Journal of Open Source Software (JOSS)?

The submission under consideration is "GeneForgeLang: A Symbolic Language for Causal Reasoning in Genetic Engineering and Molecular Bioinformatics".

Thanks
Abhi
