### 2025-09-15 d654ab7
Update paper.md
