+++ paper.bib
@@ -1,17 +1,16 @@
-
 @article{Galdzicki2014,
-  title={Standard Operating Procedures for the Synthetic Biology Open Language (SBOL)},
+  title={The Synthetic Biology Open Language (SBOL) provides a community standard for communicating designs in synthetic biology},
   author={Galdzicki, Michal and Clancy, Kevin P and Oberortner, Ernst and Pocock, Matthew and Quinn, Jason Y and Rodriguez, Carlos A and Roehner, Nicolas and Wilson, Michael L and Bartley, Bryan A and Beal, Jacob and others},
-  journal={PloS One},
-  volume={9},
-  number={10},
-  pages={e112389},
+  journal={Nature Biotechnology},
+  volume={32},
+  number={6},
+  pages={545--550},
   year={2014},
-  publisher={Public Library of Science}
+  publisher={Nature Publishing Group}
 }
 
 @article{LeDuc2022,
-  title={ProForma 2.0: unifying the encoding of proteoforms and peptidoforms},
+  title={ProForma 2.0: Unifying the Encoding of Proteoforms and Peptidoforms},
   author={LeDuc, Richard D and Deutsch, Eric W and Binz, Pierre-Alain and Fellers, Ryan T and Cesnik, Ajla J and Klein, Joshua A and Van Den Bossche, Tim and Gabriels, Ralf and Yalavarthi, Aditya and Perez-Riverol, Yasset and others},
   journal={Journal of Proteome Research},
   volume={21},
@@ -21,21 +20,24 @@
   publisher={ACS Publications}
 }
 
-@misc{GenoCAD,
-  title = {GenoCAD: A Design Tool for Synthetic Biology},
-  author = {Czar, Michael J and Cai, Yizhi and Peccoud, Jean},
-  year = {2009},
-  howpublished = {Journal of Biological Engineering, 3(1):1},
-  url = {https://jbioleng.biomedcentral.com/articles/10.1186/1754-1611-3-1}
+@article{Czar2009,
+  title={GenoCAD: A design tool for synthetic biology},
+  author={Czar, Michael J and Cai, Yizhi and Peccoud, Jean},
+  journal={Journal of Biological Engineering},
+  volume={3},
+  number={1},
+  pages={1--12},
+  year={2009},
+  publisher={Springer}
 }
 
-@inproceedings{Eugene2011,
-  author    = {Lesia Bilitchenko,Adam Liu,Sherine Cheung,Emma Weeding,Bing Xia,Mariana Leguia,J. Christopher Anderson,Douglas Densmore},
-  title     = {Eugene – A Domain Specific Language for Specifying and Constraining Synthetic Biological Parts, Devices, and Systems},
-  booktitle = {PLOS ONE},
-  year    = {2011},
-  volume  = {6},
-  number  = {4},
-  pages   = {e18882},
-  doi     = {10.1371/journal.pone.0018882}
+@article{Bilitchenko2011,
+  title={Eugene--a domain specific language for specifying and constraining synthetic biological parts, devices, and systems},
+  author={Bilitchenko, Lesia and Liu, Adam and Cheung, Sherine and Weeding, Emma and Xia, Bing and Leguia, Mariana and Anderson, J Christopher and Densmore, Douglas},
+  journal={PloS one},
+  volume={6},
+  number={4},
+  pages={e18882},
+  year={2011},
+  publisher={Public Library of Science San Francisco, USA}
 }
