### 2025-09-18 #8927 @manuelmenendezgonzalez
Dear Abhishek,

Thank you for your thorough evaluation of our GeneForgeLang submission. We
appreciate your detailed feedback and take your concerns seriously.

We believe there may have been an issue with repository synchronization at
the time of your review, which could have resulted in you evaluating an
outdated version of our project. Since your initial review, we have
undertaken significant updates to address the points you raised.
Specifically, we have:

- Completely removed all LLM-generated artifacts and placeholder
implementations: We have replaced all mock functions with fully functional
implementations based on real scientific approaches, utilizing CFD scoring
for off-target prediction and the DeepHF model for on-target efficiency.
- Fixed all non-functional examples: All code examples now work as
described, with proper plugin implementations that adhere to the correct
GFL plugin architecture.
- Addressed security issues: We have resolved all security warnings
identified by the bandit scanner, particularly those regarding random
number generation.
- Removed hallucinated references: We have corrected all bibliography
entries to ensure they reference legitimate and verifiable publications.
- Updated documentation: The `paper.md` and all associated documents now
accurately reflect the current state of the project.
- Created a complete development container setup: We have added a full
development container configuration to ensure consistent development
environments and facilitate easier reproduction of our work.

We have now synchronized all repositories and can confirm that the latest
version of our codebase, documentation, and paper accurately represents our
work. Given these comprehensive updates, we kindly ask you to review our
submission again with the updated materials.

The current repository contains fully functional implementations, proper
scientific references, and addresses all technical issues you identified.
We believe these updates demonstrate the scholarly merit and technical
rigor of our work.

Thank you for your time and consideration.

Sincerely,
Manuel

On Thu, 18 Sept 2025 at 13:29, Abhishek Tiwari ***@***.***>
wrote:

> *abhishektiwari* left a comment (openjournals/joss-reviews#8927)
> <https://github.com/openjournals/joss-reviews/issues/8927#issuecomment-3306948103>
>
> @manuelmenendezgonzalez <https://github.com/manuelmenendezgonzalez> Upon
> closer look at GeneForgeLang
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang> repository,
> I see a lot of evidence that appears to suggest that a large part of
> software and associated documents are LLM-generated, with numerous
> technical issues including non-functional examples, hallucinated references
> and security risk, repeated duplications, and placeholder functions.
>
> I am going to send this submission for a scope and scholarly merit review
> to our EiCs and request them to provide guidance inline with our interim
> position on AI written software
> <https://joss.readthedocs.io/en/latest/submitting.html#interim-position-on-ai-enabled-software>
> .
>
