+++ roadmap_phase_3/gfl-genesis/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Test package for GFL Genesis Project
-"""
