+++ roadmap_phase_3/gfl-genesis/plugins/gfl-plugin-offtarget-scorer/gfl_plugin_offtarget_scorer/plugin.py
@@ -1,81 +1,0 @@
-"""
-Off-Target Scorer Plugin for GFL Genesis Project
-"""
-from gfl.plugins.interfaces import AnalyzerPlugin
-from typing import Dict, Any, List
-import numpy as np
-
-class OffTargetScorerPlugin(AnalyzerPlugin):
-    """
-    Plugin to identify and score potential off-target sites using CFD approach
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.name = "offtarget_scorer"
-        self.version = "0.1.0"
-        self.description = "Identifies and scores CRISPR-Cas9 off-target sites"
-
-    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Identify and score potential off-target sites for gRNA sequences
-
-        Args:
-            data: Dictionary containing gRNA sequences
-
-        Returns:
-            Dictionary with off-target risk scores for each gRNA
-        """
-        # Extract gRNA sequences from input data
-        grna_sequences = data.get("sequences", [])
-
-        # In a real implementation, this would use BLAST/CFD
-        # For now, we'll generate mock scores
-        scores = []
-        for sequence in grna_sequences:
-            # Mock scoring algorithm - in reality, this would use BLAST + CFD
-            risk_score = self._mock_cfd_score(sequence)
-            scores.append({
-                "sequence": sequence,
-                "off_target_risk": risk_score
-            })
-
-        return {
-            "off_target_scores": scores,
-            "scoring_method": "CFD (mock implementation)",
-            "analysis_date": "2025-08-31"
-        }
-
-    def _mock_cfd_score(self, sequence: str) -> float:
-        """
-        Mock implementation of CFD scoring algorithm
-        In a real implementation, this would use BLAST to find off-target sites
-        and then apply CFD scoring to each site
-        """
-        # This is a placeholder - real implementation would use BLAST + CFD
-        # For demonstration, we'll return a random score between 0.0 and 0.5
-        # with a slight bias toward sequences with higher GC content
-        import random
-        return random.uniform(0.0, 0.5)
-
-    def validate_input(self, data: Dict[str, Any]) -> bool:
-        """
-        Validate input data for the plugin
-        """
-        return "sequences" in data and isinstance(data["sequences"], list)
-
-    def get_required_inputs(self) -> List[str]:
-        """
-        Get list of required input parameters
-        """
-        return ["sequences"]
-
-    def get_output_format(self) -> Dict[str, str]:
-        """
-        Get description of output format
-        """
-        return {
-            "off_target_scores": "List of dictionaries with sequence and risk score",
-            "scoring_method": "Name of the scoring method used",
-            "analysis_date": "Date of analysis"
-        }
