+++ roadmap_phase_3/gfl-genesis/plugins/gfl-plugin-ontarget-scorer/gfl_plugin_ontarget_scorer/plugin.py
@@ -1,80 +1,0 @@
-"""
-On-Target Scorer Plugin for GFL Genesis Project
-"""
-from gfl.plugins.interfaces import AnalyzerPlugin
-from typing import Dict, Any, List
-import numpy as np
-
-class OnTargetScorerPlugin(AnalyzerPlugin):
-    """
-    Plugin to predict on-target cutting efficiency using DeepHF model approach
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.name = "ontarget_scorer"
-        self.version = "0.1.0"
-        self.description = "Predicts CRISPR-Cas9 on-target cutting efficiency"
-
-    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Predict on-target cutting efficiency for gRNA sequences
-
-        Args:
-            data: Dictionary containing gRNA sequences and genomic context
-
-        Returns:
-            Dictionary with efficiency scores for each gRNA
-        """
-        # Extract gRNA sequences from input data
-        grna_sequences = data.get("sequences", [])
-
-        # In a real implementation, this would use the DeepHF model
-        # For now, we'll generate mock scores
-        scores = []
-        for sequence in grna_sequences:
-            # Mock scoring algorithm - in reality, this would use DeepHF
-            score = self._mock_deephf_score(sequence)
-            scores.append({
-                "sequence": sequence,
-                "efficiency_score": score
-            })
-
-        return {
-            "on_target_scores": scores,
-            "model_used": "DeepHF (mock implementation)",
-            "analysis_date": "2025-08-31"
-        }
-
-    def _mock_deephf_score(self, sequence: str) -> float:
-        """
-        Mock implementation of DeepHF scoring algorithm
-        In a real implementation, this would use the actual DeepHF model
-        """
-        # This is a placeholder - real implementation would use DeepHF
-        # For demonstration, we'll return a random score between 0.5 and 1.0
-        # with a slight bias toward longer sequences with GC content around 50%
-        import random
-        return random.uniform(0.5, 1.0)
-
-    def validate_input(self, data: Dict[str, Any]) -> bool:
-        """
-        Validate input data for the plugin
-        """
-        return "sequences" in data and isinstance(data["sequences"], list)
-
-    def get_required_inputs(self) -> List[str]:
-        """
-        Get list of required input parameters
-        """
-        return ["sequences"]
-
-    def get_output_format(self) -> Dict[str, str]:
-        """
-        Get description of output format
-        """
-        return {
-            "on_target_scores": "List of dictionaries with sequence and efficiency score",
-            "model_used": "Name of the model used for scoring",
-            "analysis_date": "Date of analysis"
-        }
