+++ roadmap_phase_3/gfl-genesis/gfl_genesis/__init__.py
@@ -1,5 +1,0 @@
-"""
-GFL Genesis Project Package
-"""
-__version__ = "0.1.0"
-__author__ = "GFL Genesis Team"

[…]
