### 2025-09-19 #8927 @manuelmenendezgonzalez
Hi Jesko,

Thank you for your thorough review and detailed feedback. I have
implemented comprehensive improvements addressing all your concerns. Here's
a summary of the updates:

1. Scientific Accuracy:

- CFD Implementation: The CFD computation now correctly implements the
method described in Doench et al. (2016), using Supplementary Table 19 for
position-specific mismatch penalties.
- DeepHF Reference: The reference for DeepHF has been corrected to Wang et
al. (2019) from Nature Communications.
- Implementation Transparency: Documentation now clearly explains the
relationship between our implementations and the original research.

2. Technical Issues:

- Hardcoded Dates: All hardcoded dates have been replaced with dynamic date
generation.
- Plugin Registration: A robust automatic plugin registration system has
been developed.
- Plugin Availability: `ProteinVAEGenerator` and `BayesianOptimization` are
now properly available and functional, resolving the workflow execution
issue.

3. Documentation:

- Language Translation: All Spanish documentation in
`docs/features/design_block.md`, `docs/features/optimize_block.md`,
`docs/features/with_priors_clause.md`, and
`examples/gfl-genesis/docs/project_plan.md` has been translated to English.
(Spanish is my native language, and I acknowledge that I use AI
to translate documentation).

4. Verification and Testing:

- Comprehensive Test Suite: New test scripts confirm all fixes are working
correctly, and all plugins load and function as expected.

Again, thank you for your meticulous review that helped me significantly
improve the repository and the paper. I am confident that the repository is
now ready for review and publication.

Best,
Manuel

On Thu, 18 Sept 2025 at 23:23, Jesko Wagner ***@***.***>
wrote:

> *jeskowagner* left a comment (openjournals/joss-reviews#8927)
> <https://github.com/openjournals/joss-reviews/issues/8927#issuecomment-3309767423>
>
> Hi Manuel,
>
> I remain confused by the new changes.
>
> To start with one arbitrary example, could you please explain how the CFD
> computation
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/bae88f31acbb490995c624cc112eff6be3efee60/examples/gfl-genesis/plugins/gfl-plugin-offtarget-scorer/gfl_plugin_offtarget_scorer/plugin.py#L50-L74>
> you added today follows one of the papers
> <https://doi.org/10.1038/nbt.3437> you reference in the comments of the
> corresponding section? It does not resemble my reading of that paper's
> method, which is based on results they detail in Supplementary Table 19.
>
> Equally, why is the "analysis date
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/bae88f31acbb490995c624cc112eff6be3efee60/examples/gfl-genesis/plugins/gfl-plugin-offtarget-scorer/gfl_plugin_offtarget_scorer/plugin.py#L48>"
> hardcoded to an end of August date?
>
> And what is the correct reference for DeepHF? You mention Li et al. (2022)
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/bae88f31acbb490995c624cc112eff6be3efee60/examples/gfl-genesis/plugins/gfl-plugin-ontarget-scorer/gfl_plugin_ontarget_scorer/plugin.py#L52>,
> but I could not find the corresponding paper. Is it perhaps Wang et al.
> (2019) <https://doi.org/10.1038/s41467-019-12281-8>? If so, I don't think
> this
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/bae88f31acbb490995c624cc112eff6be3efee60/examples/gfl-genesis/plugins/gfl-plugin-ontarget-scorer/gfl_plugin_ontarget_scorer/plugin.py#L49-L73>
> is a correct implementation, because that paper describes it as an online
> design tool powered by a deep learning method, whereas yours is just using
> GC content and guide length plus noise.
>
> Could you please clarify your reasoning behind these additions? I
> understand some of the changes were to remove previous placeholder
> implementations, but why did you choose the logic for these functions as
> you did, but naming them after already existing, significantly different
> methods?
>
> I have also tried running your example
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang?tab=readme-ov-file#your-first-gfl-workflow>,
> after an installation of GeneForgeLang with pip install -e .[full] as
> instructed. I get this error:
>
> AST keys: ['metadata', 'design', 'optimize']Collected entity registry: {}Validation: ✅ PassedTraceback (most recent call last):
>   File "/path/to/repo/workflow.py", line 49, in <module>
>     result = execute(ast)
>   File "/path/to/repo/gfl/api.py", line 500, in execute
>     raise ValueError(f"Plugin requirements not met: {error_msg}")ValueError: Plugin requirements not met: Design model 'ProteinVAEGenerator' not available. Available: ; Optimization strategy 'BayesianOptimization' not supported. Available: []
>
> Also, is there a reason some of the documentation
> <https://fundacion-de-neurociencias.github.io/GeneForgeLang/features/design_block/>
> is in Spanish?
>
> Best wishes,
>
> Jesko
>
