+++ paper.md
@@ -1,11 +1,12 @@
 ---
-title: 'GeneForgeLang: A Symbolic Language for Genetic Reasoning, Editing, and Simulation'
+title: 'GeneForgeLang: A Domain-Specific Language for Genomic Workflows with AI-Powered Analysis'
 tags:
   - synthetic biology
   - gene editing
-  - symbolic language
+  - domain-specific language
   - bioinformatics
   - AI-driven design
+  - CRISPR
 authors:
   - name: Manuel Menéndez González
     orcid: 0000-0002-5218-0774
@@ -14,29 +15,29 @@ affiliations:
  - index: 1
    name: Universidad de Oviedo, Spain
    ror: 006gksa02
-date: 24 May 2025
+date: 19 September 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-**GeneForgeLang (GFL)** is a symbolic language designed to represent, analyze, and simulate biomolecular processes with unprecedented clarity, granularity, and logical reasoning. GFL allows scientists and AI models to interact through a shared grammar that can describe genetic modifications, transcriptional regulation, protein folding, therapeutic edits, and pathway simulation.
+**GeneForgeLang (GFL)** is a domain-specific language designed to represent, analyze, and simulate genomic workflows with unprecedented clarity, granularity, and AI-powered analysis capabilities. GFL allows scientists and AI models to interact through a shared grammar that can describe genetic modifications, CRISPR guide RNA design, therapeutic edits, and pathway optimization.
 
-Originally conceived to address the limitations of existing descriptive formats (e.g., FASTA, SBOL, ProForma) and programming languages used in synthetic biology, GFL enables representation of **DNA, RNA, and protein constructs**, **mutation provenance**, **therapeutic delivery**, and **feedback systems**. Its formal grammar and parser support AI-enhanced interpretation, bidirectional translation from scientific literature, and automatic simulation of molecular logic.
+Originally conceived to address the limitations of existing descriptive formats and programming languages used in synthetic biology, GFL enables representation of **CRISPR gRNA designs**, **gene editing workflows**, **protein engineering processes**, and **experimental optimization strategies**. Its formal grammar and parser support AI-enhanced interpretation, bidirectional translation from scientific literature, and automatic simulation of molecular logic.
 
-Although it is currently used in the context of precision medicine and neurogenetics, GFL is fully extensible to domains like **crop engineering**, **antibiotic resistance**, **pandemic vaccine design**, and **synthetic metabolic pathways**.
+GFL is particularly well-suited for applications in **precision medicine**, **CRISPR optimization**, **protein design**, and **synthetic biology workflows**.
 
 # Statement of need
 
-With the rise of CRISPR, base editors, and programmable RNA therapies, biomedical science has entered an era of **complex multilayered interventions**. Yet, most researchers still describe their designs through prose, diagrams, or Python scripts—lacking a **semantic, symbolic, and machine-interpretable layer**.
+With the rise of CRISPR technologies, base editors, and programmable RNA therapies, biomedical science has entered an era of **complex multilayered interventions**. Yet, most researchers still describe their designs through prose, diagrams, or general-purpose programming languages—lacking a **semantic, symbolic, and machine-interpretable layer** specifically designed for genomic workflows.
 
[…]
