+++ gfl/grammar_parser.py
@@ -35,9 +35,9 @@ try:
         EnhancedValidationError,
         EnhancedValidationResult,
         ErrorCategory,
+        ErrorFix,
         ErrorSeverity,
         SourceLocation,
-        ErrorFix,
     )
 except ImportError:
     # Fallback for testing outside package context
@@ -45,9 +45,9 @@ except ImportError:
         EnhancedValidationError,
         EnhancedValidationResult,
         ErrorCategory,
+        ErrorFix,
         ErrorSeverity,
         SourceLocation,
-        ErrorFix,
     )
 
 # Import performance components with fallback
@@ -666,9 +666,7 @@ class AdvancedGFLParser:
         """Get source location for a parser token."""
         if hasattr(p, "lineno") and hasattr(p, "lexpos"):
             return SourceLocation(
-                line=p.lineno(index)
-                if callable(p.lineno)
-                else getattr(p.slice[index], "lineno", 1),
+                line=p.lineno(index) if callable(p.lineno) else getattr(p.slice[index], "lineno", 1),
                 column=getattr(p.slice[index], "lexpos", 0),
                 file_path=getattr(self, "current_file", "<input>"),
             )
@@ -706,10 +704,7 @@ class AdvancedGFLParser:
                     category=ErrorCategory.SYNTAX,
                 )
 
-                return EnhancedValidationResult(
-                    errors=[error],
-                    file_path=filename
-                )
+                return EnhancedValidationResult(errors=[error], file_path=filename)
 
 
 # Factory functions
