### 2025-09-19 ea7aed0
Final updates to address reviewer concerns - Added comprehensive documentation and verification summaries
