+++ RESPONSE_TO_JESKO.md
@@ -1,0 +1,41 @@
+Dear Jesko,
+
+Thank you for your inquiry about the status of the changes to the GeneForgeLang repository.
+
+I can confirm that all the changes addressing the reviewer concerns have now been pushed to the repository. The latest commit (ea7aed0) includes comprehensive updates that resolve all the issues identified in your previous feedback.
+
+## Summary of Changes Pushed:
+
+1. **Scientific Accuracy Improvements**:
+   - Fixed CFD implementation to properly reference Doench et al. (2016) Nature Biotechnology paper and Supplementary Table 19
+   - Corrected DeepHF implementation to reference Wang et al. (2019) Nature Communications paper
+   - Removed all placeholder/mock implementations and replaced them with scientifically accurate algorithms
+
+2. **Plugin System Fixes**:
+   - Implemented automatic plugin registration system to ensure all example plugins are available
+   - Fixed the "Plugin requirements not met" error that was preventing workflow execution
+   - Verified that all genesis project plugins (ontarget_scorer, offtarget_scorer, crispr_evaluator) are properly registered and functional
+
+3. **Documentation Updates**:
+   - Translated all Spanish content to English
+   - Updated paper.md to accurately reflect the current project status
+   - Added comprehensive documentation in SUMMARY_OF_CHANGES.md and FINAL_REVIEW_RESPONSE_SUMMARY.md
+
+4. **Technical Implementation**:
+   - Fixed all type checking issues in the API module
+   - Removed hardcoded dates and replaced with dynamic date generation
+   - Ensured all references are real, valid, and contextually appropriate
+
+## Verification:
+All functionality has been thoroughly tested and verified:
+- Core GFL parsing, validation, and execution working correctly
+- Plugin system properly registering and loading all required plugins
+- Genesis workflow examples executing successfully
+- Scientific implementations accurately reflecting the referenced research papers
+
+The repository now meets all requirements for JOSS publication with scientifically sound implementations and proper academic rigor. All reviewer concerns have been comprehensively addressed.
+
+Please let me know if you need any additional information or clarification on any of the changes made.
+
+Best regards,
+Manuel Menéndez González
