### 2025-09-19 03c3484
Corrected scientific implementations to properly reference and implement Doench et al. (2016) algorithms - Fixed Rule Set 2 and CFD implementations, removed all hallucinated content
