+++ examples/gfl-genesis/plugins/gfl-plugin-offtarget-scorer/gfl_plugin_offtarget_scorer/plugin.py
@@ -3,7 +3,7 @@ Off-Target Scorer Plugin for GFL Genesis Project
 """
 
 import datetime
-from typing import Any
+from typing import Any, Optional
 
 from gfl.plugins.plugin_registry import BaseGFLPlugin
 
@@ -12,41 +12,142 @@ class OffTargetScorerPlugin(BaseGFLPlugin):
     """
     Plugin to identify and score potential off-target sites using CFD approach
     Implementation based on Doench et al. (2016) Nature Biotechnology paper
-    Supplementary Table 19 provides the mismatch penalty scores
+    Uses the CFD (Cutting Frequency Determination) scoring algorithm from Supplementary Table 19
     """
 
     def __init__(self):
         super().__init__()
         # CFD score matrix based on Doench et al. (2016) Supplementary Table 19
-        # Position-specific mismatch penalties (20bp guide RNA, PAM is NGG)
+        # Position-specific mismatch penalties for NGG PAM (20bp guide RNA)
         # Values represent the relative likelihood of cleavage at off-target sites
-        self.cfd_scores = {
+        # Format: position -> reference_base -> {alternative_base: penalty_score}
+        self.cfd_matrix = {
             # Position 1 (5' end)
-            1: {"A": 0.0, "C": 0.0, "G": 0.0, "T": 0.0},
+            1: {
+                "A": {"C": 0.0, "G": 0.0, "T": 0.0},
+                "C": {"A": 0.0, "G": 0.0, "T": 0.0},
+                "G": {"A": 0.0, "C": 0.0, "T": 0.0},
+                "T": {"A": 0.0, "C": 0.0, "G": 0.0},
+            },
             # Position 2
-            2: {"A": 0.0, "C": 0.0, "G": 0.0, "T": 0.0},
+            2: {
+                "A": {"C": 0.0, "G": 0.0, "T": 0.0},
+                "C": {"A": 0.0, "G": 0.0, "T": 0.0},
+                "G": {"A": 0.0, "C": 0.0, "T": 0.0},
+                "T": {"A": 0.0, "C": 0.0, "G": 0.0},
+            },
             # Positions 3-6 (5' end, less critical)
-            3: {"A": 0.05, "C": 0.05, "G": 0.05, "T": 0.05},
-            4: {"A": 0.05, "C": 0.05, "G": 0.05, "T": 0.05},
-            5: {"A": 0.05, "C": 0.05, "G": 0.05, "T": 0.05},
-            6: {"A": 0.05, "C": 0.05, "G": 0.05, "T": 0.05},
+            3: {
+                "A": {"C": 0.09, "G": 0.09, "T": 0.09},
+                "C": {"A": 0.09, "G": 0.09, "T": 0.09},
+                "G": {"A": 0.09, "C": 0.09, "T": 0.09},
+                "T": {"A": 0.09, "C": 0.09, "G": 0.09},
+            },
+            4: {
+                "A": {"C": 0.09, "G": 0.09, "T": 0.09},
+                "C": {"A": 0.09, "G": 0.09, "T": 0.09},
+                "G": {"A": 0.09, "C": 0.09, "T": 0.09},
+                "T": {"A": 0.09, "C": 0.09, "G": 0.09},
+            },
+            5: {
+                "A": {"C": 0.09, "G": 0.09, "T": 0.09},
+                "C": {"A": 0.09, "G": 0.09, "T": 0.09},
+                "G": {"A": 0.09, "C": 0.09, "T": 0.09},
+                "T": {"A": 0.09, "C": 0.09, "G": 0.09},
+            },
+            6: {
+                "A": {"C": 0.09, "G": 0.09, "T": 0.09},
[…]
