### 2025-09-19 efc05d8
Add final response to Jesko with detailed explanation of corrections made
