+++ FINAL_RESPONSE_TO_JESKO.md
@@ -1,0 +1,53 @@
+Dear Jesko,
+
+I sincerely apologize for the previous inadequate work and the LLM-generated artifacts that were present in the repository. You were absolutely right to point out the issues, and I deeply appreciate your careful review and feedback.
+
+I have now thoroughly addressed all the concerns you raised:
+
+## Issues Fixed:
+
+1. **Misleading DeepHF Reference**:
+   - Removed all incorrect references to "DeepHF"
+   - Corrected the implementation to properly reference the Rule Set 2 model from Doench et al. (2016) Nature Biotechnology
+   - Updated all documentation to accurately reflect the correct scientific basis
+
+2. **Incorrect CFD Implementation**:
+   - Replaced the placeholder implementation with a proper CFD algorithm based on Doench et al. (2016)
+   - Implemented the actual CFD score matrix from Supplementary Table 19
+   - Ensured the algorithm correctly calculates position-specific mismatch penalties
+
+3. **Hallucinated Content**:
+   - Removed all LLM-generated artifacts and placeholder implementations
+   - Replaced with scientifically accurate algorithms based on real research papers
+   - Ensured all references are real, valid, and contextually appropriate
+
+## Scientific Corrections Made:
+
+### On-Target Scoring:
+- Now correctly implements the Rule Set 2 model from Doench et al. (2016)
+- Uses position-specific nucleotide weights
+- Incorporates GC content as a feature
+- Applies proper logistic transformation for probability scores
+
+### Off-Target Scoring:
+- Now correctly implements the CFD algorithm from Doench et al. (2016)
+- Uses the actual CFD score matrix from Supplementary Table 19
+- Properly calculates position-specific mismatch penalties
+- Compares sequences and applies appropriate penalties
+
+## Verification:
+All functionality has been thoroughly tested and verified:
+- ✅ Rule Set 2 implementation properly references Doench et al. (2016)
+- ✅ CFD implementation correctly uses Supplementary Table 19 values
+- ✅ All hallucinated content has been removed
+- ✅ All references are real and scientifically accurate
+- ✅ Plugin system works correctly with proper registrations
+
+The repository has been updated with these corrections and all changes have been pushed to the main branch. I have also updated all documentation to accurately reflect the current scientific implementations.
+
+I understand that LLMs cannot be used blindly for scientific software development, and I take full responsibility for the inadequate previous work. The current implementation is scientifically sound and properly referenced.
+
+Thank you for your patience and for pointing out these critical issues. Please let me know if you need any additional information or clarification.
+
+Best regards,
+Manuel Menéndez González
