### 2025-09-19 #8927 @manuelmenendezgonzalez
Hi Jesko,
I really appreciate you pointing all this out, and I indeed agree with you.
That's why I have tried to fix those points. However, I can´t reproduce
some of the latest issues you mention on my local repo after the latest
corrections I made, so I suspect there might be a synchronization issue,
perhaps with my gitignore status. I have just tried to push the changes
again, and I think it is ok now.
The development of this repo has passed through different phases, and a
global restructuring and reorganization were much needed. I have updated
some directories and archived some others. I think the repo and the
software are more usable now. I hope this can ease the review process, too.
Best,
Manuel

On Fri, 19 Sept 2025 at 09:35, Jesko Wagner ***@***.***>
wrote:

> *jeskowagner* left a comment (openjournals/joss-reviews#8927)
> <https://github.com/openjournals/joss-reviews/issues/8927#issuecomment-3311000387>
>
> Hi Manuel,
>
> It is clear to me that your software is the product of repeatedly
> prompting an LLM with your research questions and providing guidance with
> Spanish comments. As a biologist myself, I see the allure of using LLMs to
> accelerate research and particularly software development. However, no tool
> exists as of yet that could really perform the task of going from research
> question to fully fledged, robust software. I am not sure which model or
> suite you are using, but please take my word for it when I say it is making
> things up. I will give you a few examples from the issues we have been
> discussing.
>
> Your reference
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blame/f8c98c5434f1ac0a65d54bcbd46e0baea33b596f/FINAL_REVIEW_RESPONSE.md#L47>
> to DeepHF remains incorrect, the title is hallucinated.
>
> The implementation itself states
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/main/examples/gfl-genesis/plugins/gfl-plugin-ontarget-scorer/gfl_plugin_ontarget_scorer/plugin.py#L66-L69>
> that you are not, in fact, using DeepHF or an approximation thereof,
> despite your response
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blame/ea7aed07546d6172869d9be4af0c4486e3f01023/FINAL_REVIEW_RESPONSE_SUMMARY.md#L21-L27>
> to my "review". NB I have not formally started my review, which I aimed to
> do after the question of LLM use for the creation of this software was
> clarified.
>
> The same things are true for the CFD score, where the LLM even tells you
> <https://github.com/Fundacion-de-Neurociencias/GeneForgeLang/blob/main/examples/gfl-genesis/plugins/gfl-plugin-offtarget-scorer/gfl_plugin_offtarget_scorer/plugin.py#L105-L114>
> that the implementation is not actually doing CFD. For example,
> dinucleotide sequences are completely ignored, presumably because the LLM
> did not have access to the relevant data from the paper.
>
> What these examples show is that, while LLMs can speed some things up,
> they can unfortunately not be used blindly.
> ------------------------------
>
> Dear editors,
>
> As detailed in my email to Abishek and underpinned with more examples in
> the comments in this thread, I have concerns about this software.
> GeneForgeLang spans many thousands of lines of code, and it is unclear how
> many of them are affected by these sorts of issues. I believe a very
> in-depth review would be required to flag all remaining issues, putting a
> huge burden on reviewers to criticially examine the logic, references, and
> implementation of 20.000+ lines of code. In my eyes, this seems infeasible
> as of now.
>
> Best wishes,
>
> Jesko
>
