### 2025-09-19 99c03f5
🚀 Complete GeneForgeLang Professionalization
✅ JOSS Issues Resolved:
- Removed all Spanish language content (files, comments, docs)
- Eliminated incorrect scientific implementations (CFD/DeepHF claims)
- Fixed plugin availability issues with working plugin system
- Removed hardcoded values and AI-generated artifacts
- Created scientifically honest, transparent implementations

🏗️ Professional Project Structure:
- Implemented modern src-layout (src/geneforgelang/)
- Organized modular architecture (core, plugins, web, cli, utils)
- Updated to modern pyproject.toml configuration
- Created profess
[…]
