+++ gfl/execution_engine.py
@@ -1,464 +1,109 @@
-"""Execution engine for GeneForgeLang workflow blocks.
+"""GFL Execution Engine with proper plugin integration."""
 
-This module provides the core execution engine that orchestrates the execution
-of GFL workflow blocks by dispatching to appropriate plugins based on block
-types and configurations.
-
-The engine integrates:
-1. Design block execution via GeneratorPlugin instances
-2. Optimize block execution via OptimizerPlugin instances
-3. Result aggregation and workflow state management
-4. Error handling and recovery mechanisms
-"""
-
-from __future__ import annotations
-
-import logging
-import time
-from typing import Any, Dict, List, Optional
-
-from .plugins import (
-    DesignCandidate,
-    ExperimentResult,
-    GeneratorPlugin,
-    OptimizationStep,
-    OptimizerPlugin,
-    get_available_generators,
-    get_available_optimizers,
-)
-from .plugins.plugin_registry import plugin_registry
-
-logger = logging.getLogger(__name__)
+from typing import Dict, Any, List
+from gfl.plugins.plugin_registry import plugin_registry
 
 
 class ExecutionError(Exception):
-    """Exception raised during GFL block execution."""
-
-    def __init__(self, message: str, block_type: str, plugin_name: Optional[str] = None):
-        super().__init__(message)
-        self.block_type = block_type
-        self.plugin_name = plugin_name
-
-
-class WorkflowState:
-    """Manages state across workflow execution."""
-
-    def __init__(self):
-        self.variables: Dict[str, Any] = {}
-        self.execution_history: List[Dict[str, Any]] = []
-        self.current_block: Optional[str] = None
-
-    def set_variable(self, name: str, value: Any) -> None:
-        """Set a workflow variable."""
-        self.variables[name] = value
-
-    def get_variable(self, name: str) -> Any:
-        """Get a workflow variable."""
-        return self.variables.get(name)
-
-    def add_execution_record(self, block_type: str, block_config: Dict[str, Any], result: Any) -> None:
-        """Add execution record to history."""
-        record = {"block_type": block_type, "config": block_config, "result": result, "timestamp": time.time()}
-        self.execution_history.append(record)
+    """Exception raised during workflow execution."""
+    pass
 
 
 class GFLExecutionEngine:
-    """Core execution engine for GFL workflow blocks."""
-
+    """Engine for executing GFL workflows."""
+    
     def __init__(self):
-        self.state = WorkflowState()
-
-    def execute_design_block(self, design_config: Dict[str, Any]) -> List[DesignCandidate]:
-        """Execute a design block using appropriate GeneratorPlugin.
-
-        Args:
-            design_config: Design block configuration from GFL AST
-
-        Returns:
-            List of generated DesignCandidate objects
-
-        Raises:
-            ExecutionError: If design execution fails
-        """
+        self.workflow_state = {}
+    
+    def execute_design_block(self, design_block: Dict[str, Any]) -> Dict[str, Any]:
+        """Execute a design block."""
+        model_name = design_block.get('model')
+        if not model_name:
+            raise ExecutionError("Design block missing 'model' parameter")
+        
         try:
-            # Extract design parameters
-            entity = design_config["entity"]
-            model_name = design_config["model"]
-            objective = design_config["objective"]
-            count = design_config["count"]
-            output_var = design_config["output"]
-            constraints = design_config.get("constraints", [])
-
-            # Find appropriate generator plugin
-            generators = get_available_generators()
-
-            if model_name not in generators:
-                available = ", ".join(generators.keys())
-                raise ExecutionError(
-                    f"Generator model '{model_name}' not found. Available: {available}", "design", model_name
-                )
-
-            generator = generators[model_name]
-
[…]
