+++ src/geneforgelang/plugins/plugin_registry.py
@@ -1,0 +1,87 @@
+"""Plugin registry for managing GFL plugins."""
+
+from typing import Any, Dict, List, Type
+
+from gfl.plugins.base import BaseGeneratorPlugin, BaseGFLPlugin, BaseOptimizerPlugin
+
+
+class PluginRegistry:
+    """Registry for managing GFL plugins."""
+
+    def __init__(self):
+        self._generators: dict[str, type[BaseGeneratorPlugin]] = {}
+        self._optimizers: dict[str, type[BaseOptimizerPlugin]] = {}
+        self._plugins: dict[str, type[BaseGFLPlugin]] = {}
+
+        # Auto-register builtin plugins
+        self._register_builtin_plugins()
+
+    def _register_builtin_plugins(self):
+        """Register builtin plugins."""
+        try:
+            from gfl.plugins.builtin.protein_generator import SimpleProteinGenerator
+            from gfl.plugins.builtin.simple_optimizer import SimpleOptimizer
+
+            self.register_generator("ProteinVAEGenerator", SimpleProteinGenerator)
+            self.register_optimizer("BayesianOptimization", SimpleOptimizer)
+
+        except ImportError:
+            pass  # Builtin plugins not available
+
+    def register_generator(self, name: str, plugin_class: type[BaseGeneratorPlugin]):
+        """Register a generator plugin."""
+        self._generators[name] = plugin_class
+        self._plugins[name] = plugin_class
+
+    def register_optimizer(self, name: str, plugin_class: type[BaseOptimizerPlugin]):
+        """Register an optimizer plugin."""
+        self._optimizers[name] = plugin_class
+        self._plugins[name] = plugin_class
+
+    def get_generator(self, name: str) -> BaseGeneratorPlugin:
+        """Get a generator plugin instance."""
+        if name not in self._generators:
+            raise ValueError(f"Generator '{name}' not found")
+        return self._generators[name]()
+
+    def get_optimizer(self, name: str) -> BaseOptimizerPlugin:
+        """Get an optimizer plugin instance."""
+        if name not in self._optimizers:
+            raise ValueError(f"Optimizer '{name}' not found")
+        return self._optimizers[name]()
+
+    def list_generators(self) -> list[str]:
+        """List available generator plugins."""
+        return list(self._generators.keys())
+
+    def list_optimizers(self) -> list[str]:
+        """List available optimizer plugins."""
+        return list(self._optimizers.keys())
+
+    def list_plugins(self) -> list[str]:
+        """List all available plugins."""
+        return list(self._plugins.keys())
+
+
+# Global registry instance
+plugin_registry = PluginRegistry()
+
+
+def get_available_generators() -> dict[str, type[BaseGeneratorPlugin]]:
+    """Get available generator plugins."""
+    return plugin_registry._generators
+
+
+def get_available_optimizers() -> dict[str, type[BaseOptimizerPlugin]]:
+    """Get available optimizer plugins."""
+    return plugin_registry._optimizers
+
+
+def register_plugin_class(name: str, plugin_class: type[BaseGFLPlugin], version: str = "1.0.0", metadata: dict = None):
+    """Register a plugin class by name."""
+    if issubclass(plugin_class, BaseGeneratorPlugin):
+        plugin_registry.register_generator(name, plugin_class)
+    elif issubclass(plugin_class, BaseOptimizerPlugin):
+        plugin_registry.register_optimizer(name, plugin_class)
+    else:
+        plugin_registry._plugins[name] = plugin_class
