+++ examples/gfl-genesis/plugins/gfl-crispr-evaluator/gfl_crispr_evaluator/__init__.py
@@ -1,3 +1,0 @@
-"""
-GFL Plugin for CRISPR Evaluation Orchestration
-"""
