### 2025-09-19 2108b7c
Fix formatting and linting issues from pre-commit hooks
- Applied automatic formatting fixes from Black and Ruff
- Fixed line endings and mixed line ending issues
- Updated typing imports to use modern syntax
- Resolved code style inconsistencies
