### 2025-09-20 98621a5
Remove JOSS review response documents
