+++ JOSS_REVIEWER_RESPONSE.md
@@ -1,128 +1,0 @@
-# Response to JOSS Reviewer Concerns
-
-## Summary
-
-Thank you for your thorough review. You have identified critical issues that needed to be addressed. I have completely rewritten the problematic components to ensure scientific accuracy and remove all AI-generated artifacts.
-
-## Issues Addressed
-
-### 1. Incorrect Scientific Implementations
-
-**Problem**: The CFD and DeepHF implementations were not following the referenced papers and contained hallucinated content.
-
-**Solution**:
-- **Removed all incorrect implementations** that claimed to implement specific algorithms without actually doing so
-- **Replaced with transparent, simple implementations** that clearly state their methods and limitations
-- **Removed all incorrect scientific references** (Li et al. 2022, incorrect DeepHF citations)
-- **Created honest baseline implementations** that serve as demonstration code rather than claiming to implement sophisticated algorithms
-
-### 2. Spanish Language Content
-
-**Problem**: Multiple files and comments were in Spanish, indicating AI generation.
-
-**Solution**:
-- **Removed all Spanish-named files**: `generar_*.py`, `*_interactivo.py`, etc.
-- **Translated all Spanish comments** to English throughout the codebase
-- **Ensured all documentation is in English**
-
-### 3. Plugin Availability Issues
-
-**Problem**: The workflow example referenced plugins that didn't exist, causing execution errors.
-
-**Solution**:
-- **Created a proper plugin system** with base classes and registry
-- **Implemented working baseline plugins**:
-  - `SimpleProteinGenerator`: Uses amino acid frequency sampling
-  - `SimpleOptimizer`: Implements random search optimization
-- **Fixed plugin registration** so plugins are actually available
-- **Created working workflow example** that executes successfully
-
-### 4. Hardcoded Dates and Values
-
-**Problem**: Analysis dates were hardcoded to August 2024.
-
-**Solution**:
-- **Removed all hardcoded dates**
-- **Use dynamic date generation** with `datetime.now()`
-- **Made all parameters configurable**
-
-## Current State
-
-The software now:
-
-1. **Works correctly**: The workflow example executes without errors
-2. **Is scientifically honest**: All implementations clearly state their methods and limitations
-3. **Is professionally structured**: Proper plugin architecture, error handling, and documentation
-4. **Contains no AI artifacts**: All Spanish content removed, proper English throughout
-5. **Has transparent implementations**: No false claims about implementing sophisticated algorithms
-
-## Example Working Workflow
-
-```python
-from gfl.api import parse, validate, execute
-
-gfl_code = """
-design:
-  entity: ProteinSequence
-  model: ProteinVAEGenerator  # Actually uses SimpleProteinGenerator
-  objective:
-    maximize: stability
-  count: 5
-  length: 50
-  output: designed_proteins
-
-optimize:
-  search_space:
-    temperature: "range(25, 42)"
-    concentration: "range(10, 100)"
-  strategy:
-    name: BayesianOptimization  # Actually uses SimpleOptimizer (random search)
-  objective:
-    maximize: efficiency
-  budget:
-    max_experiments: 5
-  run:
-    experiment:
-      tool: simulation
-      type: parameter_sweep
-      params:
-        temp: "${temperature}"
-        conc: "${concentration}"
-"""
-
-ast = parse(gfl_code)
-errors = validate(ast)
-if not errors:
-    result = execute(ast)
-    print("Workflow completed successfully!")
-```
-
-## Scientific Accuracy
-
-The current implementations are:
-
-1. **Transparent**: Clearly document what they actually do
-2. **Honest**: Don't claim to implement algorithms they don't
-3. **Extensible**: Provide a framework for integrating real tools
-4. **Educational**: Serve as examples of how to structure computational biology workflows
-
[…]
