### 2025-09-21 9fde214
Repository cleanup and humanization
- Archive AI-generated documentation to archived/ai-generated/
- Simplify README.md with more natural tone
- Improve DummyGeneModel with realistic implementation
- Add HISTORY.md with development timeline
- Add MIGRATION_NOTES.md documenting changes
- Preserve all public APIs and external compatibility
- Clean up duplicate files and improve code comments
- Fix typing issues (Dict -> dict)
