+++ HISTORY.md
@@ -1,0 +1,55 @@
+# GeneForgeLang Development History
+
+## Early Development (2023-2024)
+
+### Initial Concept
+Started as a side project to make genomic workflows more accessible. The idea came from working with complex bioinformatics pipelines that were hard to reproduce and share.
+
+### First Implementation
+- Basic YAML parser for genomic experiments
+- Simple validation system
+- Focus on CRISPR and gene editing workflows
+
+### Key Milestones
+
+**v0.1.0 (March 2024)**
+- First working parser
+- Basic experiment validation
+- Simple CLI interface
+
+**v0.2.0 (June 2024)**
+- Added plugin system
+- Web interface prototype
+- Better error handling
+
+**v0.3.0 (September 2024)**
+- Major refactor for better performance
+- Enhanced validation with type checking
+- Improved documentation
+
+**v1.0.0 (Current)**
+- Stable API for external integrations
+- Comprehensive plugin ecosystem
+- Production-ready features
+
+## Development Notes
+
+### Design Decisions
+- Chose YAML over JSON for better readability
+- Plugin system inspired by existing bioinformatics tools
+- Web interface built with FastAPI for performance
+
+### Challenges Overcome
+- Balancing simplicity with power
+- Making the language extensible without being complex
+- Ensuring backward compatibility during major refactors
+
+### Community Contributions
+- Several plugins contributed by early users
+- Feedback helped shape the API design
+- Bug reports led to better error messages
+
+## Future Plans
+- Better integration with existing bioinformatics tools
+- More sophisticated validation rules
+- Performance optimizations for large workflows

[…]
