### 2025-09-21 b4f3f48
Remove non-functional PyPI badges from README
- Remove PyPI version badge (package not published yet)
- Remove Python Support badge (package not published yet)
- Keep only MIT License badge which works correctly
