### 2025-09-22 89dae59
Integrate DataStagingManager from GeneForge team
- Add gfl/staging module with DataStagingManager class
- Integrate data staging into GFL service /api/v2/execute endpoint
- Add data_manifest parameter to ExecuteRequest
- Implement automatic file staging and cleanup
- Add comprehensive test suite for data staging functionality
- Add high-level integration tests for complete workflow
- Update pyproject.toml with requests dependency
- Add documentation for data staging feature
- Preserve all existing APIs and maintain backward compatibility

This enables GFL workflows to work with remote data fil
[…]
