### 2025-09-23 35809df
feat(ecosystem): Complete LSP, cookiecutter, and plugin packaging
