+++ cookiecutter-gfl-plugin/{{ cookiecutter.plugin_name }}/{{ cookiecutter.module_name }}/plugin.py
@@ -1,0 +1,126 @@
+"""Plugin implementation for {{ cookiecutter.plugin_name }}."""
+
+from typing import Any, Dict, List, Optional
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+class {{ cookiecutter.module_name.split('_')[-1].title() }}Plugin:
+    """
+    {{ cookiecutter.description }}
+    
+    This plugin provides functionality for GeneForgeLang workflows.
+    """
+    
+    def __init__(self, config: Optional[Dict[str, Any]] = None):
+        """
+        Initialize the plugin.
+        
+        Args:
+            config: Optional configuration dictionary for the plugin.
+        """
+        self.config = config or {}
+        self.name = "{{ cookiecutter.plugin_name }}"
+        self.version = "{{ cookiecutter.version }}"
+        
+        logger.info(f"Plugin '{self.name}' v{self.version} initialized.")
+    
+    def validate_input(self, input_data: Any) -> bool:
+        """
+        Validate input data for the plugin.
+        
+        Args:
+            input_data: Input data to validate.
+            
+        Returns:
+            True if input is valid, False otherwise.
+        """
+        # Add your validation logic here
+        if input_data is None:
+            logger.error("Input data is None")
+            return False
+        
+        return True
+    
+    def run(self, input_data: Any, params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
+        """
+        Execute the main plugin functionality.
+        
+        Args:
+            input_data: Input data to process.
+            params: Optional parameters for the plugin execution.
+            
+        Returns:
+            Dictionary containing the results of the plugin execution.
+        """
+        logger.info(f"Running plugin '{self.name}'...")
+        
+        # Validate input
+        if not self.validate_input(input_data):
+            return {
+                "status": "error",
+                "message": "Invalid input data",
+                "plugin": self.name
+            }
+        
+        try:
+            # --- Your plugin logic here ---
+            result = self._process_data(input_data, params or {})
+            
+            logger.info(f"Plugin '{self.name}' completed successfully.")
+            return {
+                "status": "success",
+                "result": result,
+                "plugin": self.name,
+                "version": self.version
+            }
+            
+        except Exception as e:
+            logger.error(f"Plugin '{self.name}' failed: {str(e)}")
+            return {
+                "status": "error",
+                "message": str(e),
+                "plugin": self.name
+            }
+    
+    def _process_data(self, input_data: Any, params: Dict[str, Any]) -> Any:
+        """
+        Process the input data according to plugin logic.
+        
+        Args:
+            input_data: Input data to process.
+            params: Parameters for processing.
+            
+        Returns:
+            Processed data.
+        """
+        # Example processing logic:
+        # if isinstance(input_data, list):
+        #     for record in input_data:
+        #         logger.debug(f"Processing record: {record.get('id', 'unknown')}")
+        #         # Process each record...
+        
+        # Placeholder processing
+        processed_data = {
+            "input_type": type(input_data).__name__,
+            "processed_at": "2024-01-01T00:00:00Z",
+            "params_used": params
+        }
+        
+        return processed_data
+    
+    def get_metadata(self) -> Dict[str, Any]:
+        """
+        Get plugin metadata.
+        
+        Returns:
+            Dictionary containing plugin metadata.
+        """
+        return {
+            "name": self.name,
+            "version": self.version,
+            "description": "{{ cookiecutter.description }}",
+            "author": "{{ cookiecutter.author_name }}",
+            "email": "{{ cookiecutter.author_email }}"
+        }
