+++ cookiecutter-gfl-plugin/{{ cookiecutter.plugin_name }}/{{ cookiecutter.module_name }}/__init__.py
@@ -1,0 +1,6 @@
+"""{{ cookiecutter.description }}"""
+
+from .plugin import {{ cookiecutter.module_name.split('_')[-1].title() }}Plugin
+
+__version__ = "{{ cookiecutter.version }}"
+__all__ = ["{{ cookiecutter.module_name.split('_')[-1].title() }}Plugin"]
