+++ gfl_lsp/gfl_lsp/__init__.py
@@ -1,0 +1,3 @@
+"""GeneForgeLang Language Server Protocol implementation."""
+
+__version__ = "0.1.0"
