+++ gfl_lsp/gfl_lsp/server/main.py
@@ -1,0 +1,443 @@
+import asyncio
+from typing import List
+from lsprotocol.types import (
+    Diagnostic, Position, Range, CompletionItem, CompletionItemKind, 
+    CompletionList, CompletionOptions, DiagnosticSeverity, Hover, 
+    MarkupContent, MarkupKind
+)
+from pygls.server import LanguageServer
+
+# Importamos nuestras propias herramientas de GFL directamente desde el proyecto
+import sys
+import os
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', '..'))
+
+from gfl.api import parse, validate
+from gfl.error_handling import EnhancedValidationResult, ErrorSeverity
+
+# Creamos una instancia del servidor de lenguaje
+server = LanguageServer("gfl-lsp", "v0.1.0")
+
+def _validate(ls: LanguageServer, params):
+    """
+    Función interna para validar un documento y publicar los diagnósticos.
+    """
+    text_doc = ls.workspace.get_document(params.text_document.uri)
+    source = text_doc.source
+    diagnostics: List[Diagnostic] = []
+
+    try:
+        # 1. Parsear el código fuente
+        ast = parse(source)
+        
+        # 2. Validar el AST con resultado mejorado
+        validation_result = validate(ast, enhanced=True)
+        
+        # 3. Convertir los errores de GFL a Diagnósticos de LSP
+        if isinstance(validation_result, EnhancedValidationResult):
+            # Procesar errores sintácticos
+            for error in validation_result.syntax_errors:
+                line = error.location.line - 1 if error.location and error.location.line > 0 else 0
+                col = error.location.column - 1 if error.location and error.location.column > 0 else 0
+                
+                d = Diagnostic(
+                    range=Range(
+                        start=Position(line=line, character=col),
+                        end=Position(line=line, character=col + len(error.context or ''))
+                    ),
+                    message=f"[{error.code}] {error.message}",
+                    severity=DiagnosticSeverity.Error
+                )
+                diagnostics.append(d)
+            
+            # Procesar errores semánticos
+            for error in validation_result.semantic_errors:
+                line = error.location.line - 1 if error.location and error.location.line > 0 else 0
+                col = error.location.column - 1 if error.location and error.location.column > 0 else 0
+                
+                # Determinar severidad
+                severity = DiagnosticSeverity.Error  # Error por defecto
+                if error.severity == ErrorSeverity.WARNING:
+                    severity = DiagnosticSeverity.Warning
+                elif error.severity == ErrorSeverity.INFO:
+                    severity = DiagnosticSeverity.Information
+                elif error.severity == ErrorSeverity.HINT:
+                    severity = DiagnosticSeverity.Hint
+                
+                d = Diagnostic(
+                    range=Range(
+                        start=Position(line=line, character=col),
+                        end=Position(line=line, character=col + len(error.context or ''))
+                    ),
+                    message=f"[{error.code}] {error.message}",
+                    severity=severity
+                )
+                diagnostics.append(d)
+        else:
+            # Resultado legacy (lista de strings)
+            for i, error_msg in enumerate(validation_result):
+                d = Diagnostic(
+                    range=Range(
+                        start=Position(line=i, character=0),
+                        end=Position(line=i, character=50)
+                    ),
+                    message=f"[VALIDATION] {error_msg}",
+                    severity=DiagnosticSeverity.Error
+                )
+                diagnostics.append(d)
+
+    except Exception as e:
+        # Manejar errores generales del parser
+        d = Diagnostic(
+            range=Range(
+                start=Position(line=0, character=0),
+                end=Position(line=0, character=10)
+            ),
+            message=f"[PARSE_ERROR] {str(e)}",
[…]
