+++ gfl/staging/__init__.py
@@ -7,3 +7,4 @@ to local temporary directories before plugin execution.
 from .manager import DataStagingManager
 
 __all__ = ["DataStagingManager"]
+
