### 2025-09-23 fe6eb63
feat(ecosystem): Complete LSP, cookiecutter, and plugin packaging
