+++ gfl_lsp/gfl_lsp/server/main.py
@@ -1,8 +1,8 @@
 import asyncio
 from typing import List
 from lsprotocol.types import (
-    Diagnostic, Position, Range, CompletionItem, CompletionItemKind, 
-    CompletionList, CompletionOptions, DiagnosticSeverity, Hover, 
+    Diagnostic, Position, Range, CompletionItem, CompletionItemKind,
+    CompletionList, CompletionOptions, DiagnosticSeverity, Hover,
     MarkupContent, MarkupKind
 )
 from pygls.server import LanguageServer
@@ -29,17 +29,17 @@ def _validate(ls: LanguageServer, params):
     try:
         # 1. Parsear el código fuente
         ast = parse(source)
-        
+
         # 2. Validar el AST con resultado mejorado
         validation_result = validate(ast, enhanced=True)
-        
+
         # 3. Convertir los errores de GFL a Diagnósticos de LSP
         if isinstance(validation_result, EnhancedValidationResult):
             # Procesar errores sintácticos
             for error in validation_result.syntax_errors:
                 line = error.location.line - 1 if error.location and error.location.line > 0 else 0
                 col = error.location.column - 1 if error.location and error.location.column > 0 else 0
-                
+
                 d = Diagnostic(
                     range=Range(
                         start=Position(line=line, character=col),
@@ -49,12 +49,12 @@ def _validate(ls: LanguageServer, params):
                     severity=DiagnosticSeverity.Error
                 )
                 diagnostics.append(d)
-            
+
             # Procesar errores semánticos
             for error in validation_result.semantic_errors:
                 line = error.location.line - 1 if error.location and error.location.line > 0 else 0
                 col = error.location.column - 1 if error.location and error.location.column > 0 else 0
-                
+
                 # Determinar severidad
                 severity = DiagnosticSeverity.Error  # Error por defecto
                 if error.severity == ErrorSeverity.WARNING:
@@ -63,7 +63,7 @@ def _validate(ls: LanguageServer, params):
                     severity = DiagnosticSeverity.Information
                 elif error.severity == ErrorSeverity.HINT:
                     severity = DiagnosticSeverity.Hint
-                
+
                 d = Diagnostic(
                     range=Range(
                         start=Position(line=line, character=col),
@@ -97,7 +97,7 @@ def _validate(ls: LanguageServer, params):
                 severity=DiagnosticSeverity.Error
         )
         diagnostics.append(d)
-    
+
     # 4. Publicar los diagnósticos en el editor
     ls.publish_diagnostics(text_doc.uri, diagnostics)
 
@@ -113,28 +113,28 @@ def completions(params):
     items = []
     document = server.workspace.get_document(params.text_document.uri)
     current_line = document.lines[params.position.line].strip()
-    
+
     try:
         # Parsear el documento completo para obtener el AST
         source = document.source
         ast = parse(source)
-        
+
         # A. Autocompletado de Bloques de Alto Nivel
         # Si la línea actual está vacía o tiene indentación cero
         if not document.lines[params.position.line].startswith('  '):
             keywords = [
-                'import_schemas', 'rules', 'hypothesis', 'timeline', 
-                'pathways', 'complexes', 'experiment', 'analyze', 
+                'import_schemas', 'rules', 'hypothesis', 'timeline',
+                'pathways', 'complexes', 'experiment', 'analyze',
                 'guided_discovery', 'refine_data', 'optimize', 'simulate'
             ]
             for keyword in keywords:
                 items.append(CompletionItem(
-                    label=keyword, 
+                    label=keyword,
                     kind=CompletionItemKind.KEYWORD,
                     detail=f"GFL block: {keyword}",
[…]
