+++ docs/features/data_staging.md
@@ -164,4 +164,3 @@ Potential future improvements:
 - **Progress Tracking**: Real-time progress updates for large file downloads
 - **Retry Logic**: Automatic retry for failed downloads
 - **Compression**: Support for compressed file formats
-
