### 2025-09-23 69b3a8a
Proyecto Fénix: Documentation consolidation - Reorganized documentation structure and updated mkdocs configuration
