+++ docs/getting_started/quickstart_user.md
@@ -1,0 +1,221 @@
+# User Quickstart Guide
+
+This guide will help you get started with GeneForgeLang by creating your first workflow. We'll walk through installation, creating a simple workflow, and running it with plugins.
+
+## Prerequisites
+
+Before starting, make sure you have:
+1. Python 3.9 or higher installed
+2. GeneForgeLang installed (see [Installation Guide](installation.md))
+3. Basic understanding of biological workflows
+
+## Step 1: Verify Installation
+
+First, let's verify that GeneForgeLang is properly installed:
+
+```bash
+# Check GeneForgeLang installation
+python -c "import geneforgelang; print(f'GeneForgeLang version: {geneforgelang.__version__}')"
+
+# Check available plugins
+gfl --list-plugins
+```
+
+## Step 2: Create Your First Workflow
+
+Create a new file called `my_first_workflow.gfl` with the following content:
+
+```gfl
+# My First GeneForgeLang Workflow
+# This workflow demonstrates basic GFL functionality
+
+# Define input parameters
+input:
+  sequence: "ATGCGATCGATCGATCGATCGATCGATCGATCGATCGATCG"
+  database: "nt"
+
+# Run a BLAST search
+run:
+  - plugin: "blast"
+    operation: "blastn"
+    sequence: "${sequence}"
+    database: "${database}"
+    expect_threshold: 0.001
+    as_var: "blast_results"
+
+# Process the results
+process:
+  - name: "filter_hits"
+    input: "${blast_results}"
+    operation: "filter"
+    condition: "evalue < 0.01"
+    as_var: "filtered_hits"
+
+# Output the results
+output:
+  - blast_hits: "${filtered_hits}"
+  - summary: "Found ${len(filtered_hits)} significant hits"
+```
+
+## Step 3: Run Your Workflow
+
+Execute your workflow using the GFL command-line tool:
+
+```bash
+gfl run my_first_workflow.gfl
+```
+
+## Step 4: View Results
+
+The results will be displayed in the terminal and saved to output files. You can also specify output formats:
+
+```bash
+# Save results in JSON format
+gfl run my_first_workflow.gfl --output-format json --output-file results.json
+
+# Save results in YAML format
+gfl run my_first_workflow.gfl --output-format yaml --output-file results.yaml
+```
+
+## Step 5: Create a CRISPR Design Workflow
+
+Let's create a more advanced workflow using the Genesis plugins for CRISPR design:
+
+```gfl
+# CRISPR Design Workflow
+# This workflow designs and evaluates gRNA candidates
+
+# Define input sequences
+input:
+  target_sequence: "ATGCGATCGATCGATCGATCGATCGATCGATCGATCGATCG"
+  genome_context: "ACGTGCAATGGAGCGGCTTGCGGATCGATCGATCGATCGATCGATCG"
+
+# Generate gRNA candidates
+design:
+  entity: "gRNA"
+  count: 5
+  constraints:
+    - "length(20)"
+    - "gc_content(40, 60)"
+
+# Evaluate candidates
+evaluate:
+  candidates: "${design.candidates}"
+  
+  # Score on-target efficiency
+  - plugin: "gfl-ontarget-scorer"
+    input: 
+      grna_sequence: "${candidate.sequence}"
+      genome_sequence: "${genome_context}"
+    as_var: "on_target_score"
+  
+  # Score off-target risk
+  - plugin: "gfl-offtarget-scorer"
+    input:
+      grna_sequence: "${candidate.sequence}"
+    params:
+      max_mismatches: 3
+      genome_reference: "GRCh38"
+    as_var: "off_target_risk"
+  
+  # Combine scores
+  - plugin: "gfl-crispr-evaluator"
+    input:
+      grna_candidates: "${evaluate.candidates}"
+    params:
[…]
