+++ docs/ecosystem/core_plugins/samtools.md
@@ -1,0 +1,304 @@
+# SAMtools Plugin
+
+SAMtools plugin for GeneForgeLang that enables sequence alignment manipulation and analysis directly from GFL workflows.
+
+## Overview
+
+This plugin provides integration with SAMtools, allowing users to manipulate and analyze sequence alignment files (SAM/BAM/CRAM) directly from GeneForgeLang workflows. It supports key SAMtools operations including sorting, indexing, filtering, and statistics generation.
+
+## Features
+
+- **File Manipulation**: Sort, index, and convert alignment files
+- **Filtering**: Filter alignments based on various criteria
+- **Statistics**: Generate alignment statistics and metrics
+- **Configurable Parameters**: Set parameters for each SAMtools operation
+- **Structured Output**: Returns parsed results in a clean JSON format
+- **GFL Integration**: Seamlessly integrates with GeneForgeLang's plugin system
+
+## Installation
+
+```bash
+pip install gfl-plugin-samtools
+```
+
+**Note**: This plugin requires SAMtools to be installed separately. Install SAMtools using your system's package manager or download from [the official website](http://www.htslib.org/).
+
+## Usage
+
+Once installed, the plugin is automatically discovered by the GeneForgeLang service through entry points.
+
+### Example GFL Workflow
+
+```gfl
+run:
+  - plugin: "samtools"
+    operation: "sort"
+    input_bam: "/path/to/unsorted.bam"
+    output_bam: "/path/to/sorted.bam"
+    as_var: "sorting_results"
+
+  - plugin: "samtools"
+    operation: "index"
+    input_bam: "/path/to/sorted.bam"
+    as_var: "indexing_results"
+
+output:
+  - sorted_bam: "${sorting_results.output_bam}"
+  - bam_index: "${indexing_results.output_bai}"
+```
+
+### Parameters
+
+- `operation`: The SAMtools operation to perform (sort, index, view, stats, etc.)
+- `input_bam`: Path to the input BAM file
+- `output_bam`: Path to the output BAM file (for operations that create new files)
+- Additional operation-specific parameters
+
+## Supported Operations
+
+### sort
+
+Sorts SAM/BAM/CRAM files by chromosome and genomic coordinate.
+
+**Parameters:**
+- `input_bam`: Input alignment file
+- `output_bam`: Output sorted file
+- `sort_order`: Sort order (coordinate, queryname, etc.)
+- `memory_limit`: Memory limit for sorting
+
+### index
+
+Creates an index for a sorted BAM/CRAM file.
+
+**Parameters:**
+- `input_bam`: Input sorted alignment file
+- `output_bai`: Output index file (optional, auto-generated if not specified)
+
+### view
+
+Converts between SAM/BAM/CRAM formats and filters alignments.
+
+**Parameters:**
+- `input_file`: Input alignment file
+- `output_file`: Output file
+- `regions`: Genomic regions to include (optional)
+- `filters`: Filtering criteria (mapping quality, flags, etc.)
+
+### stats
+
+Generates comprehensive statistics for alignment files.
+
+**Parameters:**
+- `input_bam`: Input alignment file
+- `output_stats`: Output statistics file
+- `regions`: Genomic regions to analyze (optional)
+
+### flagstat
+
+Counts the number of alignments for each FLAG type.
+
+**Parameters:**
+- `input_bam`: Input alignment file
+- `output_flagstat`: Output flagstat file
+
+## Output Format
+
+The plugin returns structured results containing:
+
[…]
