### 2025-09-24 59e80c9
feat: Stabilization for beta release - Fixed GFL service and added QA tools
