### 2025-09-24 7612a2b
feat: Implement GFL Capability-Aware Validator
- Add comprehensive capability system with feature definitions and dependencies
- Enhance semantic validator with engine capability checking
- Support multiple engine types (basic, standard, advanced, experimental)
- Add capability warnings (not errors) for unsupported features
- Implement spatial genomic feature validation
- Add dependency checking for feature prerequisites
- Create convenience functions for easy validation
- Add comprehensive test suite and documentation

This enables the GFL ecosystem to evolve incrementally by:
- Preventing ex
[…]
