+++ test_capability_validator.py
@@ -1,0 +1,236 @@
+#!/usr/bin/env python3
+"""
+Test script for GFL capability-aware validator.
+This script demonstrates the new capability validation features.
+"""
+
+import sys
+import os
+
+sys.path.append(os.path.join(os.path.dirname(__file__), "gfl"))
+
+from gfl.parser import parse_gfl
+from gfl.semantic_validator import EnhancedSemanticValidator
+from gfl.capability_system import (
+    GFLFeature,
+    get_engine_capabilities,
+    EngineCapabilityChecker
+)
+
+
+def test_capability_validation():
+    """Test the capability-aware validator with different engine types."""
+    
+    # Test GFL content with spatial genomic features
+    gfl_content = """
+loci:
+  - id: "TestGene_Locus"
+    chromosome: "chr1"
+    start: 1000000
+    end: 1001000
+    elements:
+      - id: "TestGene_Promoter"
+        type: "promoter"
+      - id: "TestGene_Body"
+        type: "gene"
+
+rules:
+  - id: "TestSpatialRule"
+    description: "Test rule for spatial genomic reasoning"
+    if:
+      - type: "is_within"
+        element: "TestGene_Promoter"
+        locus: "TestGene_Locus"
+    then:
+      - type: "set_activity"
+        element: "TestGene_Body"
+        level: "active"
+
+simulate:
+  name: "TestSimulation"
+  description: "Test spatial genomic simulation"
+  action:
+    type: "move"
+    element: "TestGene_Promoter"
+    destination: "chr1:2000000"
+  query:
+    - type: "get_activity"
+      element: "TestGene_Body"
+
+experiment:
+  name: "BasicExperiment"
+  description: "Basic experiment that should work on all engines"
+"""
+
+    print("Testing GFL Capability-Aware Validator")
+    print("=" * 50)
+    
+    # Parse the GFL content
+    print("1. Parsing GFL content...")
+    try:
+        ast = parse_gfl(gfl_content)
+        if ast:
+            print("✓ GFL parsing successful")
+            print(f"  AST keys: {list(ast.keys())}")
+        else:
+            print("✗ GFL parsing failed")
+            return False
+    except Exception as e:
+        print(f"✗ GFL parsing error: {e}")
+        return False
+    
+    # Test with different engine capabilities
+    engine_types = ["basic", "standard", "advanced", "experimental"]
+    
+    for engine_type in engine_types:
+        print(f"\n2. Testing with {engine_type.upper()} engine...")
+        
+        try:
+            # Get capabilities for this engine type
+            capabilities = get_engine_capabilities(engine_type)
+            print(f"  Engine supports {len(capabilities)} features")
+            
+            # Create validator with engine capabilities
+            validator = EnhancedSemanticValidator(
+                file_path="test_spatial_genomic.gfl",
+                engine_capabilities=capabilities
+            )
+            
+            # Validate the AST
+            result = validator.validate_ast(ast)
+            
+            # Report results
+            print(f"  Validation result:")
+            print(f"    Errors: {len(result.errors)}")
+            print(f"    Warnings: {len(result.warnings)}")
+            
+            # Show capability warnings
+            capability_warnings = [w for w in result.warnings if hasattr(w, 'feature')]
+            if capability_warnings:
+                print(f"    Capability warnings: {len(capability_warnings)}")
+                for warning in capability_warnings:
+                    print(f"      - {warning.message}")
+                    if warning.feature:
+                        print(f"        Feature: {warning.feature.value}")
+            else:
+                print("    No capability warnings")
+                
+            # Show feature support status
+            required_features = {
[…]

[…]
