### 2025-09-24 44d4fe8
feat: Implement GFL v2.0 Multi-Omic Capabilities - Complete implementation with transcripts, proteins, metabolites blocks and external identifiers
