+++ gfl/semantic_validator.py
@@ -6,8 +6,9 @@ reporting, including location tracking, error codes, and suggested fixes.
 
 import logging
 import os
-from typing import Any, Dict, List, Optional, Union, Set
+from typing import Any, Optional, Union
 
+from gfl.capability_system import CAPABILITY_DEFINITIONS, EngineCapabilityChecker, GFLFeature, GFLValidationWarning
 from gfl.error_handling import (
     EnhancedValidationResult,
     ErrorCodes,
@@ -15,26 +16,21 @@ from gfl.error_handling import (
     SourceLocation,
 )
 from gfl.schema_loader import get_global_schema_loader, load_schemas_from_files
-from gfl.capability_system import (
-    GFLFeature,
-    EngineCapabilityChecker,
-    GFLValidationWarning,
-    CAPABILITY_DEFINITIONS
-)
 
 logger = logging.getLogger(__name__)
 
 
-def validate(ast: Dict[str, Any], file_path: Optional[str] = None, 
-             engine_capabilities: Optional[Set[GFLFeature]] = None) -> EnhancedValidationResult:
+def validate(
+    ast: dict[str, Any], file_path: Optional[str] = None, engine_capabilities: Optional[set[GFLFeature]] = None
+) -> EnhancedValidationResult:
     """Convenience function for GFL validation with capability checking.
-    
+
     Args:
         ast: The GFL AST dictionary to validate.
         file_path: Optional path to the file being validated for error reporting.
         engine_capabilities: Set of GFL features supported by the target engine.
                             If None, no capability checking is performed.
-                            
+
     Returns:
         EnhancedValidationResult with detailed error information and capability warnings.
     """
@@ -42,20 +38,21 @@ def validate(ast: Dict[str, Any], file_path: Optional[str] = None,
     return validator.validate_ast(ast)
 
 
-def validate_with_engine_type(ast: Dict[str, Any], engine_type: str, 
-                             file_path: Optional[str] = None) -> EnhancedValidationResult:
+def validate_with_engine_type(
+    ast: dict[str, Any], engine_type: str, file_path: Optional[str] = None
+) -> EnhancedValidationResult:
     """Validate GFL AST with a predefined engine type.
-    
+
     Args:
         ast: The GFL AST dictionary to validate.
         engine_type: Type of engine ("basic", "standard", "advanced", "experimental").
         file_path: Optional path to the file being validated for error reporting.
-        
+
     Returns:
         EnhancedValidationResult with detailed error information and capability warnings.
     """
     from gfl.capability_system import get_engine_capabilities
-    
+
     capabilities = get_engine_capabilities(engine_type)
     return validate(ast, file_path, capabilities)
 
@@ -67,8 +64,7 @@ class EnhancedSemanticValidator:
     including location tracking, error codes, and suggested fixes.
     """
 
-    def __init__(self, file_path: Optional[str] = None, 
-                 engine_capabilities: Optional[Set[GFLFeature]] = None):
+    def __init__(self, file_path: Optional[str] = None, engine_capabilities: Optional[set[GFLFeature]] = None):
         """Initialize validator.
 
         Args:
@@ -76,7 +72,7 @@ class EnhancedSemanticValidator:
             engine_capabilities: Set of GFL features supported by the target engine.
                                 If None, no capability checking is performed.
         """
-        self.symbol_table: Dict[str, Dict[str, Any]] = {}
+        self.symbol_table: dict[str, dict[str, Any]] = {}
         self.result = EnhancedValidationResult(file_path=file_path)
         self.current_block: Optional[str] = None
         self.nested_level = 0
@@ -84,7 +80,7 @@ class EnhancedSemanticValidator:
         self.capability_checker = EngineCapabilityChecker(engine_capabilities or set())
         self.engine_capabilities = engine_capabilities
 
[…]
