### 2025-09-25 e43cb58
feat: Complete GeneForge v2.0 Beta Launch with GFL Genesis v2 Integration
- Updated tagline to 'The Symbolic Language for Biological Reasoning' across all assets
- Integrated GFL Genesis v2 (MYC enhancer targeting) results into scientific manuscript
- Added comprehensive production execution framework with 3-phase validation
- Implemented spatial genomic reasoning capabilities (Hi-C, ATAC-seq, ChIP-seq)
- Created regulatory gRNA evaluator for enhancer targeting
- Updated README, documentation, and project configuration
- Generated complete scientific validation with dual project results
- Read
[…]
