+++ test_qa_integration.py
@@ -1,0 +1,411 @@
+#!/usr/bin/env python3
+"""
+QA Integration Test Suite for GeneForge Platform
+Executes comprehensive testing across all critical areas.
+"""
+
+import json
+import os
+import subprocess
+import sys
+import time
+from typing import Dict, List, Tuple
+
+import requests
+
+# Add gfl to path
+sys.path.append(os.path.join(os.path.dirname(__file__), "gfl"))
+
+from gfl.parser import parse_gfl
+from gfl.semantic_validator import EnhancedSemanticValidator
+from gfl.capability_system import get_engine_capabilities
+
+
+class QAIntegrationTester:
+    """Comprehensive QA testing for GeneForge platform."""
+    
+    def __init__(self):
+        self.results = {
+            "total_tests": 0,
+            "passed": 0,
+            "failed": 0,
+            "warnings": 0,
+            "bugs_found": [],
+            "test_details": {}
+        }
+        self.base_url = "http://localhost:8000"
+        
+    def run_test(self, test_name: str, test_func) -> bool:
+        """Run a single test and record results."""
+        self.results["total_tests"] += 1
+        print(f"\n🧪 Running: {test_name}")
+        
+        try:
+            success = test_func()
+            if success:
+                self.results["passed"] += 1
+                print(f"✅ PASSED: {test_name}")
+            else:
+                self.results["failed"] += 1
+                print(f"❌ FAILED: {test_name}")
+            return success
+        except Exception as e:
+            self.results["failed"] += 1
+            print(f"❌ ERROR in {test_name}: {e}")
+            self.results["bugs_found"].append({
+                "test": test_name,
+                "error": str(e),
+                "type": "exception"
+            })
+            return False
+    
+    def test_gfl_service_health(self) -> bool:
+        """Test 1: GFL Service Health Check"""
+        try:
+            response = requests.get(f"{self.base_url}/health", timeout=10)
+            if response.status_code == 200:
+                data = response.json()
+                return data.get("status") == "ok"
+            return False
+        except Exception:
+            return False
+    
+    def test_gfl_parser_basic(self) -> bool:
+        """Test 2: Basic GFL Parser Functionality"""
+        try:
+            test_gfl = """
+experiment:
+  name: "test_experiment"
+  description: "Basic parser test"
+"""
+            ast = parse_gfl(test_gfl)
+            return isinstance(ast, dict) and "experiment" in ast
+        except Exception:
+            return False
+    
+    def test_gfl_parser_multi_omic(self) -> bool:
+        """Test 3: Multi-Omic Parser Functionality"""
+        try:
+            test_gfl = """
+transcripts:
+  - id: "test_transcript"
+    gene_source: "TEST"
+    exons: [1, 2, 3]
+
+proteins:
+  - id: "test_protein"
+    translates_from: "transcript(test_transcript)"
+    domains:
+      - id: "test_domain"
+        start: 1
+        end: 100
+"""
+            ast = parse_gfl(test_gfl)
+            return ("transcripts" in ast and "proteins" in ast and 
+                    len(ast["transcripts"]) > 0 and len(ast["proteins"]) > 0)
+        except Exception:
+            return False
+    
+    def test_semantic_validator(self) -> bool:
+        """Test 4: Semantic Validator Functionality"""
+        try:
+            test_gfl = """
+transcripts:
+  - id: "test_transcript"
+    gene_source: "TEST"
+    exons: [1, 2, 3]
+"""
+            ast = parse_gfl(test_gfl)
+            capabilities = get_engine_capabilities("advanced")
+            validator = EnhancedSemanticValidator(engine_capabilities=capabilities)
+            result = validator.validate_ast(ast)
[…]
