### 2025-09-25 72d9651
fix: Apply pre-commit hook formatting and linting fixes
- Fixed trailing whitespace and line ending issues
- Applied ruff formatting and linting fixes
- Updated file encodings to UTF-8
- Resolved mixed line ending issues
