+++ examples/gfl-genesis/execute_genesis_real.py
@@ -4,37 +4,34 @@ GFL Genesis Project - Real Execution Script
 Executes the complete CRISPR gRNA discovery workflow using GeneForge v2.0
 """
 
+import json
 import os
 import sys
-import json
 import time
-import pandas as pd
+from datetime import datetime
+
 import matplotlib.pyplot as plt
 import numpy as np
-from datetime import datetime
+import pandas as pd
 
 # Add GeneForge paths
 sys.path.append("C:\\Users\\usuario\\GeneForge Ecosystem\\GeneForge\\GeneForge")
 sys.path.append("C:\\Users\\usuario\\GeneForge Ecosystem\\GeneForgeLang")
 
+
 def simulate_geneforge_engine():
     """Simulate the GeneForge engine execution with realistic results."""
     print("🚀 Starting GeneForge v2.0 Engine...")
     print("📊 Initializing Multi-Omic Discovery Pipeline...")
-    
+
     # Simulate guided discovery cycles
-    results = {
-        "cycles": [],
-        "candidates": [],
-        "convergence": [],
-        "final_rankings": []
-    }
-    
+    results = {"cycles": [], "candidates": [], "convergence": [], "final_rankings": []}
+
     print("\n🔄 Executing Guided Discovery Cycles...")
-    
+
     for cycle in range(1, 6):
         print(f"\n--- Cycle {cycle}/5 ---")
-        
+
         # Simulate candidate generation and evaluation
         cycle_candidates = []
         for i in range(10):
@@ -45,192 +42,202 @@ def simulate_geneforge_engine():
                 "off_target_score": np.random.beta(2, 8),  # Skewed toward lower scores
                 "combined_score": 0.0,
                 "genomic_position": f"chr17:{43094495 + np.random.randint(0, 10000)}",
-                "efficiency_confidence": np.random.uniform(0.7, 0.95)
+                "efficiency_confidence": np.random.uniform(0.7, 0.95),
             }
-            candidate["combined_score"] = (
-                candidate["on_target_score"] * 0.6 + 
-                (1 - candidate["off_target_score"]) * 0.4
-            )
+            candidate["combined_score"] = candidate["on_target_score"] * 0.6 + (1 - candidate["off_target_score"]) * 0.4
             cycle_candidates.append(candidate)
-        
+
         # Sort by combined score
         cycle_candidates.sort(key=lambda x: x["combined_score"], reverse=True)
-        
+
         # Store cycle results
         cycle_result = {
             "cycle": cycle,
             "best_score": cycle_candidates[0]["combined_score"],
             "avg_score": np.mean([c["combined_score"] for c in cycle_candidates]),
-            "candidates_count": len(cycle_candidates)
+            "candidates_count": len(cycle_candidates),
         }
-        
+
         results["cycles"].append(cycle_result)
         results["candidates"].extend(cycle_candidates)
         results["convergence"].append(cycle_result["best_score"])
-        
+
         print(f"  ✅ Generated {len(cycle_candidates)} candidates")
         print(f"  📈 Best score: {cycle_result['best_score']:.4f}")
         print(f"  📊 Average score: {cycle_result['avg_score']:.4f}")
-        
+
         # Simulate learning and improvement
         time.sleep(0.5)
-    
+
     # Generate final rankings
     all_candidates = results["candidates"]
     all_candidates.sort(key=lambda x: x["combined_score"], reverse=True)
[…]
