+++ test_qa_integration.py
@@ -16,30 +16,23 @@ import requests
 # Add gfl to path
 sys.path.append(os.path.join(os.path.dirname(__file__), "gfl"))
 
+from gfl.capability_system import get_engine_capabilities
 from gfl.parser import parse_gfl
 from gfl.semantic_validator import EnhancedSemanticValidator
-from gfl.capability_system import get_engine_capabilities
 
 
 class QAIntegrationTester:
     """Comprehensive QA testing for GeneForge platform."""
-    
+
     def __init__(self):
-        self.results = {
-            "total_tests": 0,
-            "passed": 0,
-            "failed": 0,
-            "warnings": 0,
-            "bugs_found": [],
-            "test_details": {}
-        }
+        self.results = {"total_tests": 0, "passed": 0, "failed": 0, "warnings": 0, "bugs_found": [], "test_details": {}}
         self.base_url = "http://localhost:8000"
-        
+
     def run_test(self, test_name: str, test_func) -> bool:
         """Run a single test and record results."""
         self.results["total_tests"] += 1
         print(f"\n🧪 Running: {test_name}")
-        
+
         try:
             success = test_func()
             if success:
@@ -52,13 +45,9 @@ class QAIntegrationTester:
         except Exception as e:
             self.results["failed"] += 1
             print(f"❌ ERROR in {test_name}: {e}")
-            self.results["bugs_found"].append({
-                "test": test_name,
-                "error": str(e),
-                "type": "exception"
-            })
+            self.results["bugs_found"].append({"test": test_name, "error": str(e), "type": "exception"})
             return False
-    
+
     def test_gfl_service_health(self) -> bool:
         """Test 1: GFL Service Health Check"""
         try:
@@ -69,7 +58,7 @@ class QAIntegrationTester:
             return False
         except Exception:
             return False
-    
+
     def test_gfl_parser_basic(self) -> bool:
         """Test 2: Basic GFL Parser Functionality"""
         try:
@@ -82,7 +71,7 @@ experiment:
             return isinstance(ast, dict) and "experiment" in ast
         except Exception:
             return False
-    
+
     def test_gfl_parser_multi_omic(self) -> bool:
         """Test 3: Multi-Omic Parser Functionality"""
         try:
@@ -101,11 +90,12 @@ proteins:
         end: 100
 """
             ast = parse_gfl(test_gfl)
-            return ("transcripts" in ast and "proteins" in ast and 
-                    len(ast["transcripts"]) > 0 and len(ast["proteins"]) > 0)
+            return (
+                "transcripts" in ast and "proteins" in ast and len(ast["transcripts"]) > 0 and len(ast["proteins"]) > 0
+            )
         except Exception:
             return False
-    
+
     def test_semantic_validator(self) -> bool:
         """Test 4: Semantic Validator Functionality"""
         try:
@@ -122,7 +112,7 @@ transcripts:
             return len(result.errors) == 0
         except Exception:
             return False
-    
+
     def test_spatial_genomic_features(self) -> bool:
         """Test 5: Spatial Genomic Features"""
         try:
@@ -154,33 +144,33 @@ rules:
             return len(result.errors) == 0
         except Exception:
             return False
-    
+
     def test_conformance_suite_v1_3_0(self) -> bool:
         """Test 6: Conformance Suite v1.3.0"""
         try:
             # Run the conformance suite test
-            result = subprocess.run([
-                sys.executable, "test_conformance_suite_v1_3_0.py"
-            ], capture_output=True, text=True, timeout=60)
-            
+            result = subprocess.run(
+                [sys.executable, "test_conformance_suite_v1_3_0.py"], capture_output=True, text=True, timeout=60
+            )
+
[…]
