+++ examples/gfl-genesis/GFL_GENESIS_FINAL_REPORT.md
@@ -2,10 +2,10 @@
 
 ## Executive Summary
 
-**Project:** GFL Genesis CRISPR gRNA Discovery  
-**Date:** September 25, 2025  
-**GeneForge Version:** v2.0 with Multi-Omic Capabilities  
-**Target Gene:** BRCA1  
+**Project:** GFL Genesis CRISPR gRNA Discovery
+**Date:** September 25, 2025
+**GeneForge Version:** v2.0 with Multi-Omic Capabilities
+**Target Gene:** BRCA1
 **Status:** ✅ **EXPERIMENT COMPLETED SUCCESSFULLY**
 
 ## Experimental Overview
@@ -23,7 +23,7 @@ The GFL Genesis project successfully demonstrated the full capabilities of GeneF
 ### Guided Discovery Workflow
 - **Strategy:** Iterative refinement with 5 discovery cycles
 - **Candidate Generation:** 50 total gRNA candidates (10 per cycle)
-- **Evaluation Metrics:** 
+- **Evaluation Metrics:**
   - On-target efficiency (60% weight)
   - Off-target risk minimization (40% weight)
 - **Learning Model:** Ensemble classifier with multi-feature analysis
@@ -159,6 +159,6 @@ The GFL Genesis project represents a landmark achievement in computational biolo
 
 ---
 
-**Report Generated:** September 25, 2025  
-**GeneForge Team:** Computational Biology Division  
+**Report Generated:** September 25, 2025
+**GeneForge Team:** Computational Biology Division
 **Next Phase:** Beta Release and Community Deployment

[…]
