### 2025-09-25 6726eba
fix: Address critical collaboration points for GFL-GF partnership
- Updated manuscript title: 'GeneForge v2.0'  'GeneForgeLang v2.0'
- Added 'Platform Execution and Visualization' section highlighting GeneForge WebApp
- Clarified nomenclature: GeneForgeLang (open-source) vs GeneForge WebApp (commercial)
- Added direct links to GeneForge WebApp (https://www.geneforge.com.es)
- Created comprehensive collaboration proposal document
- Established clear product differentiation and synergy opportunities
- Ready for joint launch coordination with GeneForge WebApp team
