+++ examples/gfl-genesis/Manuscript_GeneForge_v2_Beta.md
@@ -1,4 +1,4 @@
-# GeneForge v2.0: A Multi-Omic Reasoning Platform for Computational Biology
+# GeneForgeLang v2.0: A Multi-Omic Reasoning Platform for Computational Biology
 
 ## Authors
 GeneForge Development Team
@@ -7,7 +7,7 @@ GeneForge Research Foundation
 
 ## Abstract
 
-We present GeneForge v2.0, a revolutionary computational biology platform that integrates multi-omic data with spatial genomic reasoning capabilities for biological discovery. Building upon the foundation of GeneForgeLang (GFL), our domain-specific language for bio-design, this new version introduces comprehensive support for transcripts, proteins, metabolites, and their spatial relationships within the genome. We demonstrate the platform's capabilities through two comprehensive real-world applications: (1) the discovery of optimal gRNA candidates for CRISPR gene editing of the BRCA1 gene, and (2) the advanced regulatory targeting of MYC oncogene enhancers using spatial genomic reasoning. Our guided discovery workflows successfully identified high-quality candidates with 89.7% combined efficiency score, 84.8% on-target activity, and only 3.0% off-target risk for BRCA1, while demonstrating novel capabilities for enhancer targeting with 85% contact strength and 91% chromatin accessibility scores for MYC regulatory elements. These results validate GeneForge v2.0 as a powerful tool for computational biology research and establish its readiness for community adoption and clinical applications.
+We present GeneForgeLang v2.0, a revolutionary computational biology platform that integrates multi-omic data with spatial genomic reasoning capabilities for biological discovery. Building upon the foundation of GeneForgeLang (GFL), our domain-specific language for bio-design, this new version introduces comprehensive support for transcripts, proteins, metabolites, and their spatial relationships within the genome. We demonstrate the platform's capabilities through two comprehensive real-world applications executed on the GeneForge WebApp: (1) the discovery of optimal gRNA candidates for CRISPR gene editing of the BRCA1 gene, and (2) the advanced regulatory targeting of MYC oncogene enhancers using spatial genomic reasoning. Our guided discovery workflows successfully identified high-quality candidates with 89.7% combined efficiency score, 84.8% on-target activity, and only 3.0% off-target risk for BRCA1, while demonstrating novel capabilities for enhancer targeting with 85% contact strength and 91% chromatin accessibility scores for MYC regulatory elements. These results validate GeneForgeLang v2.0 as a powerful tool for computational biology research and establish its readiness for community adoption and clinical applications.
 
 **Keywords:** computational biology, multi-omics, CRISPR, gene editing, domain-specific language, spatial genomics, regulatory elements, enhancer targeting, Hi-C, ATAC-seq
 
@@ -157,6 +157,18 @@ The experiment leverages advanced spatial reasoning capabilities:
 - **Accessibility Scoring**: ATAC-seq data for chromatin accessibility assessment
 - **Enhancer Activity**: H3K27ac ChIP-seq for active regulatory element identification
 - **Distance Constraints**: Genomic distance calculations between regulatory elements
+
+### Platform Execution and Visualization
+
+All experimental workflows were designed, executed, and analyzed using the GeneForge WebApp (https://www.geneforge.com.es), a comprehensive SaaS platform that provides an intuitive interface for GeneForgeLang workflow development and execution. The GeneForge WebApp serves as the primary execution environment for GeneForgeLang workflows, offering:
+
+- **Visual Workflow Designer**: Drag-and-drop interface for creating complex biological workflows
[…]
