+++ COLLABORATION_PROPOSAL_GFL_GF.md
@@ -1,0 +1,132 @@
+# Propuesta de Colaboración: GeneForgeLang + GeneForge WebApp
+
+## 🎯 Objetivo
+
+Establecer una colaboración estratégica entre el proyecto open-source **GeneForgeLang (GFL)** y la plataforma comercial **GeneForge WebApp** para maximizar el impacto científico y comercial de ambos productos.
+
+## 📋 Puntos Críticos Identificados
+
+### Punto Crítico #1: Clarificación de Nomenclatura
+
+**Problema Identificado:**
+- El manuscrito científico actualmente se refiere a la plataforma como "GeneForge v2.0"
+- Esto crea confusión con el producto comercial "GeneForge WebApp"
+- Los usuarios pueden confundir ambos productos al buscar "GeneForge"
+
+**Solución Propuesta:**
+- **GeneForgeLang (GFL)**: Lenguaje de dominio específico y motor open-source
+- **GeneForge WebApp**: Plataforma SaaS comercial que utiliza GFL
+- Título del manuscrito: "GeneForgeLang v2.0: A Multi-Omic Reasoning Platform for Computational Biology"
+
+### Punto Crítico #2: Oportunidad de Sinergia
+
+**Problema Identificado:**
+- El manuscrito no menciona explícitamente GeneForge WebApp como plataforma de ejecución
+- Se pierde la oportunidad de crear un puente directo desde la ciencia al producto comercial
+
+**Solución Propuesta:**
+- Agregar sección "Platform Execution and Visualization" en el manuscrito
+- Describir GeneForge WebApp como la interfaz principal para ejecutar workflows GFL
+- Incluir enlace directo a https://www.geneforge.com.es
+
+## 🤝 Propuesta de Colaboración
+
+### Beneficios para GeneForgeLang (GFL)
+1. **Validación Científica**: El manuscrito proporciona credibilidad científica al proyecto
+2. **Adopción Comunitaria**: Los investigadores pueden usar GFL directamente o a través de GeneForge WebApp
+3. **Ecosistema Completo**: Demuestra la viabilidad de un modelo open-source + comercial
+
+### Beneficios para GeneForge WebApp
+1. **Credibilidad Científica**: El manuscrito valida científicamente la plataforma
+2. **Trazabilidad de Usuarios**: Los lectores del paper pueden encontrar fácilmente la WebApp
+3. **Diferenciación Clara**: Se establece claramente la relación entre GFL y GeneForge WebApp
+
+### Beneficios para la Comunidad Científica
+1. **Claridad de Productos**: Entienden la diferencia entre GFL (open-source) y GeneForge WebApp (comercial)
+2. **Opciones de Uso**: Pueden elegir entre usar GFL directamente o a través de la WebApp
+3. **Continuidad**: Pueden empezar con GFL y migrar a GeneForge WebApp según sus necesidades
+
+## 📝 Cambios Propuestos en el Manuscrito
+
+### 1. Título Actualizado
+```
+GeneForgeLang v2.0: A Multi-Omic Reasoning Platform for Computational Biology
+```
+
+### 2. Abstract Actualizado
+- Mencionar que los experimentos se ejecutaron en "GeneForge WebApp"
+- Establecer claramente la relación entre GFL y GeneForge WebApp
+
+### 3. Nueva Sección: "Platform Execution and Visualization"
+```markdown
+### Platform Execution and Visualization
+
[…]
