### 2025-09-25 45ab76c
fix: Apply pre-commit formatting to collaboration proposal
