### 2025-09-27 c564074
Add conformance test suite v1.4.0 for large-scale genomic editing operations
- Add v1.4.0 directory structure with large_scale_editing module
- Implement test cases for delete, insert, and invert operations
- Add comprehensive documentation for new bridge editor syntax
- Update main conformance suite README to include v1.4.0 features
- Provide standardized tests for GeneForge team validation
